% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_mutation_counts.R
\name{add_mutation_counts}
\alias{add_mutation_counts}
\title{Calculate Synonymous/Nonsynonymous mutation frequency}
\usage{
add_mutation_counts(data_frame)
}
\arguments{
\item{data_frame}{a data frame in long format with columns sample, trv_type}
}
\value{
a data frame with synonymous/nonsynonymous counts appended
}
\description{
Creates a data frame giving synonymous/nonsynonymous counts
}

