% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/lohSpec-class.R
\docType{methods}
\name{addBlankRegion}
\alias{addBlankRegion}
\alias{addBlankRegion,data.table-method}
\alias{addBlankRegion}
\title{Method addBlankRegion}
\usage{
addBlankRegion(object, ...)

\S4method{addBlankRegion}{data.table}(object, chrData, ...)
}
\arguments{
\item{object}{Object of class VarScanFormat}

\item{...}{additional arguments to passed}

\item{chrData}{of class data.table}

\item{object}{of class lohData}
}
\description{
Method addBlankRegion
}
