% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/align_waterfall.R
\name{align_waterfall}
\alias{align_waterfall}
\title{align plots}
\usage{
align_waterfall(p2, p1, p3, p4)
}
\arguments{
\item{p2}{ggplot object displaying mutation burden on gene}

\item{p1}{ggplot object displaying a mutation landscape}

\item{p3}{ggplot object displaying mutation burden on sample}

\item{p4}{ggplot object displaying clinical information "optional"}
}
\value{
a grob object
}
\description{
align mutation landscape, mutation burden on sample, and mutation burden on gene plots
}

