% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/align_y.R
\name{align_y}
\alias{align_y}
\title{align plots}
\usage{
align_y(p2, p1, p3, title)
}
\arguments{
\item{p2}{ggplot object displaying mutation burden on gene}

\item{p1}{ggplot object displaying a mutation landscape}

\item{p3}{ggplot object displaying mutation burden on sample}

\item{title}{optional a character string giving a title for the plot}
}
\value{
a grob object
}
\description{
align mutation landscape, mutation burden on sample, and mutation burden on gene plots
}

