% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/align_y_cn.R
\name{align_y_cn}
\alias{align_y_cn}
\title{align cn plots on x axis}
\usage{
align_y_cn(p1, p2)
}
\arguments{
\item{p1}{ggplot object of chromosome}

\item{p2}{ggplot object of CN}
}
\value{
ggplot object
}
\description{
given a chromosome and cn plot align plot widths
}

