% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/align_y_TvTi.R
\name{align_y_TvTi}
\alias{align_y_TvTi}
\title{align TvTi plots on y axis}
\usage{
align_y_TvTi(p1, p2)
}
\arguments{
\item{p1}{main plot}

\item{p2}{left subplot}
}
\value{
ggplot object
}
\description{
align transition/transversion plots
}

