% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/anno_trans_tranv.R
\name{anno_trans_tranv}
\alias{anno_trans_tranv}
\title{Annotate Transitions and Transversions}
\usage{
anno_trans_tranv(x)
}
\arguments{
\item{x}{Object of class data frame containing columns 'reference', 'variant'}
}
\value{
Object of class data frame with transition/transversion annotations appended
}
\description{
Given a data frame with columns reference and variant annotate the base change occurring
}

