% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{annoGenomeCoord}
\alias{annoGenomeCoord}
\title{Method annoGenomeCoord}
\usage{
annoGenomeCoord(object, ...)
}
\arguments{
\item{object}{Object of class VarScanFormat}

\item{...}{additional arguments to passed}
}
\description{
Method annoGenomeCoord
}
