% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/lohSpec-class.R
\docType{methods}
\name{arrangeLohPlots}
\alias{arrangeLohPlots}
\alias{arrangeLohPlots,lohSpecPlots-method}
\alias{arrangeLohPlots}
\title{Method arrangeLohPlots}
\usage{
arrangeLohPlots(object, ...)

\S4method{arrangeLohPlots}{lohSpecPlots}(object, sectionHeights, verbose, ...)
}
\arguments{
\item{object}{Object of class VarScanFormat}

\item{...}{additional arguments to passed}

\item{object}{of class lohData}
}
\description{
Method arrangeLohPlots
}
