% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assign_ggplotGrob_width.R
\name{assign_ggplotGrob_width}
\alias{assign_ggplotGrob_width}
\title{assign ggplotGrob width}
\usage{
assign_ggplotGrob_width(x, max_width)
}
\arguments{
\item{x}{ggplotGrob object}

\item{max_width}{grob object specifying width to reassign ggplotGrob with}
}
\value{
ggplotGrob
}
\description{
assign width of ggplotGrob object
}

