% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bamreadcount.R
\name{bamreadcount}
\alias{bamreadcount}
\title{Count nucleotide reads at SNP locations}
\usage{
bamreadcount(bamfile, genome, targetbed = NULL)
}
\arguments{
\item{bamfile}{Path to the bam file}
}
\value{
data frame
}
\description{
Given the bam file path, count the number of reads at the 24 SNP locations
}

