% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GGgenome.R
\docType{data}
\name{brcaMAF}
\alias{brcaMAF}
\title{Truncated BRCA MAF file}
\format{a data frame with 2773 observations and 55 variables}
\usage{
data(brcaMAF)
}
\description{
A data set containing 50 samples corresponding to "Breast invasive carcinoma" originating from the TCGA project in .maf format (version 2.4):
https://wiki.nci.nih.gov/display/TCGA/TCGA+MAF+Files#TCGAMAFFiles-BRCA:Breastinvasivecarcinoma, /dccfiles_prod/tcgafiles/distro_ftpusers/anonymous/tumor/brca/gsc/genome.wustl.edu/illuminaga_dnaseq/mutations/genome.wustl.edu_BRCA.IlluminaGA_DNASeq.Level_2.5.3.0/genome.wustl.edu_BRCA.IlluminaGA_DNASeq.Level_2.5.3.0.somatic.maf
}
\keyword{datasets}

