% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build.cnFreq.R
\name{build.cnFreq}
\alias{build.cnFreq}
\title{Construct CN frequency plot}
\usage{
build.cnFreq(data_frame, plotType, plot_title = NULL, background = "grey90",
  CN_low_colour = "#002EB8", CN_high_colour = "#A30000", x_lab_size = 12,
  y_lab_size = 12, facet_lab_size = 10, layers = NULL)
}
\arguments{
\item{data_frame}{object of class data frame containing columns chromosome, start, end, gain, and loss}

\item{plotType}{character string to determine whether to plot proportions or frequencies}

\item{plot_title}{character string for title of plot}

\item{background}{character string specifying backround color of plot}

\item{CN_low_colour}{character string specifying low value of colour gradient}

\item{CN_high_colour}{character string specifying high value of colour gradient}

\item{x_lab_size}{integer specifying the size of the X label}

\item{y_lab_size}{integer specifying the size of the Y label}

\item{facet_lab_size}{integer specifying the size of the faceted labels}

\item{layers}{Additional layers to be plotted, can be a theme but must be a ggplot layer}
}
\value{
ggplot object
}
\description{
given a data frame construct a plot to display proportions of losses and gains across the genome
}

