% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build.cnView.main.R
\name{build.cnView.main}
\alias{build.cnView.main}
\title{construct CN plot}
\usage{
build.cnView.main(x, chr, cnDiff = FALSE)
}
\arguments{
\item{chr}{a character string specifying chromosome}

\item{cnDiff}{Boolean specifying whether values in cn are copy number differences or actual copy number}

\item{data_frame}{a data frame with columns chromosome, coordinate, cn, p_value}
}
\value{
ggplot2 object
}
\description{
given a CN data frame plot points in ggplot
}

