% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build.main.Mutspec.R
\name{build.main.mutSpec}
\alias{build.main.mutSpec}
\title{Plot a mutation heatmap}
\usage{
build.main.mutSpec(data_frame, grid = TRUE, label_x = FALSE,
  gene_label_size = 8, file_type = "MGI", drop_mutation = FALSE,
  plot_x_title = TRUE, plot_label = FALSE, plot_label_size = 4,
  plot_palette = NULL, layers = NULL)
}
\arguments{
\item{data_frame}{a data frame in MAF format}

\item{grid}{boolean value whether to overlay a grid on the plot}

\item{label_x}{boolean value whether to label the x axis}

\item{gene_label_size}{numeric value indicating the size of the gene labels on the y-axis}

\item{file_type}{character string specifying the file type, one of 'MAF' or 'MGI'}

\item{drop_mutation}{Boolean specifying whether to drop unused "mutation type" levels from the legend}

\item{plot_x_title}{Boolean specifying whether to plot the x_axis title}

\item{plot_label}{Boolean specifying whether to plot text inside each cell}

\item{plot_label_size}{Integer specifying text size of cell labels}

\item{layers}{additional ggplot2 layers to plot}

\item{plot_pallete}{Character vector specifying colors to fill on mutation type}
}
\value{
a ggplot2 object
}
\description{
Plot a Mutation Landscape with variables sample, gene, mutation
}

