% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build.mutRecurA.mutSpec.R
\name{build.mutRecurA.mutSpec}
\alias{build.mutRecurA.mutSpec}
\title{plot mutation burden}
\usage{
build.mutRecurA.mutSpec(data_frame, coverage_space)
}
\arguments{
\item{data_frame}{a data frame in MAF format}

\item{coverage_space}{an integer specifying the coverage space in base pairs from which a mutation could occur}
}
\value{
a ggplot object
}
\description{
plot a barchart showing mutations per MB
}

