% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build.mutRecurB.mutSpec.R
\name{build.mutRecurB.mutSpec}
\alias{build.mutRecurB.mutSpec}
\title{plot mutation burden}
\usage{
build.mutRecurB.mutSpec(x)
}
\arguments{
\item{x}{a data frame containing columns sample, mut_burden}
}
\value{
a ggplot object
}
\description{
plot a barchart showing mutation burden given by data frame
}

