% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/build_chromosome.R
\name{build_chromosome}
\alias{build_chromosome}
\title{build chromosome}
\usage{
build_chromosome(data_frame, chromosome, chr_txt_angle = chr_txt_angle,
  chr_txt_size = chr_txt_size)
}
\arguments{
\item{data_frame}{a data frame with columns Chr, Coord, Tumor, Normal, Diff, p_value}

\item{chromosome}{character string specifying UCSC chromosome to plot one of chr... or all}

\item{chr_txt_angle}{integer specifying angle of text when plotting band text}

\item{chr_txt_size}{integer specifying size of text when plotting band text}
}
\value{
ggplot object
}
\description{
given a data frame with cytogenetic band locations plot chromosome in ggplot
}

