% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/build_CN_plot.R
\name{build_CN_plot}
\alias{build_CN_plot}
\title{construct CN plot}
\usage{
build_CN_plot(x, chr)
}
\arguments{
\item{chr}{a character string specifying chromosome}

\item{data_frame}{a data frame with columns Chr, Coord, Tumor, Normal, Diff, p_value}
}
\value{
ggplot2 object
}
\description{
given a CN data frame plot points in ggplot
}

