% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_gene.R
\name{build_gene}
\alias{build_gene}
\title{build gene plot}
\usage{
build_gene(data_frame, display_x_axis = T, x_limits = NULL,
  gene_colour = NULL, transcript_name = FALSE, transcript_name_size = 6)
}
\arguments{
\item{data_frame}{an object of class data frame specifying gene feature information}

\item{x_limits}{vector specifying x-axis limits of plot}

\item{gene_colour}{character specifying colour of gene to be plotted}

\item{transcript_name}{Boolean specifying whether to plot USCS transcript names}

\item{transcript_name_size}{Integer specifying the size of the transcript name text}

\item{master_gene}{an object of class data frame specifying the master gene feature information resultant of compression}

\item{display_axis}{Boolean specifying whether to display X axis coordinate values}
}
\value{
ggplot object
}
\description{
given a data frame with gene feature information build the ggplot2 object
}

