% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/build_lolli.R
\name{build_lolli}
\alias{build_lolli}
\title{Construct Lolliplot}
\usage{
build_lolli(gene_data, length, mutation_observed, mutation_cosmic)
}
\arguments{
\item{gene_data}{object of class dataframe giving protien domain and gene information}

\item{length}{integer specifying the length of the protien in amino acids}

\item{mutation_observed}{object of class data frame specifying mutations observed in input file}

\item{mutation_cosmic}{optional object of class data frame specifying mutations observed in cosmic}
}
\value{
a ggplot2 object
}
\description{
Construct Lolliplot given gene and mutation data
}

