% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_mutation_recurrence_b.R
\name{build_mutation_recurrence_b}
\alias{build_mutation_recurrence_b}
\title{plot mutation burden}
\usage{
build_mutation_recurrence_b(x)
}
\arguments{
\item{x}{a data frame containing columns sample, mut_burden}
}
\value{
a ggplot object
}
\description{
plot a barchart showing mutation burden given by data frame
}

