% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_track_name.R
\name{build_track_name}
\alias{build_track_name}
\title{build label for plot}
\usage{
build_track_name(name, bg_fill = "black", text_fill = "white",
  border = "black", size = 10)
}
\arguments{
\item{name}{character string giving the name of the track}

\item{bg_fill}{character string giving the colour to fill the label}

\item{text_fill}{character string giving the colour to fill the text}

\item{border}{character string specifying the colour to fill the border of the label}

\item{size}{integer specifying the size of the text within the label}
}
\value{
ggplot object
}
\description{
given a name create a label
}

