% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_trans_tranv.R
\name{build_trans_tranv}
\alias{build_trans_tranv}
\title{build transitions/transversions}
\usage{
build_trans_tranv(x, type = "Proportion", x_axis_text_angle = 45)
}
\arguments{
\item{x}{Object of class data frame containing columns 'trans_tranv', 'sample', 'Freq', and 'Prop'}

\item{type}{Object of class character specifying whether to plot the Proportion or Frequency, one of "Prop"}

\item{x_axis_text_angle}{Integer specifying the angle to labels on x_axis}
}
\value{
GGplot Object
}
\description{
Given a data frame with columns 'trans_tranv', 'sample', 'Freq', and 'Prop', build a transition/transversion plot
}

