% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/lohSpec-class.R
\docType{methods}
\name{buildLohFreq}
\alias{buildLohFreq}
\alias{buildLohFreq,lohData-method}
\alias{buildLohFreq}
\title{Method buildLohFreq}
\usage{
buildLohFreq(object, ...)

\S4method{buildLohFreq}{lohData}(object, plotALohCutoff, plotAType, plotAColor,
  plotALayers, verbose, ...)
}
\arguments{
\item{object}{Object of class VarScanFormat}

\item{...}{additional arguments to passed}

\item{object}{of class lohData}
}
\description{
Method buildLohFreq
}
