% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_trans_tranv_freq.R
\name{calc_trans_tranv_freq}
\alias{calc_trans_tranv_freq}
\title{Calculate Transition/Transversion Frequency}
\usage{
calc_trans_tranv_freq(x)
}
\arguments{
\item{x}{Object of class data frame containing columns 'reference', 'variant', 'sample', 'trans_tranv'}
}
\value{
Object of class data frame with Frequency and Proportion of Transistions/Transversions appended on a sample level
}
\description{
Given a data frame with columns reference, variant, and sample construct a transition/transversion plot
}

