% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcGC.R
\name{calcGC}
\alias{calcGC}
\title{Calculate GC content}
\usage{
calcGC(gr, genome)
}
\arguments{
\item{gr}{A Granges object to calculate GC content for}

\item{genome}{Object of class BSgenome specifying the genome to calculate GC content from}
}
\value{
Object of class GRanges
}
\description{
Calculate GC content for elements in a GRanges object
}

