% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Calculate_Pseudoexons.R
\name{Calculate_Pseudoexons}
\alias{Calculate_Pseudoexons}
\title{calculate pseudoexons}
\usage{
Calculate_Pseudoexons(x, chr)
}
\arguments{
\item{x}{object of class dataframe with column names start end}

\item{chr}{character string specifying the chromosome for pseudoexon calculation}
}
\value{
Object of class data frame
}
\description{
calculate pseudoexons i.e. an exonic region in one isoform but not another
}

