% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cdsFromTXID.R
\name{cdsFromTXID}
\alias{cdsFromTXID}
\title{cdsFromTXID}
\usage{
cdsFromTXID(txdb, txid)
}
\arguments{
\item{txdb}{A TxDb object for a genome}

\item{txid}{A list of TXIDs}
}
\value{
Object of class Granges
}
\description{
Return CDS coordinates as a GRanges object given transcript IDs
}

