% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clinical-class.R
\docType{methods}
\name{Clinical-class}
\alias{Clinical-class}
\alias{Clinical}
\alias{Clinical}
\title{Class Clinical}
\usage{
\S4method{initialize}{Clinical}(.Object, path, inputData, inputFormat,
  legendColumns, palette, clinicalLayers, verbose)

Clinical(path, inputData = NULL, inputFormat = c("wide", "long"),
  legendColumns = 1, palette = NULL, clinicalLayers = NULL,
  verbose = FALSE)
}
\arguments{
\item{path}{String specifying the path to clinical data, file must have the
column "sample".}

\item{inputData}{Optional data.table or data.frame object holding clinical
data, used only if path is not specified.}

\item{inputFormat}{String specifying the input format of the data given, one
of wide or long format.}

\item{legendColumns}{Integer specifying the number of columns in the legend.}

\item{palette}{Named character vector supplying colors for clinical variables.}

\item{clinicalLayers}{list of ggplot2 layers to be passed to the plot.}

\item{verbose}{Boolean specifying if progress should be reported.}
}
\description{
An S4 class to store clinical information and plots

Initalizer method for the Clinical class

Constructor for the Clinical class.
}
\section{Slots}{

\describe{
\item{\code{clinicalGrob}}{gtable object for the clinical plot.}

\item{\code{clinicalLayers}}{list of ggtheme or ggproto objects used to build the plot.}

\item{\code{clinicalData}}{data.table object to store clinical data}
}}

