% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CN_cohort_plot.R
\name{CN_cohort_plot}
\alias{CN_cohort_plot}
\title{Construct CN cohort plot}
\usage{
CN_cohort_plot(data_frame, genome = "hg19", plot_title = NULL,
  background = "grey90", CN_low_colour = "#002EB8",
  CN_high_colour = "#A30000", x_lab_size = 12, y_lab_size = 12,
  facet_lab_size = 10)
}
\arguments{
\item{data_frame}{object of class data frame containing columns Chromosome, Start, Stop, SegMean, Sample.name}

\item{genome}{character string specifying UCSC genome from which data is derived}

\item{plot_title}{character string for title of plot}

\item{background}{character string specifying backround color of plot}

\item{CN_low_colour}{character string specifying low value of colour gradient}

\item{CN_high_colour}{character string specifying high value of colour gradient}

\item{x_lab_size}{integer specifying the size of the X label}

\item{facet_lab_size}{integer specifying the size of the faceted labels}

\item{Y_lab_size}{integer specifying the size of the Y label}
}
\value{
ggplot object
}
\description{
given a data frame construct a plot to display CN information for a group of samples
}

