% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CN_plot.R
\name{CN_plot}
\alias{CN_plot}
\title{plot CN values}
\usage{
CN_plot(x, genome = "hg19", chr = "chr1", chr_txt_angle = 45,
  chr_txt_size = 5)
}
\arguments{
\item{x}{a data frame with columns Chr, Coord, Tumor, Normal, Diff, p_value}

\item{genome}{character string specifying UCSC genome to use}

\item{chr}{character string specifying UCSC chromosome to plot one of chr... or all}

\item{chr_txt_angle}{integer specifying angle of text when plotting band text}

\item{chr_txt_size}{integer specifying size of text when plotting band text}
}
\value{
ggplot object
}
\description{
given data frame with columns Chr, Coord, Tumor, Normal, Diff, p_value plot CN
}

