% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cnFreq.qual.R
\name{cnFreq.qual}
\alias{cnFreq.qual}
\title{check input to cnFreq}
\usage{
cnFreq.qual(x)
}
\arguments{
\item{x}{a data frame with columns chromosome, start, end, gain, and loss, or chromosome, start, end, segmean, and sample}
}
\value{
list containing data frame passing quality checks and the type of plot (proportional losses/gain or frequency of losses/gains)
}
\description{
Perform a data quality check for inputs to cnFreq
}

