% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnFreq_buildMain.R
\name{cnFreq_buildMain}
\alias{cnFreq_buildMain}
\title{Construct CN frequency plot}
\usage{
cnFreq_buildMain(x, plotType, dummy_data, plot_title = NULL,
  CN_low_colour = "#002EB8", CN_high_colour = "#A30000", x_lab_size = 12,
  y_lab_size = 12, facet_lab_size = 10, plotLayer = NULL)
}
\arguments{
\item{x}{object of class data frame containing columns chromosome,
start, end, gain, and loss}

\item{plotType}{character string to determine whether to plot proportions or
frequencies}

\item{dummy_data}{Object of class data frame containing columns chromosome,
start, end, cn, sample. Used for defining chromosome boundaries}

\item{plot_title}{character string for title of plot}

\item{CN_low_colour}{character string specifying low value of colour gradient}

\item{CN_high_colour}{character string specifying high value of colour
gradient}

\item{x_lab_size}{integer specifying the size of the X label}

\item{y_lab_size}{integer specifying the size of the Y label}

\item{facet_lab_size}{integer specifying the size of the faceted labels}

\item{plotLayer}{Additional layer to be plotted, can be a theme but must be a
ggplot layer}
}
\value{
ggplot object
}
\description{
given a data frame construct a plot to display proportions of losses and
gains across the genome
}

