% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnFreq_qual.R
\name{cnFreq_qual}
\alias{cnFreq_qual}
\title{check input to cnFreq}
\usage{
cnFreq_qual(x)
}
\arguments{
\item{x}{a data frame with columns chromosome, start, end, segmean, and sample}
}
\value{
data frame passing quality checks
}
\description{
Perform a data quality check for input to cnFreq
}

