% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinedCnLoh-class.R
\name{cnLoh-class}
\alias{cnLoh-class}
\alias{cnLoh}
\title{Class cnLoh}
\usage{
cnLoh(cnInput, lohInput, samples, chromosomes, BSgenome, windowSize, step,
  normal, plotAColor, plotALayers, plotBAlpha, somaticLohCutoff,
  plotBTumorColor, plotBNormalColor, plotBLayers, plotCLimits, plotCLowColor,
  plotCHighColor, plotCLayers, sectionHeights, verbose)
}
\arguments{
\item{samples}{Character vector specifying samples to plot. If not NULL
all samples in "input" not specified with this parameter are removed.}

\item{chromosomes}{Character vector specifying chromosomes to plot. If not NULL
all chromosomes in "input" not specified with this parameter are removed.}

\item{BSgenome}{Object of class BSgenome to extract genome wide chromosome 
coordinates}

\item{input}{Object of class cnLohDataFormat}
}
\description{
An S4 class for the cn, somatic loh, and germline loh plots

Constructor for the cnLoh class
}
\section{Slots}{

\describe{
\item{\code{cnData}}{data.table object for cn plot}

\item{\code{cnPlot}}{gtable object for the cn plot}

\item{\code{somaticLohData}}{data.table object for the somatic loh plot}

\item{\code{somaticLohPlot}}{gtable object for the somatic loh plot}

\item{\code{germlineLohData}}{data.table object for the germline loh plot}

\item{\code{germlineLohData}}{gtable object for the germline loh plot}
}}

