% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinedCnLoh-class.R
\docType{class}
\name{cnLohData}
\alias{cnLohData}
\alias{cnLohData}
\title{Private Class cnLohData}
\usage{
cnLohData(cnInput, lohInput, samples, chromosomes, BSgenome, windowSize, step,
  normal, verbose = FALSE)
}
\arguments{
\item{object}{Object of class cnLohDataFormat}
}
\description{
An S4 class for the data to plot cn, somatic LOH, and germline LOH plots

Constructor for the cnLohData class
}
