% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cnSpec.qual.R
\name{cnSpec.qual}
\alias{cnSpec.qual}
\title{Construct CN cohort plot}
\usage{
cnSpec.qual(x, y, genome)
}
\arguments{
\item{x}{object of class data frame containing columns Chromosome, Start, Stop, SegMean, Sample.name}

\item{y}{object of class data frame containing user supplied chromosome locations}

\item{genome}{character string specifying a user supplied genome}
}
\value{
character string specifying input passed quality check
}
\description{
given a data frame construct a plot to display CN information for a group of samples
}

