% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnSpec_buildMain.R
\name{cnSpec_buildMain}
\alias{cnSpec_buildMain}
\title{Construct CN cohort plot}
\usage{
cnSpec_buildMain(data_frame, dummy_data, plot_title = NULL,
  CN_low_colour = "#002EB8", CN_high_colour = "#A30000", x_lab_size = 12,
  y_lab_size = 12, facet_lab_size = 10, layers = NULL,
  CNscale = "absolute")
}
\arguments{
\item{data_frame}{object of class data frame containing columns chromosome,
start, end, cn, sample}

\item{dummy_data}{Object of class data frame containing columns chromosome,
start, end, cn, sample. Used for defining chromosome boundaries}

\item{plot_title}{character string for title of plot}

\item{CN_low_colour}{character string specifying low value of colour gradient}

\item{CN_high_colour}{character string specifying high value of colour
gradient}

\item{x_lab_size}{integer specifying the size of the X label}

\item{y_lab_size}{integer specifying the size of the Y label}

\item{facet_lab_size}{integer specifying the size of the faceted labels}

\item{layers}{Additional layers to be plotted, can be a theme but must be a
ggplot layer}

\item{CNscale}{Character string specifying if copy number calls supplied are
relative (i.e.copy neutral == 0) or absolute (i.e. copy neutral ==2). One of 
"relative" or "absolute"}
}
\value{
ggplot object
}
\description{
given a data frame construct a plot to display CN information for a group
of samples
}

