% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnSpec_qual.R
\name{cnSpec_qual}
\alias{cnSpec_qual}
\title{Construct CN cohort plot}
\usage{
cnSpec_qual(x, y, genome)
}
\arguments{
\item{x}{object of class data frame containing columns chromosome, start,
stop, segmean, sample}

\item{y}{object of class data frame containing user supplied chromosome
locations}

\item{genome}{character string specifying a user supplied genome}
}
\value{
character string specifying input passed quality check
}
\description{
given a data frame construct a plot to display CN information for a group of
samples
}

