% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cnView.qual.R
\name{cnView.qual}
\alias{cnView.qual}
\title{check input to cnView}
\usage{
cnView.qual(x, y, genome)
}
\arguments{
\item{x}{a data frame with columns chromosome, coordinate, cn, p_value}

\item{y}{a data frame with columns "chrom", "chromStart", "chromEnd", "name", "gieStain"}

\item{genome}{character string specifying UCSC genome to use}

\item{chr}{character string specifying UCSC chromosome to plot one of chr... or all}

\item{chr_txt_angle}{integer specifying angle of text when plotting band text}

\item{chr_txt_size}{integer specifying size of text when plotting band text}
}
\value{
a list of data frames passing quality checks
}
\description{
Perform a data quality check for inputs to cnView
}

