% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cnView.R
\name{cnView}
\alias{cnView}
\title{plot CN values}
\usage{
cnView(x, y = NULL, z = NULL, genome = "hg19", chr = "chr1",
  main.cnDiff = FALSE, chr_txt_angle = 45, chr_txt_size = 5)
}
\arguments{
\item{x}{a data frame with columns chromosome, coordinate, cn, p_value}

\item{y}{a data frame with columns "chrom", "chromStart", "chromEnd", "name", "gieStain"}

\item{z}{a data frame with columns "chromosome", "start", "end", "segmean"}

\item{genome}{character string specifying UCSC genome to use}

\item{chr}{character string specifying UCSC chromosome to plot one of chr... or all}

\item{main.cnDiff}{Boolean specifying whether values in cn are copy number differences or actual copy number}

\item{chr_txt_angle}{integer specifying angle of text when plotting band text}

\item{chr_txt_size}{integer specifying size of text when plotting band text}
}
\value{
ggplot object
}
\description{
given data frame with columns Chr, Coord, Tumor, Normal, Diff, p_value plot CN
}

