% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cnView_align.R
\name{cnView_align}
\alias{cnView_align}
\title{align cn plots on x axis}
\usage{
cnView_align(p1, p2)
}
\arguments{
\item{p1}{ggplot object of chromosome}

\item{p2}{ggplot object of CN}
}
\value{
ggplot object
}
\description{
given a chromosome and cn plot align plot widths
}

