% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cnView_subsetChr.R
\name{cnView_subsetChr}
\alias{cnView_subsetChr}
\title{subset based on chr}
\usage{
cnView_subsetChr(x, chr)
}
\arguments{
\item{x}{a data frame with columns chromosome}

\item{chr}{character string specifying UCSC chromosome to subset on}
}
\value{
object of class data frame
}
\description{
given a data frame subset out specific a chromosome
}

