% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compIdent_bamRcnt.R
\name{comIdent_bamRcnt}
\alias{comIdent_bamRcnt}
\alias{compIdent_bamRcnt}
\title{Count nucleotide reads at SNP locations}
\usage{
compIdent_bamRcnt(bamfile, genome, targetbed = NULL)
}
\arguments{
\item{bamfile}{Path to the bam file}

\item{genome}{Object of class BSgenome corresponding to a genome of interest}

\item{targetbed}{Object of class data frame containing target locations in
.bed format and containing columns chr, start, end}
}
\value{
object of class data frame containing readcount information
}
\description{
Given the bam file path, count the number of reads at specified snp locations
}

