% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compare_identities_plot.R
\name{compare_identities_plot}
\alias{compare_identities_plot}
\title{Compare sample identities}
\usage{
compare_identities_plot(count_tables, sample_names)
}
\arguments{
\item{sample_names}{list of names associated with each sample in counts list}

\item{counts}{list of readcount tables}
}
\value{
none
}
\description{
Produce an identity SNP plot displaying VAFs of 24 SNP locations and coverage
information to compare multiple sample identities
}

