% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparemysamples.R
\name{comparemysamples}
\alias{comparemysamples}
\title{Compare the identities of multiple samples}
\usage{
comparemysamples(x)
}
\arguments{
\item{x}{data frame of bam files and associated sample names}
}
\value{
gridExtra object
}
\description{
Given the bam file path, count the number of reads at the 24 SNP locations
}

