% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compIdent_bamRcnt.R
\name{compIdent_bamRcnt}
\alias{compIdent_bamRcnt}
\title{Count nucleotide reads at SNP locations}
\usage{
compIdent_bamRcnt(bamfile, genome, target = NULL, debug = FALSE)
}
\arguments{
\item{bamfile}{Path to the bam file}

\item{genome}{Object of class BSgenome corresponding to a genome of interest}

\item{target}{Object of class data frame containing target locations in
1-base format and containing columns "chr", "start", "end", "var", "name"}

\item{debug}{Boolean specifying if test datasets should be used for
debugging.}
}
\value{
object of class data frame containing readcount information
}
\description{
Given the bam file path, count the number of reads at specified snp locations
}

