% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compIdent_bamRcnt_qual.R
\name{compIdent_bamRcnt_qual}
\alias{compIdent_bamRcnt_qual}
\title{Count nucleotide reads at SNP locations}
\usage{
compIdent_bamRcnt_qual(bai, genome, targetbed)
}
\arguments{
\item{bai}{Path to the bam index file}

\item{genome}{Object of class BSgenome corresponding to a genome of interest}

\item{targetbed}{Object of class data frame containing target locations in .bed format}
}
\value{
list of data objects passing quality checks
}
\description{
Given the bam file path, count the number of reads at the 24 SNP locations
}

