% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compIdent_format.R
\name{compIdent_format}
\alias{compIdent_format}
\title{Format readcount tables from compIdent}
\usage{
compIdent_format(x)
}
\arguments{
\item{x}{Named list of data frames with rows of the data frame corresponding
to target locations.}
}
\value{
Formated data frame
}
\description{
Format readcount tables from compIdent for input into compIdent_buildMain
}
