% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/construct_gene.R
\name{construct_gene}
\alias{construct_gene}
\title{Construct gene information}
\usage{
construct_gene(gene, domain_data, length)
}
\arguments{
\item{gene}{character string specifying gene name}

\item{domain_data}{object of class data frame specifying protien domain information, obtained from fetchDomain}

\item{length}{integer specifying length of transcript in amino acids}
}
\value{
object of class data frame giving gene and domain information
}
\description{
Build gene for input into build_lolli
}

