% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/correct_spread_coord_points.R
\name{correct_spread_coord_points}
\alias{correct_spread_coord_points}
\title{correct spread coordinates}
\usage{
correct_spread_coord_points(x)
}
\arguments{
\item{x}{list}
}
\value{
numeric vector of spread points
}
\description{
correct skewing of coord points introduced by spread_coord_points for
}

