% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covBars.R
\name{covBars}
\alias{covBars}
\title{Construct coverage cohort plot}
\usage{
covBars(x, col = NULL, plot_title = NULL, background = "grey90",
  x_lab_size = 12, y_lab_size = 12, facet_lab_size = 10, layers = NULL)
}
\arguments{
\item{x}{object of class data frame containing rows for the coverage and columns the sample names}

\item{col}{vector of colors for the coverage bars}
}
\value{
ggplot object
}
\description{
given a matrix construct a plot to display coverage as percentage bars for a group of samples
}

