% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covBars_buildMain.R
\name{covBars_buildMain}
\alias{covBars_buildMain}
\title{Construct coverage cohort plot}
\usage{
covBars_buildMain(data_frame, col, plot_title = NULL, x_lab_size = 12,
  y_lab_size = 12, facet_lab_size = 10, layers = NULL)
}
\arguments{
\item{data_frame}{object of class data frame containing columns depth,
sample, bp}

\item{col}{vector of colors for the coverage bars}

\item{plot_title}{character string for title of plot}

\item{x_lab_size}{integer specifying the size of the X label}

\item{y_lab_size}{integer specifying the size of the Y label}

\item{facet_lab_size}{integer specifying the size of the faceted labels}

\item{layers}{Additional layers to be plotted, can be a theme but must be a
ggplot layer}
}
\value{
ggplot object
}
\description{
given a data frame construct a plot to display coverage as percentage bars
for a group of samples
}

