% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covBars_qual.R
\name{covBars_qual}
\alias{covBars_qual}
\title{Construct coverage cohort plot}
\usage{
covBars_qual(x)
}
\arguments{
\item{x}{object of class matrix containing rows for the coverage and columns
the sample names}
}
\value{
a list of data frame and color vector
}
\description{
given a matrix construct a plot to display coverage as percentage bars for a
group of samples
}
