% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenVisR.R
\docType{data}
\name{cytoGeno}
\alias{cytoGeno}
\title{cytogenetic band information for select genomes}
\format{a data frame with 3207 observations and 6 variables}
\usage{
data(cytoGeno)
}
\description{
A data set containing cytogenetic band information for all chromosomes in the following genomes "hg38", "hg19", "mm10", "mm9", "rn5", obtained from the UCSC sql database at genome-mysql.cse.ucsc.edu
}
\keyword{datasets}

