% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dodge_coord_x.R
\name{dodge_coord_x}
\alias{dodge_coord_x}
\title{dodge coordinates}
\usage{
dodge_coord_x(x, spread_degree)
}
\arguments{
\item{x}{numeric vector of position coordinates on x axis}
}
\value{
numeric vector of dodged position coordinates on x axis
}
\description{
given amino acid position coordinates on, before grouping and dodge on x axis
}

