% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extr_ggplotGrob_width.R
\name{extr_ggplotGrob_width}
\alias{extr_ggplotGrob_width}
\title{extract ggplotGrob width}
\usage{
extr_ggplotGrob_width(x)
}
\arguments{
\item{x}{ggplotGrob object}
}
\value{
ggplotGrob width parameters
}
\description{
extract plot width of ggplotGrob object
}

