% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extrcds.R
\name{extrcds}
\alias{extrcds}
\title{Extract CDS}
\usage{
extrcds(txdb, gr, reduce = FALSE, gaps = FALSE)
}
\arguments{
\item{txdb}{A TxDb object for a genome}

\item{gr}{A Granges object specifying the region of interest}

\item{reduce}{Boolean specifying whether to collapse isoforms}

\item{gaps}{Boolean specifying whether to report space between CDS instead of CDS}
}
\value{
Object of class Granges list
}
\description{
Extract CDS coordinates within a GRanges object given a transcription database
}

