% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extrfiveUTR.R
\name{extrfiveUTR}
\alias{extrfiveUTR}
\title{Extract 5' UTR}
\usage{
extrfiveUTR(txdb, gr, reduce = FALSE, gaps = FALSE)
}
\arguments{
\item{txdb}{A TxDb object for a genome}

\item{gr}{A Granges object specifying the region of interest}

\item{reduce}{Boolean specifying whether to collapse isoforms}

\item{gaps}{Boolean specifying whether to report space between 5'UTR instead of 5'UTR}
}
\value{
Object of class Granges list
}
\description{
Extract 5' UTR coordinates within a GRanges object given a transcription database
}

