% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extrthreeUTR.R
\name{extrthreeUTR}
\alias{extrthreeUTR}
\title{Extract 3' UTR}
\usage{
extrthreeUTR(txdb, gr, reduce = FALSE, gaps = FALSE)
}
\arguments{
\item{txdb}{A TxDb object for a genome}

\item{gr}{A Granges object specifying the region of interest}

\item{reduce}{Boolean specifying whether to collapse isoforms}

\item{gaps}{Boolean specifying whether to report space between 3'UTR instead of 3'UTR}
}
\value{
Object of class Granges list
}
\description{
Extract 3' UTR coordinates within a GRanges object given a transcription database
}

