% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetchDomain.R
\name{fetchDomain}
\alias{fetchDomain}
\title{fetch protein domains}
\usage{
fetchDomain(uniprotID)
}
\arguments{
\item{uniprotID}{String specifying uniprot ID}
}
\value{
data frame of protien domains and start/stop coordinates
}
\description{
Retrieve protein domains from interprot database given uniprot ID
}

