% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetchfiveUTRlen.R
\name{fetchfiveUTRlen}
\alias{fetchfiveUTRlen}
\title{fetch 5'UTR length}
\usage{
fetchfiveUTRlen(enstID, dataset = "hsapiens_gene_ensembl")
}
\arguments{
\item{enstID}{a character string specifying the ensembl transcript ID}

\item{dataset}{a character string specifying the name of the biomart dataset to use}
}
\value{
an integer corresponding to the 5'UTR length
}
\description{
Retrieve 5'UTR length for a given ensembl transcript
}

