% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format_intronic_space.R
\name{format_intronic_space}
\alias{format_intronic_space}
\title{format intronic space}
\usage{
format_intronic_space(introns, exons, gr)
}
\arguments{
\item{introns}{list of data frames containing intronic features}

\item{exons}{list of data frames contatining exonic/UTR features}

\item{Granges}{object specifying the ROI}
}
\value{
Object of class data frame
}
\description{
format intronic space
}

