% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatIntronic.R
\name{formatIntronic}
\alias{formatIntronic}
\title{format intronic region}
\usage{
formatIntronic(txdb, gr)
}
\arguments{
\item{txdb}{A TxDb object for a genome}

\item{gr}{object of class GRanges specifying the ROI}
}
\value{
list of data frames
}
\description{
given a Granges object obtain and format the intronic region
}

