% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatUTR.R
\name{formatUTR}
\alias{formatUTR}
\title{format UTR}
\usage{
formatUTR(txdb, gr, genome, reduce = FALSE)
}
\arguments{
\item{txdb}{A TxDb object for a genome}

\item{gr}{A Granges object to format}

\item{genome}{Object of class BSgenome specifying the genome for GC content calculation}

\item{reduce}{Boolean specifying whether to collapse isoforms in the Granges object ROI}
}
\value{
Object of class data frame
}
\description{
given a Granges object specifying a region of interest, format into a form recognizable by ggplot2
}

