% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCov.R
\name{genCov}
\alias{genCov}
\title{produce a coverage plot}
\usage{
genCov(x, txdb, gr, genome, reduce = FALSE, gene.colour = NULL,
  gene_name = "Gene", gene.layers = NULL, label.bg_fill = "black",
  label.text_fill = "white", label.border = "black", label.size = 10,
  label.width_ratio = c(1, 10), cov.colour = "blue",
  cov.plot_type = "line", cov.layers = NULL, base = c(10, 2, 2),
  transform = c("Intron", "CDS", "UTR"), gene.plot_transcript_name = TRUE,
  gene.transcript_name_size = 4, isoformSel = NULL)
}
\arguments{
\item{x}{named list containing data frames with columns "end" and "cov" consisting of read pileups at bases of interest}

\item{txdb}{A TxDb object for a genome}

\item{gr}{A Granges object specifying a region of interest}

\item{genome}{Object of class BSgenome specifying the genome}

\item{reduce}{Boolean specifying whether to collapse isoforms in the ROI}

\item{gene.colour}{character string specifying the colour of the gene to be plotted}

\item{gene_name}{character string specifying the name of the gene or ROI}

\item{gene.layers}{additional ggplot2 layers for the gene plot}

\item{label.bg_fill}{character string giving the colour to fill the label}

\item{label.text_fill}{character string giving the colour to fill label text}

\item{label.border}{character string specifying the colour to fill the border of the label}

\item{label.size}{integer specifying the size of the text within the label}

\item{label.width_ratio}{integer vector of length 2 giving the ratio of track labels to plot}

\item{cov.colour}{character string specifying the colour of the data in the coverage plot}

\item{cov.plot_type}{character string specifying one of line, area or bar for coverage data display}

\item{cov.layers}{additional ggplot2 layers for the coverage plot}

\item{base}{A vector of log bases to transform the data, corresponding to the elements of transform}

\item{transform}{A vector of strings designating what objects to log transform accepts "Intron", "CDS", and "UTR"}

\item{gene.plot_transcript_name}{Boolean specifying whether to plot the transcript name}

\item{gene.transcript_name_size}{Integer specifying the size of the transcript name text}

\item{isoformSel}{Character vector giving the names (from the txdb object) of isoforms within the region of interest to display}
}
\value{
ggplot object
}
\description{
produce a coverage plot displaying gene and coverage information for a region of interest
}
\examples{
# Load transcript meta data
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene

# Load BSgenome
library(BSgenome.Hsapiens.UCSC.hg19)
genome <- BSgenome.Hsapiens.UCSC.hg19

# Define a region of interest
gr <- GRanges(seqnames=c("chr10"),
ranges=IRanges(start=c(89622195), end=c(89729532)), strand=strand(c("+")))

# Define preloaded coverage data as named list
data <- list("PTEN" = ptenCOV)

# Call genCov
genCov(data, txdb, gr, genome, gene.transcript_name_size=3)
}

