% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genCov_alignPlot.R
\name{genCov_alignPlot}
\alias{genCov_alignPlot}
\title{align plots on an axis}
\usage{
genCov_alignPlot(plot_list, axis = "both")
}
\arguments{
\item{plot_list}{list of ggplot objects}

\item{axis}{character string to specify the axis to align plotting space on,
one of both, width, height}
}
\value{
ggplotGrob object
}
\description{
given a list of plots, align them on plotting space
}

