% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genCov_assign_ggplotGrob_height.R
\name{genCov_assign_ggplotGrob_height}
\alias{genCov_assign_ggplotGrob_height}
\title{assign ggplotGrob height}
\usage{
genCov_assign_ggplotGrob_height(x, max_height)
}
\arguments{
\item{x}{ggplotGrob object}

\item{max_height}{grob object specifying width to reassign ggplotGrob with}
}
\value{
ggplotGrob
}
\description{
assign height of ggplotGrob object
}

