% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCov_assign_ggplotGrob_width.R
\name{genCov_assign_ggplotGrob_width}
\alias{genCov_assign_ggplotGrob_width}
\title{assign ggplotGrob width}
\usage{
genCov_assign_ggplotGrob_width(x, max_width)
}
\arguments{
\item{x}{ggplotGrob object}

\item{max_width}{grob object specifying width to reassign ggplotGrob with}
}
\value{
ggplotGrob
}
\description{
assign width of ggplotGrob object
}
