% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCov_buildCov.R
\name{genCov_buildCov}
\alias{genCov_buildCov}
\title{build coverage plot}
\usage{
genCov_buildCov(data_frame, x_limits = NULL, display_x_axis = TRUE,
  colour = "blue", plot_type = "line", layers = NULL)
}
\arguments{
\item{data_frame}{an object of class data frame containing columns
stop and cov}

\item{x_limits}{vector giving x-axis limits for plot, inferred from data
if not specified}

\item{display_x_axis}{boolean specifying whether to plot x-axis labels}

\item{colour}{character string specifying the color of the data in the plot}

\item{plot_type}{character string specifying one of line, area, bar for
data display}

\item{layers}{additional ggplot2 layers to plot}
}
\value{
ggplot object
}
\description{
given data build a coverage plot to represent the data
}

