% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCov_buildTrack.R
\name{genCov_buildTrack}
\alias{genCov_buildTrack}
\title{build label for plot}
\usage{
genCov_buildTrack(name, bg_fill = "black", text_fill = "white",
  border = "black", size = 10)
}
\arguments{
\item{name}{character string giving the name of the track}

\item{bg_fill}{character string giving the colour to fill the label}

\item{text_fill}{character string giving the colour to fill the text}

\item{border}{character string specifying the colour to fill the border of
the label}

\item{size}{integer specifying the size of the text within the label}
}
\value{
ggplot object
}
\description{
given a name create a label
}

