% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCov_extr_ggplotGrob_height.R
\name{genCov_extr_ggplotGrob_height}
\alias{genCov_extr_ggplotGrob_height}
\title{extract ggplotGrob height}
\usage{
genCov_extr_ggplotGrob_height(x)
}
\arguments{
\item{x}{ggplotGrob object}
}
\value{
ggplotGrob height parameters
}
\description{
extract plot height of ggplotGrob object
}

