% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCov_extr_ggplotGrob_width.R
\name{genCov_extr_ggplotGrob_width}
\alias{genCov_extr_ggplotGrob_width}
\title{extract ggplotGrob width}
\usage{
genCov_extr_ggplotGrob_width(x)
}
\arguments{
\item{x}{ggplotGrob object}
}
\value{
ggplotGrob width parameters
}
\description{
extract plot width of ggplotGrob object
}
