% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCov_qual.R
\name{genCov_qual}
\alias{genCov_qual}
\title{Perform quality control on genCov data}
\usage{
genCov_qual(x = x, txdb = txdb, gr = gr, genome = genome)
}
\arguments{
\item{x}{named list containing data frames with columns end and cov}

\item{txdb}{A TxDb object for a genome}

\item{gr}{A Granges object specifying a region of interest}

\item{genome}{Object of class BSgenome specifying the genome}
}
\value{
a list of objects passing basic quality control
}
\description{
Ensure data input into genCov is of the proper type and format
}
