% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCov_trackViz.R
\name{genCov_trackViz}
\alias{genCov_trackViz}
\title{Overlay tracks with plots}
\usage{
genCov_trackViz(..., bgFill = "black", textFill = "white",
  border = "black", size = 10, axis_align = "none", widthRatio = c(1,
  10), list = TRUE)
}
\arguments{
\item{...}{named list of ggplot2 plots}

\item{bgFill}{character string giving the colour to fill the label}

\item{textFill}{character string giving the colour to fill the text}

\item{border}{character string specifying the colour to fill the border of
the label}

\item{size}{integer specifying the size of the text within the label}

\item{axis_align}{character string specifying axis to align plotting space
on, one of 'both', 'height', 'width', 'none'}

\item{widthRatio}{vector of length 2 giving the ratio of track labels to
plots}

\item{list}{boolean specifying whether plots are in a named list or specified
individually via ...}
}
\value{
ggplotGrob object
}
\description{
given a named list of plots, display them on tracks
}

