% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gene_sort.R
\name{gene_sort}
\alias{gene_sort}
\title{sort MAF File by gene}
\usage{
gene_sort(data_frame)
}
\arguments{
\item{data_frame}{a data frame in MAF format}
}
\value{
a data frame sorted such that genes with more mutations are ordered first
}
\description{
order a MAF file ranking genes with more mutations higher
}

