% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneViz_buildGene.R
\name{geneViz_buildGene}
\alias{geneViz_buildGene}
\title{build gene plot}
\usage{
geneViz_buildGene(data_frame, display_x_axis = TRUE, x_limits = NULL,
  gene_colour = NULL, transcript_name = FALSE, transcript_name_size = 4,
  layers = NULL)
}
\arguments{
\item{data_frame}{an object of class data frame specifying gene feature
information}

\item{display_x_axis}{Boolean specifying whether to display X axis coordinate
values}

\item{x_limits}{vector specifying x-axis limits of plot}

\item{gene_colour}{character specifying colour of gene to be plotted}

\item{transcript_name}{Boolean specifying whether to plot USCS transcript
names}

\item{transcript_name_size}{Integer specifying the size of the transcript
name text}

\item{layers}{additional ggplot2 layers to plot}
}
\value{
ggplot object
}
\description{
given a data frame with gene feature information build the ggplot2 object
}
