% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneViz_cdsFromTXID.R
\name{geneViz_cdsFromTXID}
\alias{geneViz_cdsFromTXID}
\title{cdsFromTXID}
\usage{
geneViz_cdsFromTXID(txdb, txid)
}
\arguments{
\item{txdb}{A TxDb object for a genome}

\item{txid}{A list of TXIDs}
}
\value{
Object of class Granges
}
\description{
Return CDS coordinates as a GRanges object given transcript IDs
}

