% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneViz_extrCDS.R
\name{geneViz_extrCDS}
\alias{geneViz_extrCDS}
\title{Extract CDS}
\usage{
geneViz_extrCDS(txdb = NULL, gr = NULL, reduce = FALSE, gaps = FALSE)
}
\arguments{
\item{txdb}{A TxDb object for a genome}

\item{gr}{A Granges object specifying the region of interest}

\item{reduce}{Boolean specifying whether to collapse isoforms}

\item{gaps}{Boolean specifying whether to report space between CDS instead of
CDS}
}
\value{
Object of class Granges list
}
\description{
Extract CDS coordinates within a GRanges object given a transcription
database
}

