% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneViz_extrUTR.R
\name{geneViz_extrUTR}
\alias{geneViz_extrUTR}
\title{Extract UTR}
\usage{
geneViz_extrUTR(txdb = txdb, gr = gr, reduce = FALSE, gaps = FALSE)
}
\arguments{
\item{txdb}{A TxDb object for a genome}

\item{gr}{A Granges object specifying the region of interest}

\item{reduce}{Boolean specifying whether to collapse isoforms}

\item{gaps}{Boolean specifying whether to report space between UTR instead of UTR}
}
\value{
Object of class Granges list
}
\description{
Extract UTR coordinates within a GRanges object given a transcription database
}

