% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneViz_formatUTR.R
\name{geneViz_formatUTR}
\alias{geneViz_formatUTR}
\title{format UTR}
\usage{
geneViz_formatUTR(txdb = NULL, gr = NULL, genome = NULL, reduce = FALSE)
}
\arguments{
\item{txdb}{A TxDb object for a genome}

\item{gr}{A Granges object to format}

\item{genome}{Object of class BSgenome specifying the genome for GC content
calculation}

\item{reduce}{Boolean specifying whether to collapse isoforms in the Granges
object ROI}
}
\value{
Object of class data frame
}
\description{
given a Granges object specifying a region of interest, format into a form
recognizable by ggplot2
}

