% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneViz_Granges2dataframe.R
\name{geneViz_Granges2dataframe}
\alias{geneViz_Granges2dataframe}
\title{Convert Granges object to dataframe}
\usage{
geneViz_Granges2dataframe(gr)
}
\arguments{
\item{gr}{A Granges object to convert to data frame}
}
\value{
Object of class data frame
}
\description{
Convert a Granges object with meta data GC content to a object of class data
frame
}

