% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneViz_mapCoordSpace.R
\name{geneViz_mapCoordSpace}
\alias{geneViz_mapCoordSpace}
\title{Map regions to transformed space}
\usage{
geneViz_mapCoordSpace(master, coord)
}
\arguments{
\item{master}{an object of class data frame containing columns start, end,
width, type, trans_start, trans_end representing a master genomic region with
features from isoforms merged}

\item{coord}{an object of class data frame containing columns start and end
to map to transformed space}
}
\value{
an object of class data frame identical to coord but with extra
columns for transformed coord
}
\description{
Reference a master genomic region file to map original positions to a
transformed space
}

