% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneViz_mergeRegions.R
\name{geneViz_mergeRegions}
\alias{geneViz_mergeRegions}
\title{Create Region Table}
\usage{
geneViz_mergeRegions(gene_features, gr, base, transform)
}
\arguments{
\item{gene_features}{A dataframe specifying features of a gene}

\item{gr}{Granges object specifying the ROI}

\item{base}{A vector of log bases to transform the data, corresponding to the
elements of transform}

\item{transform}{A vector of strings designating what objects to log
transform}
}
\value{
Master region table data frame
}
\description{
Create a master region table by merging isoforms
}

