% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneViz_mergeTypes.R
\name{geneViz_mergeTypes}
\alias{geneViz_mergeTypes}
\title{Merge Typed Region Tables}
\usage{
geneViz_mergeTypes(master)
}
\arguments{
\item{master}{An unsorted dataframe of CDS and UTR elements before merging}
}
\value{
master A sorted dataframe of merged CDS and UTR elements
}
\description{
Create a master region table by merging isoforms
}

