% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_cytobands.R
\name{get_cytobands}
\alias{get_cytobands}
\title{obtain cytogenetic bands}
\usage{
get_cytobands(genome)
}
\arguments{
\item{genome}{character string giving a UCSC genome}
}
\value{
object of class data frame
}
\description{
given a genome query UCSC for cytogenetic band locations
}

