% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/lohSpec-class.R
\docType{methods}
\name{getChrBoundaries}
\alias{getChrBoundaries}
\alias{getChrBoundaries,data.table-method}
\alias{getChrBoundaries}
\title{Method getChrBoundaries}
\usage{
getChrBoundaries(object, ...)

\S4method{getChrBoundaries}{data.table}(object, genome)
}
\arguments{
\item{object}{Object of class VarScanFormat}

\item{...}{additional arguments to passed}

\item{genome}{character specifying which genome to use}

\item{object}{of class lohData}
}
\value{
Data.table with chr and start/stop positions
}
\description{
Method getChrBoundaries
}
