% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Clinical-class.R,
%   R/MutSpectra-class.R
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,Clinical-method}
\alias{getData}
\alias{getData,ClinicalData-method}
\alias{getData}
\alias{getData,MutSpectraPrimaryData-method}
\alias{getData}
\title{Method getData}
\usage{
getData(object, ...)

\S4method{getData}{Clinical}(object, ...)

\S4method{getData}{ClinicalData}(object, ...)

\S4method{getData}{MutSpectraPrimaryData}(object, ...)
}
\arguments{
\item{object}{Object of class Clinical,}

\item{...}{additional arguments to passed}
}
\description{
Method getData
}
\details{
The getData method is an accessor function used to access data held
in GenVisR objects.
}
