% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/VEP_Virtual-class.R,
%   R/VEP-class.R
\docType{methods}
\name{getDescription}
\alias{getDescription}
\alias{getDescription,VEP_Virtual-method}
\alias{getDescription,VEP-method}
\title{Method getDescription}
\usage{
getDescription(object, ...)

\S4method{getDescription}{VEP_Virtual}(object, ...)

\S4method{getDescription}{VEP}(object, ...)
}
\arguments{
\item{object}{Object of class VEP}

\item{...}{additional arguments to passed}
}
\description{
Method getDescription
}
