% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MutSpectra-class.R,
%   R/Rainfall-class.R, R/Waterfall-class.R, R/combinedCnLoh-class.R,
%   R/lohSpec-class.R
\docType{methods}
\name{getGrob}
\alias{getGrob}
\alias{getGrob,MutSpectraPlots-method}
\alias{getGrob}
\alias{getGrob,MutSpectra-method}
\alias{getGrob}
\alias{getGrob,RainfallPlots-method}
\alias{getGrob}
\alias{getGrob,Rainfall-method}
\alias{getGrob}
\alias{getGrob,WaterfallPlots-method}
\alias{getGrob}
\alias{getGrob,Waterfall-method}
\alias{getGrob}
\alias{getGrob,cnLohPlots-method}
\alias{getGrob}
\alias{getGrob,cnLoh-method}
\alias{getGrob}
\alias{getGrob,lohSpecPlots-method}
\alias{getGrob}
\alias{getGrob,lohSpec-method}
\alias{getGrob}
\title{Method getGrob}
\usage{
getGrob(object, ...)

\S4method{getGrob}{MutSpectraPlots}(object, index = 1, ...)

\S4method{getGrob}{MutSpectra}(object, index = 1, ...)

\S4method{getGrob}{RainfallPlots}(object, index = 1, ...)

\S4method{getGrob}{Rainfall}(object, index = 1, ...)

\S4method{getGrob}{WaterfallPlots}(object, index = 1, ...)

\S4method{getGrob}{Waterfall}(object, index = 1, ...)

\S4method{getGrob}{cnLohPlots}(object, index = 1, ...)

\S4method{getGrob}{cnLoh}(object, index = 1, ...)

\S4method{getGrob}{lohSpecPlots}(object, index = 1, ...)

\S4method{getGrob}{lohSpec}(object, index = 1, ...)
}
\arguments{
\item{object}{Object of clas MutSpectra}

\item{...}{additional arguments to passed}

\item{index}{integer specifying the plot index to extract}
}
\description{
Method getGrob
}
