% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/VEP_Virtual-class.R,
%   R/VEP-class.R
\docType{methods}
\name{getHeader}
\alias{getHeader}
\alias{getHeader,VEP_Virtual-method}
\alias{getHeader}
\alias{getHeader,VEP-method}
\alias{getHeader}
\title{Method getHeader}
\usage{
getHeader(object, ...)

\S4method{getHeader}{VEP_Virtual}(object, ...)

\S4method{getHeader}{VEP}(object, ...)
}
\arguments{
\item{object}{Object of class VEP}

\item{...}{additional arguments to passed}
}
\description{
Method getHeader
}
