% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getLohData}
\alias{getLohData}
\title{Method getLohData}
\usage{
getLohData(object, ...)
}
\arguments{
\item{object}{Object of class VarScanFormat}

\item{...}{additional arguments to passed}
}
\description{
Method getLohData
}
