% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/lohSpec-class.R
\docType{methods}
\name{getLohFreq}
\alias{getLohFreq}
\alias{getLohFreq,data.table-method}
\alias{getLohFreq}
\title{Method getLohFreq}
\usage{
getLohFreq(object, ...)

\S4method{getLohFreq}{data.table}(object, plotALohCutoff, chrData, verbose, ...)
}
\arguments{
\item{object}{Object of class VarScanFormat}

\item{...}{additional arguments to passed}

\item{chrData}{of class data.table}

\item{object}{of class lohData}
}
\description{
Method getLohFreq
}
