% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/combinedCnLoh-class.R,
%   R/lohSpec-class.R
\docType{methods}
\name{getLohStepCalculation}
\alias{getLohStepCalculation}
\alias{getLohStepCalculation}
\alias{getLohStepCalculation,list-method}
\alias{getLohStepCalculation}
\title{Method getLohStepCalculation}
\usage{
getLohStepCalculation(object, ...)

\S4method{getLohStepCalculation}{list}(object, step, ...)

\S4method{getLohStepCalculation}{list}(object, step, ...)
}
\arguments{
\item{object}{Object of class VarScanFormat}

\item{...}{additional arguments to passed}

\item{step}{integer}

\item{object}{of class data.table}

\item{object}{of class lohData}

\item{step}{integer}
}
\description{
Method getLohStepCalculation
}
