% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GMS-class.R,
%   R/MutationAnnotationFormat-class.R, R/VEP-class.R
\name{getPath}
\alias{getPath}
\alias{getPath,GMS-method}
\alias{getPath,MutationAnnotationFormat-method}
\alias{getPath,VEP-method}
\title{Method getPath}
\usage{
getPath(object, ...)

\S4method{getPath}{GMS}(object, ...)

\S4method{getPath}{MutationAnnotationFormat}(object, ...)

\S4method{getPath}{VEP}(object, ...)
}
\arguments{
\item{object}{Object of class VEP, GMS, or MutationAnnotationFormat}

\item{...}{additional arguments to passed}
}
\description{
Method getPath
}
