% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GMS-class.R,
%   R/MutationAnnotationFormat-class.R
\docType{methods}
\name{getSample}
\alias{getSample}
\alias{getSample,GMS-method}
\alias{getSample,GMS}
\alias{getSample,MutationAnnotationFormat-method}
\alias{getSample,MutationAnnotationFormat}
\title{Method getSample}
\usage{
getSample(object, ...)

\S4method{getSample}{GMS}(object, ...)

\S4method{getSample}{MutationAnnotationFormat}(object, ...)
}
\arguments{
\item{object}{Object of class VEP, GMS, or MutationAnnotationFormat}

\item{...}{additional arguments to passed}
}
\description{
Method getSample
}
