% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/VarScanFormat_Virtual-class.R
\docType{methods}
\name{getVarScan}
\alias{getVarScan}
\alias{getVarScan,VarScanFormat_Virtual-method}
\alias{getVarScan}
\title{Method getVarScan}
\usage{
getVarScan(object, ...)

\S4method{getVarScan}{VarScanFormat_Virtual}(object, ...)
}
\arguments{
\item{object}{Object of class VarScanFormat}

\item{...}{additional arguments to passed}
}
\description{
Method getVarScan
}
