% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/VCF_Virtual-class.R
\docType{methods}
\name{getVcf}
\alias{getVcf}
\alias{getVcf}
\title{Method getVcf}
\usage{
getVcf(object, ...)

\S4method{getVcf}{VCF_Virtual}(object, ...)
}
\arguments{
\item{...}{additional arguments to passed}
}
\description{
Method getVcf
}
