% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GMS-class.R,
%   R/MutationAnnotationFormat-class.R, R/VEP-class.R
\docType{methods}
\name{getVersion}
\alias{getVersion}
\alias{getVersion,GMS-method}
\alias{getVersion}
\alias{getVersion,MutationAnnotationFormat-method}
\alias{getVersion}
\alias{getVersion,VEP-method}
\alias{getVersion}
\title{Method getVersion}
\usage{
getVersion(object, ...)

\S4method{getVersion}{GMS}(object, ...)

\S4method{getVersion}{MutationAnnotationFormat}(object, ...)

\S4method{getVersion}{VEP}(object, ...)
}
\arguments{
\item{object}{Object of class VEP, GMS, or MutationAnnotationFormat}

\item{...}{additional arguments to passed}
}
\description{
Method getVersion
}
