% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMS_v4-class.R
\docType{class}
\name{GMS_v4-class}
\alias{GMS_v4-class}
\alias{GMS_v4}
\alias{GMS_v4}
\title{Class GMS_v4}
\usage{
\S4method{initialize}{GMS_v4}(.Object, gmsData)

GMS_v4(gmsData)
}
\arguments{
\item{gmsData}{data.table object containing a gms annotation file conforming
to the version 4 specifications.}
}
\description{
An S4 class to represent data in gms annotation version 4,
inherits from the GMS_Virtual class.

Initalizer method for the GMS_v4 sub-class

Constructor for the GMS_v4 sub-class
}
\section{Slots}{

\describe{
\item{\code{position}}{data.table object containing column names "chromosome_name",
"start", "stop".}

\item{\code{mutation}}{data.table object containing column names "reference",
"variant", "trv_type".}

\item{\code{sample}}{data.table object containing columns names "sample".}

\item{\code{meta}}{data.table object containing meta data.}
}}

