% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMS_Virtual-class.R
\docType{class}
\name{GMS_Virtual-class}
\alias{GMS_Virtual-class}
\title{Class GMS_Virtual}
\description{
An S4 class to act as a virtual class for GMS version sub-classes.
}
\section{Slots}{

\describe{
\item{\code{position}}{data.table object holding genomic positions.}

\item{\code{mutation}}{data.table object holding mutation status data.}

\item{\code{sample}}{data.table object holding sample data.}

\item{\code{meta}}{data.table object holding all other meta data.}
}}

