% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Granges2dataframe.R
\name{Granges2dataframe}
\alias{Granges2dataframe}
\title{Convert Granges object to dataframe}
\usage{
Granges2dataframe(gr)
}
\arguments{
\item{gr}{A Granges object to convert to data frame}
}
\value{
Object of class data frame
}
\description{
Convert a Granges object with meta data GC content to a object of class data frame
}

