% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/group_mutation_coord.R
\name{group_mutation_coord}
\alias{group_mutation_coord}
\title{group coordinates}
\usage{
group_mutation_coord(x)
}
\arguments{
\item{x}{numeric vector of position coordinates on x axis}
}
\value{
numeric vector of grouped position coordinates on x axis
}
\description{
given amino acid position coordinates on, before grouping and dodge on x axis
}

