% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenVisR.R
\docType{data}
\name{HCC1395_Germline}
\alias{HCC1395_Germline}
\title{Germline Calls}
\format{a data frame with 9200 observations and 5 variables}
\usage{
data(HCC1395_Germline)
}
\value{
Object of class data frame
}
\description{
A data set containing downsampled Germline calls originating from the HCC1395
breast cancer cell line.
}
\keyword{datasets}

