% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenVisR.R
\docType{data}
\name{hg19chr}
\alias{hg19chr}
\title{hg19 chromosome boundaries}
\format{a data frame with 24 observations and 3 variables}
\usage{
data(hg19chr)
}
\value{
Object of class data frame
}
\description{
A data set containg chromosome boundaries corresponding to hg19.
}
\keyword{datasets}
