% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideoView_buildMain.R
\name{ideoView_buildMain}
\alias{ideoView_buildMain}
\title{build chromosome}
\usage{
ideoView_buildMain(data_frame, chromosome, chr_txt_angle = chr_txt_angle,
  chr_txt_size = chr_txt_size, layers = NULL)
}
\arguments{
\item{data_frame}{a data frame with columns chrom, chromStart, chromEnd,
name, gieStain, height_min, height_max, alternate, bandcenter, text_y, arm}

\item{chromosome}{character string specifying UCSC chromosome to plot one of
chr... or all}

\item{chr_txt_angle}{integer specifying angle of text when plotting band text}

\item{chr_txt_size}{integer specifying size of text when plotting band text}

\item{layers}{additional ggplot2 layers to plot}
}
\value{
ggplot object
}
\description{
given a data frame with cytogenetic band locations plot chromosome in ggplot
}

