% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ideoView_formatCytobands.R
\name{ideoView_formatCytobands}
\alias{ideoView_formatCytobands}
\title{reformat cytogenetic band data frame}
\usage{
ideoView_formatCytobands(data_frame, chromosome)
}
\arguments{
\item{data_frame}{a data frame retrieved from UCSC giving cytogenetic
information}

\item{chromosome}{character string specifying chromosome of interest from
UCSC data frame}
}
\value{
object of class data frame
}
\description{
given a data frame of cytogenetic bands, format it for ggplot call
}

