% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideoView_qual.R
\name{ideoView_qual}
\alias{ideoView_qual}
\title{Check input to ideoView}
\usage{
ideoView_qual(x)
}
\arguments{
\item{x}{a data frame with rows representing cytogenetic bands for a genome.
The data frame should have columns "chrom", "chromStart", "chromEnd", "name",
"gieStain'.}
}
\value{
data frame
}
\description{
Check that input to ideoView is properly formatted
}

