% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intronic_transform.R
\name{intronic_transform}
\alias{intronic_transform}
\title{transform intronic space}
\usage{
intronic_transform(x)
}
\arguments{
\item{x}{object of class data frame containing columns Type, end, start}
}
\value{
Object of class data frame
}
\description{
transform intronic space
}

