% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-lohSpec.R
\name{lohSpec_buildMain}
\alias{lohSpec_buildMain}
\title{Plot LOH data}
\usage{
lohSpec_buildMain(x, dummyData, colourScheme = "inferno", plotLayer = NULL)
}
\arguments{
\item{x}{object of class dataframe with loh difference
calculations and column names "window_start", "window_stop", "chromosome",
"sample", and "loh_diff"}

\item{dummyData}{object of class dataframe with column names "chromosome",
"start", "end" specifying chromosome boundaries}

\item{colourScheme}{Character vector specifying the colour scale to use from
the viridis package. One of "viridis", "magma", "plasma", or "inferno".}

\item{plotLayer}{Valid ggplot2 layer to be added to the plot.
for the legend's parameters}
}
\value{
object of class ggplot2
}
\description{
Build a ggplot2 object displaying calculated LOH data
}
