% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/lohSpec-class.R
\docType{methods}
\name{lohSpec_buildMainPlot}
\alias{lohSpec_buildMainPlot}
\alias{lohSpec_buildMainPlot,lohData-method}
\alias{lohSpec_buildMainPlot}
\title{Method lohSpec_buildMainPlot}
\usage{
lohSpec_buildMainPlot(object, ...)

\S4method{lohSpec_buildMainPlot}{lohData}(object, gradientMidpoint,
  gradientColors, plotBLayers, verbose, ...)
}
\arguments{
\item{object}{Object of class VarScanFormat}

\item{...}{additional arguments to passed}

\item{object}{of class lohData}
}
\description{
Method lohSpec_buildMainPlot
}
