% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lohSpec_fileGlob.R
\name{lohSpec_fileGlob}
\alias{lohSpec_fileGlob}
\title{Grab data for lohSpec}
\usage{
lohSpec_fileGlob(path, fileExt, step, window_size, gender)
}
\arguments{
\item{path}{character string specifying which directory contains
the sample information stored as datasets with columns "chromosome",
"position", "n_vaf", "t_vaf", and "sample" (required if x is not specified)}

\item{fileExt}{character string specifying the file extensions of files}

\item{step}{integer with the length of divisions (bp) in chromosomes}

\item{window_size}{Integer value specifying the size of the window in base
pairs in which to calculate the mean Loss of Heterozygosity.}

\item{gender}{vector of length equal to the number of samples, consisting of
elements from the set {"M", "F"}}
}
\value{
object of class data frame from data specified in path for lohSpec
}
\description{
Look in the specified file path and grab data with the proper extension for
lohSpec
}

