% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/lohSpec-class.R
\docType{data}
\name{lohSpec_qual}
\alias{lohSpec_qual}
\alias{lohSpec_qual,data.table-method}
\alias{lohSpec_qual}
\title{Method lohSpec_qual}
\format{An object of class \code{NULL} of length 0.}
\usage{
\S4method{lohSpec_qual}{data.table}(object)
}
\arguments{
\item{object}{Object of class VarScanFormat}

\item{...}{additional arguments to passed}

\item{object}{of class lohData}
}
\value{
Data.table with quality check
}
\description{
Method lohSpec_qual
}
\keyword{datasets}
