% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lohSpec_tileCalc.R
\name{lohSpec_tileCalc}
\alias{lohSpec_tileCalc}
\title{Calculate loh difference}
\usage{
lohSpec_tileCalc(window_data, normal)
}
\arguments{
\item{window_data}{object of class data frame with columns "chromosome", 
"position", "n_vaf", "t_vaf", "sample", "bin", "window_start", "window_stop"}

\item{normal}{integer specifying the subtraction value from tumor VAF}
}
\value{
object of class dataframe containing mean LOH difference calculations
and column names "window_start", "window_stop", "chromosome", "position", 
"n_vaf", "t_vaf", "sample", "loh_diff"
}
\description{
Obtain LOH on an entire chromsomes from samples in a cohort
}

