% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lohSpec_tilePosition.R
\name{lohSpec_tilePosition}
\alias{lohSpec_tilePosition}
\title{Obtain window information}
\usage{
lohSpec_tilePosition(out, window_size)
}
\arguments{
\item{out}{object of class dataframe with columns 'chromosome', 
'position', 'n_vaf', 't_vaf', and 'sample'}

\item{window_size}{integer with the size of the sliding window (bp) to be 
applied}
}
\value{
list containing window start/stop positions for each chromosome 
from each sample to perform LOH calculations
}
\description{
Calculate window positions to perform LOH calculation
}
