% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lohSpec_tileWindow.R
\name{lohSpec_tileWindow}
\alias{lohSpec_tileWindow}
\title{Obtain LOH data}
\usage{
lohSpec_tileWindow(loh_data, window_size, normal)
}
\arguments{
\item{loh_data}{data frame with columns "chromosome", "position", "n_vaf",
"t_vaf", "sample" giving raw vaf calls for germline variants}

\item{window_size}{integer with the size of the sliding window (bp) to be 
applied}

\item{normal}{integer specifying the normal VAF frequency used in LOH 
calculations}
}
\value{
object of class dataframe containing LOH data
}
\description{
Obtain LOH heatmap on entire chromsomes from samples in a cohort
}

