% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lohView.R
\name{lohView}
\alias{lohView}
\title{plot LOH data}
\usage{
lohView(x, y = NULL, genome = "hg19")
}
\arguments{
\item{x}{object of class data frame with columns 'chromosome', 'position',
'n_vaf', 't_vaf', 'sample'}

\item{y}{object of class data frame with columns 'chromosome', 'start',
'end' specifying chromosomal boundaries for a genome assembly (optional)}

\item{genome}{character string specifying the genome assembly from which
input data is based}
}
\value{
ggplot object
}
\description{
Produce a graphic visualizing Loss of Heterozygosity in a cohort
}

