% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lohView_buildMain.R
\name{lohView_buildMain}
\alias{lohView_buildMain}
\title{construct loh plot}
\usage{
lohView_buildMain(x, y, chr, layers = NULL)
}
\arguments{
\item{x}{a data frame with columns chromosome, position, n_vaf, t_vaf, sample}

\item{y}{a data frame with columns chromosome, coordinate for plotting
chromosome boundaries}

\item{chr}{a character string specifying chromosome}

\item{layers}{additional ggplot2 layers to add}
}
\value{
ggplot2 object
}
\description{
given a loh data frame plot points in ggplot
}

