% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lohView_fileGlob.R
\name{lohView_fileGlob}
\alias{lohView_fileGlob}
\title{Grab data for lohView}
\usage{
lohView_fileGlob(path, fileExt, step, gender)
}
\arguments{
\item{path}{character string specifying which directory contains
the sample information stored as datasets with columns "chromosome",
"position", "n_vaf", "t_vaf", and "sample" (required if x is not specified)}

\item{fileExt}{character string specifying the file extensions of files}

\item{step}{integer with the length of divisions (bp) in chromosomes}
}
\value{
object of class data frame from data specified in path for lohView
}
\description{
Look in the specified file path and grab data with the proper extension for
lohView
}

