% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lohView_lohCalc.R
\name{lohView_lohCalc}
\alias{lohView_lohCalc}
\title{Calculate loh difference}
\usage{
lohView_lohCalc(window_data, out, normal)
}
\arguments{
\item{window_data}{object of class data frame with columns
'window_start' and 'window_stop}

\item{out}{object of class dataframe with columns 'chromosome',
'position', 'n_vaf', 't_vaf', and 'sample'}

\item{normal}{integer specifying the subtraction value from tumor VAF}
}
\value{
object of class dataframe containing mean LOH difference calculations
and column names "window_start", "window_stop", "chromosome", "position",
"n_vaf", "t_vaf", "sample", "loh_diff"
}
\description{
Obtain LOH on an entire chromsomes from samples in a cohort
}
