% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lohView_qual.R
\name{lohView_qual}
\alias{lohView_qual}
\title{check input to lohView}
\usage{
lohView_qual(x, y, genome)
}
\arguments{
\item{x}{object of class data frame with rows representing germline calls.
The data frame must contain columns with the following names "chromosome",
"position", "n_vaf", "t_vaf", "sample".}

\item{y}{a data frame with columns "chrom", "chromStart", "chromEnd", "name",
"gieStain"}

\item{genome}{character string specifying UCSC genome to use}
}
\value{
a list of data frames passing quality checks
}
\description{
Perform a data quality check for inputs to lohView
}

