% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lohView_tileWindow.R
\name{lohView_tileWindow}
\alias{lohView_tileWindow}
\title{Obtain LOH data}
\usage{
lohView_tileWindow(loh_data, window_size, normal)
}
\arguments{
\item{loh_data}{data frame with columns "chromosome", "position", "n_vaf",
"t_vaf", "sample" giving raw vaf calls for germline variants}

\item{window_size}{integer with the size of the sliding window (bp) to be
applied}

\item{normal}{integer specifying the normal VAF frequency used in LOH
calculations}

\item{tile}{object specifying whether the loh calculations should be tiled}
}
\value{
object of class dataframe containing LOH data
}
\description{
Obtain LOH heatmap on entire chromsomes from samples in a cohort
}

