% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lohView_tilePosition.R, R/lohView_windowPosition.R
\name{lohView_windowPosition}
\alias{lohView_tilePosition}
\alias{lohView_windowPosition}
\title{Obtain window information}
\usage{
lohView_tilePosition(out, window_size)

lohView_windowPosition(out, step, window_size)
}
\arguments{
\item{out}{object of class dataframe with columns 'chromosome',
'position', 'n_vaf', 't_vaf', and 'sample'}

\item{window_size}{integer with the size of the sliding window (bp) to be
applied}

\item{step}{integer with the length of divisions (bp) in chromosomes}

\item{out}{object of class dataframe with columns 'chromosome',
'position', 'n_vaf', 't_vaf', and 'sample'}

\item{window_size}{integer with the size of the sliding window (bp) to be
applied}
}
\value{
list containing window start/stop positions for each chromosome
from each sample to perform LOH calculations

list containing window start/stop positions for each chromosome
from each sample to perform LOH calculations
}
\description{
Calculate window positions to perform LOH calculation

Calculate window positions to perform LOH calculation
}
