% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lolliplot.tanscriptID2uniprotID.R
\name{lolliplot.transcriptID2uniprotID}
\alias{lolliplot.transcriptID2uniprotID}
\title{Convert trascript ID to uniprot ID}
\usage{
lolliplot.transcriptID2uniprotID(transcriptID, up)
}
\arguments{
\item{transcriptID}{String specifying ensembl transcript ID}

\item{up}{uniprot data object from UniProt.ws}
}
\value{
String specifying uniprot ID
}
\description{
Convert an ensembl transcript ID to the corresponding uniprot ID
}

