% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lolliplot_AA2sidechain.R
\name{lolliplot_AA2sidechain}
\alias{lolliplot_AA2sidechain}
\title{Convert AA to side chain classification}
\usage{
lolliplot_AA2sidechain(x)
}
\arguments{
\item{x}{Character of length 1 giving the 1 letter amino acid code}
}
\value{
Object of class character
}
\description{
Given the 1 letter code an amino acid, return the side chian classification
}

