% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{lolliplot_Codon2AA}
\alias{lolliplot_Codon2AA}
\title{Convert Codon to AA}
\usage{
lolliplot_Codon2AA(x)
}
\arguments{
\item{x}{Character string of length 1 giving the DNA codon to convert}
}
\value{
Character corresponding to the residue for the given codon
}
\description{
Convert a Codon to the appropriate amino acid
}
