% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolliplot_constructGene.R
\name{lolliplot_constructGene}
\alias{lolliplot_constructGene}
\title{Construct gene information}
\usage{
lolliplot_constructGene(gene, domain_data, length)
}
\arguments{
\item{gene}{character string specifying gene name}

\item{domain_data}{object of class data frame specifying protien domain
information, obtained from lolliplot_fetchDomain, should contain columns
giving "description", "start", "end"}

\item{length}{integer specifying length of transcript in amino acids}
}
\value{
object of class data frame giving gene and domain information
}
\description{
Build gene for input into lolliplot_buildMain
}

