% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolliplot_DNAconv.R
\name{lolliplot_DNAconv}
\alias{lolliplot_DNAconv}
\title{Convert DNA character string}
\usage{
lolliplot_DNAconv(x, to = "residue")
}
\arguments{
\item{x}{Character string of nucleotides to convert}

\item{to}{Character string specifying conversion to do, one of "codon", 
"residue", "sidechain"}
}
\value{
Converted string of nucleotides as character vector
}
\description{
Convert a character string of nucleotides to amino acids or side chain class
}

