% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lolliplot_dodgeCoordX.R
\name{lolliplot_dodgeCoordX}
\alias{lolliplot_dodgeCoordX}
\title{dodge coordinates}
\usage{
lolliplot_dodgeCoordX(x, rep.fact = 5000, rep.dist.lmt = 500,
  attr.fact = 0.1, adj.max = 0.1, adj.lmt = 0.5, iter.max = 50000)
}
\arguments{
\item{x}{numeric vector of position coordinates on x axis}

\item{rep.fact}{repulsive factor for plotted mutations observed track}

\item{rep.dist.lmt}{repulsive distance limit for plotted mutations observed
track}

\item{attr.fact}{attraction factor for plotted mutations observed track}

\item{adj.max}{maximum position change for each iteration observed track}

\item{adj.lmt}{position adjustment limit which simulation stops observed
track}

\item{iter.max}{maximum iterations beyond which to stop the simulation
observed track}
}
\value{
numeric vector of dodged position coordinates on x axis
}
\description{
given amino acid position coordinates on, before grouping and dodge on x axis
}

