% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolliplot_dodgeCoordY.R
\name{lolliplot_dodgeCoordY}
\alias{lolliplot_dodgeCoordY}
\title{dodge coordinates}
\usage{
lolliplot_dodgeCoordY(x, track = "top")
}
\arguments{
\item{x}{data frame containing columns coord_x_dodge}

\item{track}{character vector, one of "top", "bottom" specifying whether to
dodge in a positive or negative fashion}
}
\value{
numeric vector of dodged position coordinates on y axis
}
\description{
given a data frame, dodge x coordinates ontop of each other
}

