% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolliplot_groupMutcoord.R
\name{lolliplot_groupMutcoord}
\alias{lolliplot_groupMutcoord}
\title{group coordinates}
\usage{
lolliplot_groupMutcoord(x)
}
\arguments{
\item{x}{numeric vector of position coordinates on x axis}
}
\value{
numeric vector of grouped position coordinates on x axis
}
\description{
given amino acid position coordinates on, before grouping and dodge on x axis
}

