% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{lolliplot_reduceLolli}
\alias{lolliplot_reduceLolli}
\title{Reduce Lolli}
\usage{
lolliplot_reduceLolli(x, max = NULL)
}
\arguments{
\item{x}{Data frame with column name mutation_coord to reduce lollis on}

\item{max}{Integer specifying the maximum number of lollis to allow}
}
\value{
Object of class data frame taking the reduced form of x
}
\description{
Reduce lollis stacked ontop of each other to the amount specified
}
