% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolliplot_transcriptID2codingSeq.R
\name{lolliplot_transcriptID2codingSeq}
\alias{lolliplot_transcriptID2codingSeq}
\title{fetch protein length}
\usage{
lolliplot_transcriptID2codingSeq(transcriptID, species = "hsapiens")
}
\arguments{
\item{transcriptID}{character string giving ensembl transcript id}

\item{species}{character string to use when searching for ensemblMart dataset}
}
\value{
length in residues of ensembl transcript id
}
\description{
Retrieve protein length from ensembl database given enseml transcript id
}

