% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lolliplot_transcriptID2length.R
\name{lolliplot_transcriptID2length}
\alias{lolliplot_transcriptID2length}
\title{fetch protein length}
\usage{
lolliplot_transcriptID2length(transcriptID)
}
\arguments{
\item{transcriptID}{character string giving ensembl transcript id}
}
\value{
length in residues of ensembl transcript id
}
\description{
Retrieve protein length from ensembl database given enseml transcript id
}

