% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MAF_to_anno.R
\name{MAF_to_anno}
\alias{MAF_to_anno}
\title{Convert MAF File}
\usage{
MAF_to_anno(x, label_col)
}
\arguments{
\item{x}{a data frame in MAF format}

\item{label_col}{Character string specifying the column name of a label column}
}
\value{
a data frame coerced from MAF to TGI format
}
\description{
Convert columns of a mutation annotation file "MAF" into a format recognizable by internal functions
}

