% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map_coord_space.R
\name{map_coord_space}
\alias{map_coord_space}
\title{Map regions to transformed space}
\usage{
map_coord_space(master, coord)
}
\arguments{
\item{master}{an object of class data frame containing columns start, end, width, type, trans_start, trans_end representing a master genomic region with features from isoforms merged}

\item{coord}{an object of class data frame containing columns start and end to map to transformed space}
}
\value{
an object of class data frame identical to coord but with extra columns for transformed coord
}
\description{
Reference a master genomic region file to map original positions to a transformed space
}

