% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergeRegions.R
\name{mergeRegions}
\alias{mergeRegions}
\title{Create Region Table}
\usage{
mergeRegions(gene_features, gr, base = exp(1))
}
\arguments{
\item{gr}{Granges object specifying the ROI}

\item{txdb}{A TxDb object for a genome}
}
\value{
Master region table data frame
}
\description{
Create a master region table by merging isoforms
}

