% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergeTypeRegions.R
\name{mergeTypeRegions}
\alias{mergeTypeRegions}
\title{Create Typed Region Table}
\usage{
mergeTypeRegions(type.master)
}
\arguments{
\item{type.master}{A dataframe of all elements of a certain type, such as CDS}
}
\value{
type.master A dataframe of merged elements of a certain type
}
\description{
Create a master region table by merging isoforms
}

