% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MGI_to_anno.R
\name{MGI_to_anno}
\alias{MGI_to_anno}
\title{Convert MGI File}
\usage{
MGI_to_anno(x)
}
\arguments{
\item{x}{a data frame in MGI internal format}
}
\value{
a data frame coerced from MGI to internal annotation format
}
\description{
Convert columns of a mutation annotation file "MGI" into a format recognizable by internal functions
}

