% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_align.R
\name{multi_align}
\alias{multi_align}
\title{align CN/LOH plots on x axis}
\usage{
multi_align(p1, p2)
}
\arguments{
\item{p1}{ggplot object of chromosome}

\item{p2}{ggplot object of CN or LOH}
}
\value{
ggplot object
}
\description{
given a chromosome and CN/LOH plot align plot widths
}

