% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multi_chrBound.R
\name{multi_chrBound}
\alias{multi_chrBound}
\title{retrieve and format CN_cohort plot supplemental data}
\usage{
multi_chrBound(x)
}
\arguments{
\item{x}{data frame containg columns chromosome, start, end}
}
\value{
object of class data frame formatted to internal specifications
}
\description{
given a genome obtain Start and Stop positions for all chromosomes in the
genome
}

