% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_cytobandRet.R
\name{multi_cytobandRet}
\alias{multi_cytobandRet}
\title{Retrieve cytogenetic bands}
\usage{
multi_cytobandRet(genome)
}
\arguments{
\item{genome}{character string giving a UCSC genome}
}
\value{
object of class data frame
}
\description{
given a genome query UCSC for cytogenetic band locations
}
