% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_selectOut.R
\name{multi_selectOut}
\alias{multi_selectOut}
\title{Choose output}
\usage{
multi_selectOut(data, plot, out = "plot", draw = "FALSE")
}
\arguments{
\item{data}{Data object to output}

\item{plot}{Plot object to output}

\item{out}{Character vector specifying the the object to output, one of
"data", "grob", or "plot".}

\item{draw}{Boolean specifying if the input to plot needs to be drawn}
}
\value{
One of the following, a list of dataframes containing data to be
plotted, a grob object, or a plot.
}
\description{
Selector for choosing output for GenVisR functions
}
