% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_subsetChr.R
\name{multi_subsetChr}
\alias{multi_subsetChr}
\title{subset based on chr}
\usage{
multi_subsetChr(x, chr)
}
\arguments{
\item{x}{a data frame with columns chromosome}

\item{chr}{character string specifying UCSC chromosome to subset on}
}
\value{
object of class data frame
}
\description{
given a data frame subset out specific a chromosome
}
