% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutation_heatmap.R
\name{mutation_heatmap}
\alias{mutation_heatmap}
\title{Plot a mutation landscape}
\usage{
mutation_heatmap(data_frame, recurrence_cutoff = 0, grid = TRUE,
  label_x = FALSE, title = "NULL", gene_label_size = 8,
  coverage_space = 63564965, file_type = "TGI", genes = NULL)
}
\arguments{
\item{data_frame}{a data frame in annotation format}

\item{recurrence_cutoff}{an integer value to remove genes that do not have x number of mutations}

\item{grid}{a boolean value to overlay a grid on the primary plot}

\item{label_x}{a boolean value to plot samples on the x axis}

\item{title}{a character string for the plot title}

\item{gene_label_size}{an integer specifying the size of labels on Y axis}

\item{coverage_space}{an integer specifying the size in bp of the genome covered from which mutations could be called}

\item{file_type}{a character string specifying the file format of the data frame, one of "TGI", "MAF"}

\item{genes}{a character vector specifying genes to plot}
}
\value{
a grob for plotting
}
\description{
Plot a mutation landscape plot for a cohort in an annotation file
}

