% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutation_sample_subset.R
\name{mutation_sample_subset}
\alias{mutation_sample_subset}
\title{Mutation Sample Cutoff}
\usage{
mutation_sample_subset(x, genes)
}
\arguments{
\item{x}{a data frame in long format with columns 'gene', 'trv_type'}

\item{genes}{character vector listing genes to plot}
}
\value{
a subset data frame
}
\description{
Subset a MAF file keeping only entries in a selection of genes
}

