% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutation_silent_rmv.R
\name{mutation_silent_rmv}
\alias{mutation_silent_rmv}
\title{Silent Mutation Removal}
\usage{
mutation_silent_rmv(x)
}
\arguments{
\item{x}{a data frame with columns 'sample', 'gene', 'trv_type'}
}
\value{
a subset data frame
}
\description{
Subset a MAF file keeping only samples/genes that have >= 1 non-silent mutations
}

