% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationAnnotationFormat-class.R
\docType{class}
\name{MutationAnnotationFormat-class}
\alias{MutationAnnotationFormat-class}
\alias{MutationAnnotationFormat}
\title{Class MutationAnnotationFormat}
\usage{
MutationAnnotationFormat(path, version = "auto", verbose = FALSE)
}
\arguments{
\item{path}{String specifying the path to a MAF file.}

\item{version}{String specifying the version of the MAF file, if set to auto
the version will be obtained from the header in the MAF file.}

\item{verbose}{Boolean specifying if progress should be reported while reading
in the MAF file.}
}
\description{
An S4 class acting as a container for MutationAnnotationFormat version sub-classes.

Constructor for the MutationAnnotationFormat container class.
}
\section{Slots}{

\describe{
\item{\code{path}}{Character string specifying the path of the MAF file read in.}

\item{\code{version}}{Numeric value specifying the version of the MAF file.}

\item{\code{mafObject}}{MutationAnnotationFormat object which inherits from
MutationAnnotationFormat_Virtual class.}
}}

\seealso{
\code{\link{Waterfall}}

\code{\link{MutSpectra}}
}
