% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationAnnotationFormat_v2.4.R
\docType{class}
\name{MutationAnnotationFormat_v2.4-class}
\alias{MutationAnnotationFormat_v2.4}
\alias{MutationAnnotationFormat_v2.4-class}
\title{Class MutationAnnotationFormat_v2.4}
\usage{
\S4method{initialize}{MutationAnnotationFormat_v2.4}(.Object, mafPath,
  mafVersion, mafData)
}
\description{
An S4 class to represent data in mutation annotation format version 2.4

Initalizer method of MutationAnnotationFormat_v2.4 class
}
\section{Slots}{

\describe{
\item{\code{position}}{test}
}}

