% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutationObs.R
\name{mutationObs}
\alias{mutationObs}
\title{format mutation observations}
\usage{
mutationObs(data, fill_value, label_column, rep.fact, rep.dist.lmt, attr.fact,
  adj.max, adj.lmt, iter.max, ensembl.dataset = "hsapiens_gene_ensembl")
}
\arguments{
\item{data}{object of class data frame with columns trv_type and amino acid change}

\item{ensembl.dataset}{character string specifing the ensembl dataset to query when converting c. notation to p. notation}
}
\value{
object of class data frame giving mutation observations
}
\description{
Create a data frame of mutation observations
}

