% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutSpec.add_gene_to_NA.R
\name{mutSpec.add_gene_to_NA}
\alias{mutSpec.add_gene_to_NA}
\title{Assign NA samples a gene}
\usage{
mutSpec.add_gene_to_NA(data_frame)
}
\arguments{
\item{data_frame}{a data frame in MAF format}
}
\value{
a data frame with NA values in a gene column coerced to the top gene
name
}
\description{
Replace NA values in a gene column with the top gene name
}

