% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutSpec.convert_to_boolean.R
\name{mutSpec.convert_to_boolean}
\alias{mutSpec.convert_to_boolean}
\title{Convert counts to a boolean}
\usage{
mutSpec.convert_to_boolean(x)
}
\arguments{
\item{x}{a data frame in MAF format}
}
\value{
a data frame with values converted to boolean
}
\description{
Convert a data frame such that mutation type are converted to boolean instead
of counts
}

