% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutSpec.mutation_recurrence_subset.R
\name{mutSpec.mutation_recurrence_subset}
\alias{mutSpec.mutation_recurrence_subset}
\title{Mutation Recurrence Cutoff}
\usage{
mutSpec.mutation_recurrence_subset(data_frame, recurrence_cutoff)
}
\arguments{
\item{data_frame}{a data frame in long format with columns 'gene', 'trv_type'}

\item{recurrence_cutoff}{an integer, remove genes whose total mutations are
less than this amount}
}
\value{
a subset data frame
}
\description{
Subset a MAF file keeping only samples that meet a mutation recurrence cutoff
}

