% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutSpec.mutation_silent_rmv.R
\name{mutSpec.mutation_silent_rmv}
\alias{mutSpec.mutation_silent_rmv}
\title{Silent Mutation Removal}
\usage{
mutSpec.mutation_silent_rmv(x)
}
\arguments{
\item{x}{a data frame with columns 'sample', 'gene', 'trv_type'}
}
\value{
a subset data frame
}
\description{
Subset a MAF file setting keeping only sample information if a mutation
is silent
}

