% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutSpec.qual.R
\name{mutSpec.qual}
\alias{mutSpec.qual}
\title{Check input to mutSpec}
\usage{
mutSpec.qual(x, y, z, file_type, label_col)
}
\arguments{
\item{x}{a data frame in annotation format}

\item{y}{a data frame containing clinical data or a null object}

\item{z}{a data frame containing mutation burden information or a null object}

\item{file_type}{Character string specifying the input format to expect in x}

\item{label_col}{Character string specifying the column name of a label
column}
}
\value{
a list of data frames passing quality checks
}
\description{
Perform a data quality check on input to mutSpec
}

