% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutSpec.sample_sort.R
\name{mutSpec.sample_sort}
\alias{mutSpec.sample_sort}
\title{sort samples in a MAF file}
\usage{
mutSpec.sample_sort(x)
}
\arguments{
\item{x}{a data frame in long format with column names "sample",
"gene", "trv_type"}
}
\value{
a vector of samples in a sorted order
}
\description{
perform a hiearchial sort on samples based on the presence of mutations in
an ordered list of genes
}

