% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/VEP_Virtual-class.R
\docType{methods}
\name{parseDescription}
\alias{parseDescription}
\alias{parseDescription,VEP_Virtual-method}
\alias{parseDescription,VEP_Virtual}
\title{Method parseDescription}
\usage{
parseDescription(object, ...)

\S4method{parseDescription}{VEP_Virtual}(object, vepHeader, ...)
}
\arguments{
\item{object}{Object of class VEP_Virtual, needed for method dispatch on class.}

\item{vepHeader}{List of character vectors obtained from the vep header}
}
\description{
Method parseDescription
}
