% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/VEP_Virtual-class.R
\docType{methods}
\name{parseExtra}
\alias{parseExtra}
\alias{parseExtra,VEP_Virtual-method}
\alias{parseExtra,VEP_Virtual}
\title{Method parseExtra}
\usage{
parseExtra(object, ...)

\S4method{parseExtra}{VEP_Virtual}(object, vepData, ...)
}
\arguments{
\item{object}{Object of class VEP_Virtual, needed for method dispatch on class.}

\item{...}{additional arguments to passed}

\item{vepData}{Object of class data.table holding vep annotated data}
}
\description{
Method parseExtra
}
