% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/VEP_Virtual-class.R
\docType{methods}
\name{parseHeader}
\alias{parseHeader}
\alias{parseHeader,VEP_Virtual-method}
\alias{parseHeader,VEP_Virtual}
\title{Method parseHeader}
\usage{
parseHeader(object, ...)

\S4method{parseHeader}{VEP_Virtual}(object, vepHeader, ...)
}
\arguments{
\item{object}{Object of class VEP_Virtual, needed for method dispatch on class.}

\item{...}{additional arguments to passed}

\item{vepHeader}{List of character vectors obtained from the vep header}
}
\description{
Method parseHeader
}
