% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{plot mutation recurrence in genes}
\usage{
plot_bar(data_frame)
}
\arguments{
\item{data_frame}{a data frame in MAF format}
}
\value{
a ggplot object
}
\description{
plot a bar graph displaying the percentage of samples with a mutation
}

