% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_chromosome.R
\name{plot_chromosome}
\alias{plot_chromosome}
\title{plot chromosome}
\usage{
plot_chromosome(genome = "hg19", chromosome = "chr1", chr_txt_angle = 45,
  chr_txt_size = 5)
}
\arguments{
\item{genome}{character string specifying UCSC genome}

\item{chromosome}{character string specifying UCSC chromosome to plot}

\item{chr_txt_angle}{integer specifying angle of text when plotting band text}

\item{chr_txt_size}{integer specifying size of text when plotting band text}
}
\value{
ggplot object
}
\description{
plot chromosome in ggplot using UCSC database
}

