% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_coverage.R
\name{plot_coverage}
\alias{plot_coverage}
\title{produce a coverage plot}
\usage{
plot_coverage(coverage_data, txdb, gr, genome, reduce = F,
  gene_colour = NULL, gene_name = "test", bg_fill = "black",
  text_fill = "white", border = "black", size = 10, width_ratio = c(1,
  10), colour = "blue", plot_type = "line", cores = 1, base = exp(1),
  transform = c("Intron", "CDS", "UTR"))
}
\arguments{
\item{coverage_data}{named list containing data frames with columns stop and cov}

\item{txdb}{A TxDb object for a genome}

\item{gr}{A Granges object specifying a region of interest}

\item{genome}{Object of class BSgenome specifying the genome}

\item{reduce}{Boolean specifying whether to collapse isoforms in the ROI}

\item{gene_colour}{character string specifying the colour of the gene to be plotted}

\item{gene_name}{character string specifying the name of the gene or ROI}

\item{bg_fill}{character string giving the colour to fill the label}

\item{text_fill}{character string giving the colour to fill the text}

\item{border}{character string specifying the colour to fill the border of the label}

\item{size}{integer specifying the size of the text within the label}

\item{width_ratio}{vector of length 2 giving the ratio of track labels to plot}

\item{colour}{character string specifying the color of the data in the plot}

\item{plot_type}{character string specifying one of line, area for data display}

\item{cores}{Integer specifying the number of cores to use for processing}

\item{base}{The log base to transform the data}

\item{transform}{A vector of strings designating what objects to log transform}
}
\value{
ggplot object
}
\description{
produce a coverage plot displaying gene and coverage information
}

