% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_mutation_recurrence.R
\name{plot_mutation_recurrence}
\alias{plot_mutation_recurrence}
\title{plot mutation burden}
\usage{
plot_mutation_recurrence(data_frame, coverage_space)
}
\arguments{
\item{data_frame}{a data frame in MAF format}

\item{coverage_space}{an integer specifying the coverage space in base pairs from which a mutation could occur}
}
\value{
a ggplot object
}
\description{
plot a barchart showing mutations per MB
}

