% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenVisR.R
\docType{data}
\name{ptenCov}
\alias{ptenCov}
\title{Coverage for the pten gene}
\format{a data frame with 10734 observations and 4 variables}
\usage{
data(ptenCov)
}
\value{
Object of class data frame
}
\description{
A data set containing coverage data for pten obtained using "bedtools multicov" from "Convergent loss of PTEN leads to clinical resistance to a PI3K-alpha inhibitor" PMID:25409150
}
\keyword{datasets}

