% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reformat_cn.R
\name{reformat_cn}
\alias{reformat_cn}
\title{reformat CN data frame}
\usage{
reformat_cn(data_frame)
}
\arguments{
\item{data_frame}{a data frame with columns Chr, Coord, Tumor, Normal, Diff, p_value}
}
\value{
object of class data frame
}
\description{
given a CN data frame reformat it into a data frame build_CN_plot can understand
}

