% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rm.indel.R
\name{rm.indel}
\alias{rm.indel}
\title{Remove indels}
\usage{
rm.indel(x)
}
\arguments{
\item{x}{Object of class data frame containing columns 'reference', 'variant'}
}
\value{
Object of class data frame with indels removed
}
\description{
Given a data frame with columns reference and variants remove all indels
from data
}

