% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rm.mnuc.R
\name{rm.mnuc}
\alias{rm.mnuc}
\title{Remove multinucleotide codes}
\usage{
rm.mnuc(x)
}
\arguments{
\item{x}{Object of class data frame containing columns 'reference', 'variant'}
}
\value{
Object of class data frame with multi nucleotide codes removed
}
\description{
Given a data frame with columns reference and variants remove all multinucleotides from data
}

