% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample_sort.R
\name{sample_sort}
\alias{sample_sort}
\title{sort samples in a MAF file}
\usage{
sample_sort(data_frame)
}
\arguments{
\item{data_frame}{a data frame in MAF format}
}
\value{
a vector of samples in a sorted order
}
\description{
perform a hiearchial sort on samples based on the presence of mutations in an ordered list of genes
}

