% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenVisR.R
\docType{data}
\name{SNPloci}
\alias{SNPloci}
\title{Identity snps}
\format{a data frame with 24 observations and 3 variables}
\usage{
data(SNPloci)
}
\value{
Object of class data frame
}
\description{
A data set containing locations of 24 identity snps originating from:
Pengelly et al. Genome Med. 2013, PMID 24070238
}
\keyword{datasets}

