% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spread_coord_points.R
\name{spread_coord_points}
\alias{spread_coord_points}
\title{spread coordinates}
\usage{
spread_coord_points(x, spread_degree)
}
\arguments{
\item{x}{list}
}
\value{
numeric vector of spread points
}
\description{
given a data frame with median and group spreadd coordinates within that group
}

