% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuralVariant-class.R
\docType{class}
\name{StructuralVariant}
\alias{StructuralVariant}
\title{Class StructuralVariant}
\description{
An S4 class for the Structural Variant plot object
}
\section{Slots}{

\describe{
\item{\code{primaryData}}{data.table object storing primarydata used for plotting}

\item{\code{geneData}}{data.table object storing annotated gene files}

\item{\code{Grob}}{gtable object for the structural variant plot}
}}

