% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subset_chr.R
\name{subset_chr}
\alias{subset_chr}
\title{subset based on chr}
\usage{
subset_chr(x, chr)
}
\arguments{
\item{x}{a data frame with columns Chr, Coord, Tumor, Normal, Diff, p_value}

\item{chr}{character string specifying UCSC chromosome to plot one of chr... or all}
}
\value{
object of class data frame
}
\description{
given a data frame with cytogenetic band locations subset out specific chromosome
}

