% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructuralVariant-class.R
\name{svData}
\alias{svData}
\alias{StructuralVariant}
\title{Constuctor for the Structural Variant class}
\usage{
StructuralVariant(input, BSgenome = NULL, filter = TRUE, svType = NULL,
  svOrder = c("TRA", "BND", "DEL", "DUP", "INV", "INS"), maxSvSize = NULL,
  sample = NULL, chromosomes = NULL, ensembl = ensembl,
  attributes = attributes, filters = filters, annotate = TRUE,
  geneAnnotationFlank = 10000, plotSpecificGene = FALSE,
  plotGene1 = FALSE, plotGene2 = FALSE, chrGap = 5e+06, genome = "hg19",
  cytobandColor = c("White", "Grey"), sampleColor = NULL, verbose = FALSE,
  plotALayers = NULL, plotBLayers = NULL, plotCLayers = NULL,
  outputDir = "~/Desktop", plotWidth = 15, plotHeight = 12)
}
\arguments{
\item{BSgenome}{Object of class BSgenome to extract genome wide chromosome
coordinates}

\item{filter}{Boolean specifying if SV calls that did not pass should be removed}

\item{svType}{Character vector specifying which structural variant types to annotate/visualize}

\item{svOrder}{Character vector specifying the deleterious order of sv types (most to least deleterious)}

\item{maxSvSize}{Numeric specifying the maximum size of SV events (DEL/DUP/INV only)}

\item{sample}{Character vector specifying which samples to annotate/visualize}

\item{chromosomes}{Character vector specifying chromosomes to annotate/visualize}

\item{ensembl}{Object of class Mart to use in biomaRt query}

\item{attributes}{Character vector specifying which attributes to retrieve from biomaRt query}

\item{filters}{Character vector specifying which filters to use in biomaRt query}

\item{annotate}{Boolean specifying if the user wants to obtain mutated gene counts and annotate SV events}

\item{geneAnnotationFlank}{Integer specifying the size of the flanks of each SV event
to include in the annotation step}

\item{plotSpecificGene}{Character vector specifying which genes to plot}

\item{plotGene1}{Boolean specifying if TRA genes should be plotted}

\item{plotGene2}{Boolean specifying if non-TRA genes should be plotted}

\item{chrGap}{Integer specifying the size of the gap between the 1st and 2nd chromosome}

\item{genome}{Character vector specifying which genome to use to obtain chromosome bands. 
Serves as input into the getCytobands function of karyoploteR.}

\item{cytobandColor}{Character vector specifying what to color the chromosome bands}

\item{sampleColor}{Character vector specifying colors to plot for each sample}

\item{verbose}{Boolean specifying if status messages should be reported}

\item{plotBLayers}{List of ggplot2 layers to be passed to chromosome plot}

\item{plotCLayers}{List of ggplot2 layers to be passed to non-translocation plot}

\item{outputDir}{Character value for directory to output SV visualizations}

\item{plotWidth}{Integer for width of SV visualizations}

\item{plotHeight}{Integer for height of SV visualizations}

\item{object}{Object of class VCF}

\item{object}{OBject of class VCF}

\item{plotALAyers}{List of ggplot2 layers to be passed to translocation plot}
}
\description{
Constuctor for the Structural Variant class
}
