% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TGI_to_anno.R
\name{TGI_to_anno}
\alias{TGI_to_anno}
\title{Convert TGI File}
\usage{
TGI_to_anno(x)
}
\arguments{
\item{x}{a data frame in TGI internal format}
}
\value{
a data frame coerced from TGI to internal annotation format
}
\description{
Convert columns of a mutation annotation file "TGI" into a format recognizable by internal functions
}

