% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trackViz.R
\name{trackViz}
\alias{trackViz}
\title{plot with named tracka}
\usage{
trackViz(..., bg_fill = "black", text_fill = "white", border = "black",
  size = 10, axis_align = "none", width_ratio = c(1, 10),
  nested_list = F)
}
\arguments{
\item{...}{named list of plots}

\item{bg_fill}{character string giving the colour to fill the label}

\item{text_fill}{character string giving the colour to fill the text}

\item{border}{character string specifying the colour to fill the border of the label}

\item{size}{integer specifying the size of the text within the label}

\item{axis_align}{character string specifying axis to align plotting space on, one of 'both', 'height', 'width', 'none'}

\item{width_ratio}{vector of length 2 giving the ratio of track labels to plot}

\item{nested_list}{boolean specifying whether plots are in a named list nested in another list}
}
\value{
ggplotGrob object
}
\description{
given a named list of plots, plot them on tracks
}

