% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transcriptID2length.R
\name{transcriptID2length}
\alias{transcriptID2length}
\title{Convert trascript ID to protien length}
\usage{
transcriptID2length(transcriptID)
}
\arguments{
\item{transcriptID}{String specifying ensembl transcript ID}
}
\value{
numeric value specifying protien length
}
\description{
Retrieve the length of an ensembl transcript ID in amino acids
}

