% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transcriptID2sequence.R
\name{transcriptID2sequence}
\alias{transcriptID2sequence}
\title{Convert trascript ID to protien length}
\usage{
transcriptID2sequence(transcriptID, up)
}
\arguments{
\item{transcriptID}{String specifying ensembl transcript ID}

\item{up}{uniprot data object from UniProt.ws}
}
\value{
character vector giving protien sequence
}
\description{
Retrieve the length of an ensembl transcript ID in amino acids
}

