% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tanscriptID2uniprotID.R
\name{transcriptID2uniprotID}
\alias{transcriptID2uniprotID}
\title{Convert trascript ID to uniprot ID}
\usage{
transcriptID2uniprotID(transcriptID, up)
}
\arguments{
\item{transcriptID}{String specifying ensembl transcript ID}

\item{taxId}{integer specifying the uniprot taxonomy id for the species of interest}
}
\value{
String specifying uniprot ID
}
\description{
Convert an ensembl transcript ID to the corresponding uniprot ID
}

