% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transition_transversion_plot.R
\name{transition_transversion_plot}
\alias{transition_transversion_plot}
\title{plot transitions/transversions}
\usage{
transition_transversion_plot(x, type = "Proportion", x_axis_text_angle = 45)
}
\arguments{
\item{x}{Object of class data frame containing columns 'reference', 'variant'}

\item{type}{Object of class character specifying whether to plot the Proportion or Frequency, one of "Prop"}

\item{x_axis_text_angle}{Integer specifying the angle to labels on x_axis}
}
\value{
Object of class data frame with indels removed
}
\description{
Given a data frame with columns reference, variant, and sample construct a transition/transversion plot
}

