% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TvTi.convMaf.R
\name{TvTi.convMAF}
\alias{TvTi.convMAF}
\alias{TvTi.convMaf}
\title{Convert .maf format to internal format}
\usage{
TvTi.convMaf(x)
}
\arguments{
\item{x}{Object of class data frame containing columns 'Tumor_Sample_Barcode', 'Reference_Allele' 'Tumor_Seq_Allele1', 'Tumor_Seq_Allele2'}
}
\value{
a data frame, with column names 'sample', 'reference', 'variant'
}
\description{
Convert data frame in .maf format to an internally recogized format
}

