% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TvTi.qual.R
\name{TvTi.qual}
\alias{TvTi.qual}
\title{Check input to TvTi}
\usage{
TvTi.qual(x, y = NULL, file_type = "MAF")
}
\arguments{
\item{x}{Object of class data frame containing columns 'sample', reference',
'variant' for 'MGI' file or 'Tumor_Sample_Barcode', 'Reference_Allele',
'Tumor_Seq_Allele1', 'Tumor_Seq_Allele2' for 'MAF' file}

\item{y}{Object of class data frame containing columns "Prop", "trans_tranv"}

\item{file_type}{Character string spedifying th input file type expected}
}
\value{
a data frame, or list of data frames passing quality checks
}
\description{
Perform quality check for input to function TvTi
}

