% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TvTi.R
\name{TvTi}
\alias{TvTi}
\title{plot transitions/transversions}
\usage{
TvTi(x, y = NULL, type = "Proportion", label_x_axis = TRUE,
  x_axis_text_angle = 45, palette = c("#7BC374", "#EFCD8D", "#8763A0",
  "#6677A0", "#EDEE8D", "#EF8D8D"), file_type = "MAF")
}
\arguments{
\item{x}{Object of class data frame containing columns 'sample', reference', 'variant'}

\item{y}{named vector containing expected transition/transversion rates}

\item{type}{Object of class character specifying whether to plot the Proportion or Frequency, one of "Prop"}

\item{label_x_axis}{boolean specifying wheter to label x axis}

\item{x_axis_text_angle}{Integer specifying the angle to labels on x_axis}

\item{palette}{Character vector of length 6 specifying colors for trans/tranv type}

\item{file_type}{Character string specifying the format the input is in, one of 'MAF', 'MGI'}
}
\value{
Object of class data frame with indels removed
}
\description{
Given a data frame with columns reference, variant, and sample construct a transition/transversion plot
}

