% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TvTi_alignPlot.R
\name{TvTi_alignPlot}
\alias{TvTi_alignPlot}
\title{align TvTi plots on y axis}
\usage{
TvTi_alignPlot(p1, p2)
}
\arguments{
\item{p1}{main plot}

\item{p2}{left subplot}
}
\value{
ggplot object
}
\description{
align transition/transversion plots
}

