% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TvTi_annoTransTranv.R
\name{TvTi_annoTransTranv}
\alias{TvTi_annoTransTranv}
\title{Annotate Transitions and Transversions}
\usage{
TvTi_annoTransTranv(x)
}
\arguments{
\item{x}{Object of class data frame containing columns 'reference', 'variant'}
}
\value{
Object of class data frame with transition/transversion annotations
appended
}
\description{
Given a data frame with columns reference and variant annotate the base
change occurring
}

