% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TvTi_buildMain.R
\name{TvTi_buildMain}
\alias{TvTi_buildMain}
\title{build transitions/transversions}
\usage{
TvTi_buildMain(x, y = NULL, type = "Proportion", label_x_axis = TRUE,
  x_axis_text_angle = 45, palette = c("#D53E4F", "#FC8D59", "#FEE08B",
  "#E6F598", "#99D594", "#3288BD"), plot_expected = FALSE,
  tvti.layers = NULL, expec.layers = NULL, title_x_axis = TRUE)
}
\arguments{
\item{x}{Object of class data frame containing columns 'trans_tranv',
'sample', 'Freq', and 'Prop'}

\item{y}{Object of class data frame containing columns 'Prop', 'trans_tranv'
for display of expected results}

\item{type}{Object of class character specifying whether to plot the
Proportion or Frequency, one of "Prop"}

\item{label_x_axis}{boolean specifying wheter to label x axis}

\item{x_axis_text_angle}{Integer specifying the angle to labels on x_axis}

\item{palette}{Character vector of length 6 specifying colors for
trans/tranv type}

\item{plot_expected}{Boolean specifying if this is the main TvTi plot or a
sub plot for expected values}

\item{tvti.layers}{Additional ggplot2 layers for the main plot}

\item{expec.layers}{Additional ggplot2 layers for the expected values plot}

\item{title_x_axis}{boolean specifying whether to display an x axis title}
}
\value{
GGplot Object
}
\description{
Given a data frame with columns 'trans_tranv', 'sample', 'Freq', and 'Prop',
build a transition/transversion plot
}
