% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TvTi_calcTransTranvFreq.R
\name{TvTi_calcTransTranvFreq}
\alias{TvTi_calcTransTranvFreq}
\title{Calculate Transition/Transversion Frequency}
\usage{
TvTi_calcTransTranvFreq(x)
}
\arguments{
\item{x}{Object of class data frame containing columns 'reference',
'variant', 'sample', 'trans_tranv'}
}
\value{
Object of class data frame with Frequency and Proportion of
Transistions/Transversions appended on a sample level
}
\description{
Given a data frame with columns reference, variant, sample, and trans/tranv
calculate the frequencies of transitions and transversion occuring.
}
