% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TvTi_rmIndel.R
\name{TvTi_rmIndel}
\alias{TvTi_rmIndel}
\title{Remove indels}
\usage{
TvTi_rmIndel(x)
}
\arguments{
\item{x}{Object of class data frame containing columns 'reference', 'variant'}
}
\value{
Object of class data frame with indels removed
}
\description{
Given a data frame with columns reference and variants remove all indels
from data
}
