% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TvTi_rmMnuc.R
\name{TvTi_rmMnuc}
\alias{TvTi_rmMnuc}
\title{Remove multinucleotide codes}
\usage{
TvTi_rmMnuc(x)
}
\arguments{
\item{x}{Object of class data frame containing columns 'reference', 'variant'}
}
\value{
Object of class data frame with multi nucleotide codes removed
}
\description{
Given a data frame with columns reference and variants remove all
multinucleotides from data
}

