% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantCallFormat-class.R
\docType{class}
\name{VariantCallFormat-class}
\alias{VariantCallFormat-class}
\alias{VariantCallFormat}
\alias{extractVariantCallFormat}
\title{Class VariantCallFormat}
\usage{
extractVariantCallFormat(path = NULL, data = NULL, version = "auto",
  svCaller = NULL, paired = paired, tumorColumn = tumorColumn,
  verbose = FALSE)
}
\arguments{
\item{path}{String specifying the path to a VCF file. Can accept wildcards 
if multiple VCF files exist (see details).}

\item{data}{data.table object storing a VCF file. Overrides "path" if 
specified}

\item{version}{String specifying the version of the VCF file, if set to auto
the version will be obtained from the header in the VCF file}

\item{svCaller}{String specifying the structural variant caller used}

\item{paired}{Boolean specifiying if the svCaller was run with the paired option 
(i.e. tumor-normal)}

\item{tumorColumn}{Integer specifying the column number with the tumor read support information. 
Only used when paired=TRUE.}

\item{verbose}{Bolean specifying if progress should be reported while reading
in the VCF file}
}
\description{
Class VariantCallFormat

Constructor for the VCF container class
}
\details{
When specifying a path to a VCF file, the option exists to either
specify the full path to a vcf file or to us wildcards to specify multiple 
files. When specifying a full path, the initializer will check if a column
named "sample" containing the relevant sample for each row exists. If such a 
column is not found, the initializer will assume this file correspnds to 
only one sample and populate a sample column accordingly. Alternatively, if 
multiple files are specified at once using a wildcard, the initializer will 
aggregate all the files and use the filenames minus any extension to 
populate the "sample" column.
}
\section{Slots}{

\describe{
\item{\code{path}}{Character string specifying the path of the VCF file read in}

\item{\code{version}}{Character string specifiying the version of the vcf file}

\item{\code{vcfObject}}{vcf object which inherits from VCF_Virtual class}
}}

