% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarScanFormat-class.R
\docType{class}
\name{VarScanFormat-class}
\alias{VarScanFormat-class}
\alias{VarScanFormat}
\title{Class VarScanFormat}
\usage{
VarScanFormat(path = NULL, varscanData = NULL, varscanType = "LOH",
  verbose = FALSE)
}
\arguments{
\item{path}{String specifying the path to a VarScan file.}

\item{verbose}{Boolean specifying if progress should be reported while reading
in the VarScan. file.}
}
\description{
Class VarScanFormat

Constructor for the VarScanFormat container class.
}
\section{Slots}{

\describe{
\item{\code{path}}{Character string specifying the path of the VarScan file read in.}
}}

\seealso{
\code{\link{lohSpec}}
}
