% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarScanFormat_Virtual-class.R
\docType{class}
\name{VarScanFormat_Virtual-class}
\alias{VarScanFormat_Virtual-class}
\title{Class VarScanFormat_Virtual}
\description{
An S4 class to act as a virtual class for VarScanFormat version sub-classes.
}
\section{Slots}{

\describe{
\item{\code{varscan}}{data.table object holding varscan data.}

\item{\code{sample}}{data.table object holding sample data.}
}}

