% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Manta_v4.1-class.R
\docType{class}
\name{VCF_Manta_v4.1}
\alias{VCF_Manta_v4.1}
\alias{VCF_Manta_v4.1}
\title{Class VCF_Manta_v4.1}
\usage{
VCF_Manta_v4.1(vcfData, vcfHeader, paired, tumorColumn)
}
\arguments{
\item{vcfData}{data.table object containing a VCF file conforming to the 
version 4.1 specifications}

\item{vcfHeader}{Object of class list containing character vectors for vcf
header information}

\item{paired}{Boolean object specifying if the svCaller was ran in paired mode}

\item{tumorColumn}{String specifying the name of the sample column with read support information}
}
\description{
An S4 class to represent data in vcf version 4.1 format, inherits from the 
VCF_Virtual class

Constuctor for the VCF_Manta_v4.1 sub-class
}
\section{Slots}{

\describe{
\item{\code{header}}{data.table object containing header information}

\item{\code{meta}}{data.table object containing meta information lines}

\item{\code{vcfHeader}}{data.table object containing header for vcf data}

\item{\code{vcfData}}{data.table object containing vcf data lines}

\item{\code{sample}}{data.table object containing sample information}
}}

