% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF_Virtual-class.R
\docType{class}
\name{VCF_Virtual}
\alias{VCF_Virtual}
\title{Class VCF_Virtual}
\description{
An S4 class to act as a virtual class for VCF version sub-classes
}
\section{Slots}{

\describe{
\item{\code{vcf}}{data.table object holding varscan data}

\item{\code{sample}}{data.table object holding sample data}
}}

