% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEP_v88-class.R
\docType{class}
\name{VEP_v88-class}
\alias{VEP_v88-class}
\alias{VEP_v88}
\title{Class VEP_v88}
\usage{
VEP_v88(vepData, vepHeader)
}
\arguments{
\item{vepData}{data.table object containing a VEP annotation file conforming
to the version 88 specifications.}

\item{vepHeader}{Object of class list containing character vectors for vep
header information.}
}
\description{
An S4 class to represent data in variant effect predictor version 88 format,
inherits from the VEP_Virtual class.
}
\section{Slots}{

\describe{
\item{\code{header}}{data.table object containing header information}

\item{\code{description}}{data.table object containing column descriptions}

\item{\code{position}}{data.table object containing column names "chromosome_name",
"start", "stop".}

\item{\code{mutation}}{data.table object containing column names "reference",
"variant", "trv_type".}

\item{\code{sample}}{data.table object containing columns names "sample".}

\item{\code{meta}}{data.table object containing meta data.}
}}

