% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VEP_Virtual-class.R
\docType{class}
\name{VEP_Virtual-class}
\alias{VEP_Virtual-class}
\title{Class VEP_Virtual}
\description{
An S4 class to act as a virtual class for VEP version sub-classes, under development!!!
}
\section{Slots}{

\describe{
\item{\code{header}}{data.table object holding header information.}

\item{\code{description}}{data.table object holding column descriptions}

\item{\code{position}}{data.table object holding genomic positions.}

\item{\code{mutation}}{data.table object holding mutation status data.}

\item{\code{sample}}{data.table object holding sample data.}

\item{\code{meta}}{data.table object holding all other meta data.}
}}

