% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Waterfall-class.R
\docType{methods}
\name{Waterfall-class}
\alias{Waterfall}
\alias{Waterfall-class}
\title{Class Waterfall}
\usage{
\S4method{initialize}{Waterfall}(.Object, input, labelColumn, samples, coverage,
  noSynonymous, genes, mutationHierarchy, verbose)

Waterfall(input, labelColumn = NULL, samples = NULL, coverage = NULL,
  noSynonymous = FALSE, genes = NULL, mutationHierarchy = NULL,
  verbose = FALSE)
}
\arguments{
\item{input}{MutationAnnotationFormat class holding genomic information.}

\item{samples}{Character vector specifying samples to plot. If not NULL
all samples in "input" not specified with this parameter are removed.}

\item{coverage}{Integer specifying the size in base pairs of the genome
covered by sequence data from which mutations could be called. Required for
the mutation burden sub-plot (see details and vignette).}

\item{noSynonymous}{Boolean specifying if silent mutations should be removed
from the plot, this will not affect the mutation burden calculation.}

\item{mutationHierarchy}{Data.table object with rows specifying the order of
mutations from most to least deleterious and column names "mutation" and
"color". Used to change the default colors and/or to give priority to a
mutation for the same gene/sample (see details and vignette).}

\item{verbose}{Boolean specifying if status messages should be reported}
}
\description{
An S4 class for the waterfall plot object

Initalizer method for the Waterfall class

Constructor for the Waterfall class.
}
\section{Slots}{

\describe{
\item{\code{plotA}}{gtable object for the top sub-plot.}

\item{\code{plotB}}{gtable object for the left sub-plot.}

\item{\code{plotC}}{gtable object for the main plot.}

\item{\code{plotD}}{gtable object for the bottom sub-plot.}

\item{\code{Grob}}{gtable object for the arranged plot.}

\item{\code{primaryData}}{data.table object storing the primary data, should have
column names sample, gene, mutation, label.}

\item{\code{simpleMutationCounts}}{data.table object storing simplified mutation
counts, should have column names sample, mutation, Freq, mutationBurden}

\item{\code{complexMutationData}}{data.table object storing mutation counts per
mutation type should have column names sample, mutation, Freq, mutationBurden.}

\item{\code{MutationHierarchy}}{data.table object storing the hierarchy of mutation
type in order of most to least important and the mapping of mutation type to
color. Should have column names mutation and color.}
}}

