% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall.R
\name{waterfall}
\alias{waterfall}
\title{Construct a waterfall plot}
\usage{
waterfall(x, clinData = NULL, clinLegCol = 1, clinVarCol = NULL,
  clinVarOrder = NULL, clinLayer = NULL, mutBurden = NULL,
  main.recurrence_cutoff = 0, main.grid = TRUE, main.label_x = FALSE,
  main.gene_label_size = 8, coverageSpace = 44100000, file_type = "MAF",
  main.genes = NULL, main.samples = NULL, drop_mutation = FALSE,
  rmv_silent = FALSE, main.label_col = NULL, main.plot_label_size = 4,
  main.palette = NULL, sampRecur.layers = NULL, main.layers = NULL,
  mutRecur.layers = NULL, main.plot_label_angle = 0,
  variant_class_order = NULL)
}
\arguments{
\item{x}{Object of class data frame representing annotated mutations. The 
data frame supplied must have one of the following sets of column names
("Tumor_Sample_Barcode", "Hugo_Symbol", "Variant Classification") for
fileType="MAF", ("sample","gene_name","trv_type") for fileType="MGI" or
("sample", "gene", "variant_class") for fileType="Custom". This columns 
should represent samples in a cohort, gene with mutation, and the mutation
type respectively.}

\item{clinData}{Object of class data frame with rows representing clinical
data. The data frame should be in "long format" and columns must be names as
"sample", "variable", and "value" (optional see details and vignette).}

\item{clinLegCol}{Integer specifying the number of columns in the legend for
the clinical data, only valid if argument is supplied to parameter clinData.}

\item{clinVarCol}{Named character vector specifying the mapping of colours
to variables in the variable column of the data frame supplied to clinData
(ex. "variable"="colour").}

\item{clinVarOrder}{Character vector specifying the order in which to plot
variables in the variable column of the argument given to the parameter
clinData. The argument supplied to this parameter should have the same unique
length and values as in the variable column of the argument supplied to 
parameter clinData (see vignette).}

\item{clinLayer}{Valid ggplot2 layer to be added to the clinical sub-plot.}

\item{mutBurden}{an optional data frame containing columns sample, mut_burden
with sample levels matching those supplied in x}

\item{main.recurrence_cutoff}{integer specifying removal of entries not seen
in at least "x" percent of samples}

\item{main.grid}{a boolean value to overlay a grid on the primary plot}

\item{main.label_x}{a boolean value to plot samples on the x axis}

\item{main.gene_label_size}{an integer specifying the size of labels on Y
axis}

\item{coverageSpace}{an integer specifying the size in bp of the genome
covered from which mutations could be called}

\item{file_type}{a character string specifying the file format of the data
frame, one of "MGI", "MAF", "Custom"}

\item{main.genes}{a character vector specifying genes to plot}

\item{main.samples}{character vector containing sample names to subset the
input from x on}

\item{drop_mutation}{Boolean specifying whether to drop unused
"mutation type" levels from the legend}

\item{rmv_silent}{Boolean specifying wheter to remove silent mutations from
the left side and main plot}

\item{main.label_col}{Character string specifying an optional column name
from which to derive cell labels from}

\item{main.plot_label_size}{Integer specifying the size of labels for cells
in the main panel}

\item{main.palette}{Character vector specifying colors to fill on mutation
type}

\item{sampRecur.layers}{Additional ggplot2 layers to plot on the sample
recurence chart}

\item{main.layers}{Additional ggplot2 layers to plot on the main panel}

\item{mutRecur.layers}{Additional ggplot2 layers to plot on the mutation
burden data}

\item{main.plot_label_angle}{Integer specifying the degree of rotation for
text if main.label_col is specified}

\item{variant_class_order}{character vector giving the hierarchical order of
mutation types to plot (required if file_type="Custom")}
}
\value{
a grob for plotting
}
\description{
Given a data frame construct a water fall plot showing the mutation burden
and mutation type on a gene and sample level.
}
\examples{
waterfall(brcaMAF)
}

