% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_align.R
\name{waterfall_align}
\alias{waterfall_align}
\title{align plots}
\usage{
waterfall_align(genes, heatmap, burden, clinical, proportions, section_heights)
}
\arguments{
\item{genes}{ggplot object displaying mutation burden on gene}

\item{heatmap}{ggplot object displaying a mutation landscape}

\item{burden}{ggplot object displaying mutation burden on sample}

\item{clinical}{ggplot object displaying clinical information "optional"}

\item{proportions}{ggplot object displaying proportion of mutation types "optional"}

\item{section_heights}{Heights of each section (should sum to one)}
}
\value{
a grob object
}
\description{
align mutation landscape, mutation burden on sample, and mutation burden on
gene plots
}
