% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_build_proportions.R
\name{waterfall_build_proportions}
\alias{waterfall_build_proportions}
\title{Build mutational profile plot}
\usage{
waterfall_build_proportions(data_frame, plot_palette, file_type, layers,
  x_label)
}
\arguments{
\item{data_frame}{input data.frame}

\item{plot_palette}{Color palette to use}

\item{file_type}{MAF, etc}

\item{layers}{layer(s) to add to this plot object}

\item{x_label}{label this plot?}
}
\value{
a ggplot object
}
\description{
Builds a ggplot object showing individuals' mutational profile
}

