% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{waterfall_buildGenePrevelance}
\alias{waterfall_buildGenePrevelance}
\title{plot mutation recurrence in genes}
\usage{
waterfall_buildGenePrevelance(data_frame, gene_label_size = 8,
  layers = NULL)
}
\arguments{
\item{data_frame}{a data frame in MAF format}

\item{gene_label_size}{numeric value indicating the size of the gene labels
on the y-axis}

\item{layers}{additional ggplot2 layers}
}
\value{
a ggplot object
}
\description{
plot a bar graph displaying the percentage of samples with a mutation
}
