% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_buildMain.R
\name{waterfall_buildMain}
\alias{waterfall_buildMain}
\title{Plot a mutation heatmap}
\usage{
waterfall_buildMain(data_frame, grid = TRUE, label_x = FALSE,
  file_type = "MGI", drop_mutation = FALSE, plot_x_title = TRUE,
  plot_label = FALSE, plot_label_size = 4, plot_palette = NULL,
  layers = NULL, plot_label_angle = 0)
}
\arguments{
\item{data_frame}{a data frame in MAF format}

\item{grid}{boolean value whether to overlay a grid on the plot}

\item{label_x}{boolean value whether to label the x axis}

\item{file_type}{character string specifying the file type, one of 'MAF' or
'MGI'}

\item{drop_mutation}{Boolean specifying whether to drop unused
"mutation type" levels from the legend}

\item{plot_x_title}{Boolean specifying whether to plot the x_axis title}

\item{plot_label}{Boolean specifying whether to plot text inside each cell}

\item{plot_label_size}{Integer specifying text size of cell labels}

\item{plot_palette}{Character vector specifying colors to fill on mutation
type}

\item{layers}{additional ggplot2 layers to plot}

\item{plot_label_angle}{angle at which to plot label text if plot_label is
true}
}
\value{
a ggplot2 object
}
\description{
Plot a Mutation Landscape with variables sample, gene, mutation
}

