% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{waterfall_buildMutBurden_A}
\alias{waterfall_buildMutBurden_A}
\title{plot mutation burden}
\usage{
waterfall_buildMutBurden_A(x, coverage_space, layers = NULL)
}
\arguments{
\item{x}{a data frame in MAF format}

\item{coverage_space}{an integer specifying the coverage space in base pairs
from which a mutation could occur}

\item{layers}{Additional ggplot2 layers to plot}
}
\value{
a ggplot object
}
\description{
plot a barchart showing mutations per MB
}
