% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_buildMutBurden_B.R
\name{waterfall_buildMutBurden_B}
\alias{waterfall_buildMutBurden_B}
\title{plot mutation burden}
\usage{
waterfall_buildMutBurden_B(x, layers = NULL)
}
\arguments{
\item{x}{a data frame containing columns sample, mut_burden}

\item{layers}{additional ggplot2 layers to plot}
}
\value{
a ggplot object
}
\description{
plot a barchart showing mutation burden given by data frame
}
