% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{waterfall_calcMutFreq}
\alias{waterfall_calcMutFreq}
\title{Calculate Synonymous/Nonsynonymous mutation frequency}
\usage{
waterfall_calcMutFreq(x)
}
\arguments{
\item{x}{data frame in long format with columns sample, trv_type}
}
\value{
a data frame with synonymous/nonsynonymous counts appended
}
\description{
Creates a data frame giving synonymous/nonsynonymous counts on a sample level
}
