% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_Custom2anno.R
\name{waterfall_Custom2anno}
\alias{waterfall_Custom2anno}
\title{Convert Custom File}
\usage{
waterfall_Custom2anno(x, label_col, variant_class_order)
}
\arguments{
\item{x}{Object of class data frame with rows representing mutations and
containing columns "key, "sample", "gene", "variant_class".}

\item{label_col}{Character string specifying the column name of a
label column (optional)}
}
\value{
a data frame coerced from custom to annotation format
}
\description{
Convert columns of a Custom annotation file into a format
recognizable by internal functions
}
