% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_geneAlt.R
\name{waterfall_geneAlt}
\alias{waterfall_geneAlt}
\title{mutation sample cutoff gene based}
\usage{
waterfall_geneAlt(x, genes)
}
\arguments{
\item{x}{a data frame in long format with columns 'gene', 'trv_type'}

\item{genes}{character vector listing genes to plot}
}
\value{
a subset data frame
}
\description{
Subset a internal mutSpec file keeping only samples within the specified gene
 list
}
