% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_geneSort.R
\name{waterfall_geneSort}
\alias{waterfall_geneSort}
\title{sort waterfall file by gene}
\usage{
waterfall_geneSort(x, geneOrder = NULL)
}
\arguments{
\item{x}{Data frame with columns names "gene", "trv_type".}

\item{geneOrder}{Character vector specifying the order in which to plot
genes.}
}
\value{
Character vector of ordered genes
}
\description{
order a waterfall file ranking genes with more mutations higher if a gene
order is unspecified.
}
