% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/waterfall_hierarchyTRV.R
\name{waterfall_hierarchyTRV}
\alias{waterfall_hierarchyTRV}
\title{Hiearchical removal of MAF entries}
\usage{
waterfall_hierarchyTRV(x, file_type, variant_class_order)
}
\arguments{
\item{x}{a data frame in long format with columns sample, gene,
trv_type}

\item{file_type}{The type of file to act on one of 'MAF", "MGI", "Custom"}

\item{variant_class_order}{character vector giving the hierarchical order of
mutation types to plot}
}
\value{
a data frame with multiple mutations in the same sample/gene
collapsed on the most deleterious
}
\description{
Remove MAF entries with the same gene/sample in an ordered fashion such that
the most deleterious are retained
}

