% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{waterfall_MAF2anno}
\alias{waterfall_MAF2anno}
\title{Convert MAF File}
\usage{
waterfall_MAF2anno(x, label_col)
}
\arguments{
\item{x}{a data frame in MAF format}

\item{label_col}{Character string specifying the column name of a
label column}
}
\value{
a data frame coerced from MAF to TGI format
}
\description{
Convert columns of a mutation annotation file "MAF" into a format
recognizable by internal functions
}
