% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_MGI2anno.R
\name{waterfall_MGI2anno}
\alias{waterfall_MGI2anno}
\title{Convert MGI File}
\usage{
waterfall_MGI2anno(x, label_col)
}
\arguments{
\item{x}{a data frame in MGI internal format}

\item{label_col}{Character string specifying the column name of a label
column}
}
\value{
a data frame coerced from MGI to internal annotation format
}
\description{
Convert columns of a mutation annotation file "MGI" into a format
recognizable by internal functions
}
