% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{waterfall_NA2gene}
\alias{waterfall_NA2gene}
\title{Assign NA samples a gene}
\usage{
waterfall_NA2gene(x)
}
\arguments{
\item{x}{a data frame in anno format}
}
\value{
a data frame with NA values in a gene column coerced to the top gene
name
}
\description{
Replace NA values in a gene column with the top gene name
}
