% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_palette_names.R
\name{waterfall_palette_names}
\alias{waterfall_palette_names}
\title{waterfall_palette_names}
\usage{
waterfall_palette_names(palette, file_type, data_frame)
}
\arguments{
\item{palette}{Named colour vector as input}

\item{file_type}{Which file type is involved?}

\item{data_frame}{Only used if file_type is "custom"}
}
\value{
a named list of "breaks" and "labels"
}
\description{
Make labels and breaks for palettes
}
\details{
waterfall_palette_names
}

