% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_qual.R
\name{waterfall_qual}
\alias{waterfall_qual}
\title{Check input to mutSpec}
\usage{
waterfall_qual(x, y, z, file_type, label_col, variant_class_order)
}
\arguments{
\item{x}{a data frame in annotation format}

\item{y}{a data frame containing clinical data or a null object}

\item{z}{a data frame containing mutation burden information or a null object}

\item{file_type}{Character string specifying the input format to expect in x}

\item{label_col}{Character string specifying the column name of a label}

\item{variant_class_order}{Character vector specifying the hierarchical order
of mutation types to plot}
}
\value{
a list of data frames passing quality checks
}
\description{
Perform a data quality check on input to mutSpec
}

