% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_rmvSilent.R
\name{waterfall_rmvSilent}
\alias{waterfall_rmvSilent}
\title{Silent Mutation Removal}
\usage{
waterfall_rmvSilent(x)
}
\arguments{
\item{x}{a data frame with columns 'sample', 'gene', 'trv_type'}
}
\value{
a subset data frame
}
\description{
Subset a MAF file setting keeping only sample information if a mutation
is silent
}
