% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_sampAlt.R
\name{waterfall_sampAlt}
\alias{waterfall_sampAlt}
\title{mutation sample subset sample based}
\usage{
waterfall_sampAlt(x, samples)
}
\arguments{
\item{x}{a data frame in long format with columns 'sample', 'trv_type'}

\item{samples}{character vector giving samples to plot}
}
\value{
a subset data frame
}
\description{
Alter a mutSpec input file keeping/adding entries in a selection of samples
}

