% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_select_palette.R
\name{waterfall_select_palette}
\alias{waterfall_select_palette}
\title{Helper function to select a colour palette}
\usage{
waterfall_select_palette(file_type, custom_palette = NULL)
}
\arguments{
\item{file_type}{Which file tyoe is used?}

\item{custom_palette}{Nullable custom colour palette}
}
\value{
A vector of colours to be used as palette
}
\description{
waterfall_select_palette
}

