% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_VEP2anno.R
\name{waterfall_VEP2anno}
\alias{waterfall_VEP2anno}
\title{Convert VEP File}
\usage{
waterfall_VEP2anno(x, label_col)
}
\arguments{
\item{x}{a data frame in VEP format}

\item{label_col}{Character string specifying the column name of a
label column}
}
\value{
a data frame coerced from VEP to intermal expected format
}
\description{
Convert columns of a variant effect predictor file "VEP" into a format
recognizable by internal functions
}

