% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfallAlign.R
\name{waterfallAlign}
\alias{waterfallAlign}
\title{align plots}
\usage{
waterfallAlign(p2, p1, p3, p4)
}
\arguments{
\item{p2}{ggplot object displaying mutation burden on gene}

\item{p1}{ggplot object displaying a mutation landscape}

\item{p3}{ggplot object displaying mutation burden on sample}

\item{p4}{ggplot object displaying clinical information "optional"}
}
\value{
a grob object
}
\description{
align mutation landscape, mutation burden on sample, and mutation burden on
gene plots
}

