% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GMS_Virtual-class.R,
%   R/GMS-class.R, R/MutationAnnotationFormat_Virtual-class.R,
%   R/MutationAnnotationFormat-class.R, R/VEP_Virtual-class.R, R/VEP-class.R
\docType{methods}
\name{writeData}
\alias{writeData}
\alias{writeData,GMS_Virtual-method}
\alias{writeData,GMS_Virtual}
\alias{writeData,GMS-method}
\alias{writeData,MutationAnnotationFormat_Virtual-method}
\alias{writeData,MutationAnnotationFormat_Virtual}
\alias{writeData,MutationAnnotationFormat-method}
\alias{writeData,VEP_Virtual-method}
\alias{writeData,VEP_Virtual}
\alias{writeData,VEP-method}
\title{Method writeData}
\usage{
writeData(object, ...)

\S4method{writeData}{GMS_Virtual}(object, file, sep, ...)

\S4method{writeData}{GMS}(object, file, ...)

\S4method{writeData}{MutationAnnotationFormat_Virtual}(object, file, sep, ...)

\S4method{writeData}{MutationAnnotationFormat}(object, file, ...)

\S4method{writeData}{VEP_Virtual}(object, file, sep, ...)

\S4method{writeData}{VEP}(object, file, ...)
}
\arguments{
\item{object}{Object of class VEP}

\item{...}{additional arguments to passed}

\item{file}{Character string specifying a file to send output to.}

\item{sep}{Delimiter used when writing output, defaults to tab.}
}
\description{
Method writeData
}
\details{
The writeData method is used to output data held in GenVisR objects
to a file.
}
