% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build.cnView.main.R
\name{build.cnView.main}
\alias{build.cnView.main}
\title{construct CN plot}
\usage{
build.cnView.main(x, y, z = NULL, chr, cnDiff = FALSE, layers = NULL)
}
\arguments{
\item{x}{a data frame with columns chromosome, coordinate, cn, p_value}

\item{y}{a data frame with columns chromosome, coordinate for plotting boundaries}

\item{z}{a data frame with columns chromsome, start, end, segmean specifying segments called from copy number (optional)}

\item{chr}{a character string specifying chromosome}

\item{cnDiff}{Boolean specifying whether values in cn are copy number differences or actual copy number}

\item{layers}{additional ggplot2 layers to add}
}
\value{
ggplot2 object
}
\description{
given a CN data frame plot points in ggplot
}

