% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build.mutOccur.mutSpec.R
\name{build.mutOccur.mutSpec}
\alias{build.mutOccur.mutSpec}
\title{plot mutation recurrence in genes}
\usage{
build.mutOccur.mutSpec(data_frame, layers = NULL)
}
\arguments{
\item{data_frame}{a data frame in MAF format}

\item{layers}{additional ggplot2 layers}
}
\value{
a ggplot object
}
\description{
plot a bar graph displaying the percentage of samples with a mutation
}

