% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_coverage.R
\name{build_coverage}
\alias{build_coverage}
\title{build coverage plot}
\usage{
build_coverage(data_frame, x_limits = NULL, display_x_axis = TRUE,
  colour = "blue", plot_type = "line")
}
\arguments{
\item{data_frame}{an object of class data frame containing columns stop and cov}

\item{display_x_axis}{boolean specifying whether to plot x-axis labels}

\item{colour}{character string specifying the color of the data in the plot}

\item{plot_type}{character string specifying one of line, area for data display}

\item{xlimits}{vector giving x-axis limits for plot, inferred from data if not specified}
}
\value{
ggplot object
}
\description{
given data build a coverage plot to represent the data
}

