% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_gene.R
\name{build_gene}
\alias{build_gene}
\title{build gene plot}
\usage{
build_gene(data_frame)
}
\arguments{
\item{data_frame}{an object of class data frame specifying gene feature information}
}
\value{
ggplot object
}
\description{
given a data frame with gene feature information build the ggplot2 object
}

