% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CN_dummy_data.R
\name{CN_dummy_data}
\alias{CN_dummy_data}
\title{retrieve and format CN_cohort plot supplemental data}
\usage{
CN_dummy_data(x)
}
\arguments{
\item{x}{data frame containg columns chromosome, start, end}
}
\value{
object of class data frame formatted to internal specifications
}
\description{
given a genome obtain Start and Stop positions for all chromosomes in the genome
}

