% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cnFreq.R
\name{cnFreq}
\alias{cnFreq}
\title{Construct copy number frequency plot}
\usage{
cnFreq(x, CN_low_cutoff = 1.5, CN_high_cutoff = 2.5, plot_title = NULL,
  background = "grey90", CN_low_colour = "#002EB8",
  CN_high_colour = "#A30000", x_lab_size = 12, y_lab_size = 12,
  facet_lab_size = 10, layers = NULL)
}
\arguments{
\item{x}{object of class data frame with rows representing the proportion of CN losses/gains across the genome (default), or actual CN values. The formor option contains columns chromosome, start, end, gain, and loss, and the latter option contains columns chromosome, start, end, segmean, and sample}

\item{CN_low_cutoff}{numeric value representing the cutoff for copy number losses. Only used if x represents CN values.}

\item{CN_high_cutoff}{numeric value representing the cutoff for copy number gains. Only used if x represents CN values.}

\item{plot_title}{character string for title of plot}

\item{background}{character string specifying backround color of plot}

\item{CN_low_colour}{character string specifying low value of colour gradient}

\item{CN_high_colour}{character string specifying high value of colour gradient}

\item{x_lab_size}{integer specifying the size of the x label}

\item{y_lab_size}{integer specifying the size of the y label}

\item{facet_lab_size}{integer specifying the size of the faceted labels}

\item{layers}{a valid ggplot layer to over-ride default parameters}
}
\value{
ggplot object
}
\description{
given a matrix construct a plot to display copy number changes across the genome for a group of samples
}

