% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cnFreq_qual.R
\name{cnFreq_qual}
\alias{cnFreq_qual}
\title{check input to cnFreq}
\usage{
cnFreq_qual(x)
}
\arguments{
\item{x}{a data frame with columns chromosome, start, end, gain, and loss,
or chromosome, start, end, segmean, and sample}
}
\value{
list containing data frame passing quality checks and the type
of plot (proportional losses/gain or frequency of losses/gains)
}
\description{
Perform a data quality check for inputs to cnFreq
}

