% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CN_cohort_plot.R, R/cnSpec.R, R/cnSpec.qual.R
\name{cnSpec}
\alias{cnSpec}
\alias{cnSpec.qual}
\title{Construct CN cohort plot}
\usage{
cnSpec(x, y = NULL, genome = "hg19", plot_title = NULL,
  background = "grey90", CN_low_colour = "#002EB8",
  CN_high_colour = "#A30000", x_lab_size = 12, y_lab_size = 12,
  facet_lab_size = 10, layers = NULL)

cnSpec(x, y = NULL, genome = "hg19", plot_title = NULL,
  background = "grey90", CN_low_colour = "#002EB8",
  CN_high_colour = "#A30000", x_lab_size = 12, y_lab_size = 12,
  facet_lab_size = 10, layers = NULL)

cnSpec.qual(x, y)
}
\arguments{
\item{x}{object of class data frame containing columns Chromosome, Start, Stop, SegMean, Sample.name}

\item{y}{object of class data frame containing columns chromosome start end}

\item{genome}{character string specifying UCSC genome from which data is derived}

\item{plot_title}{character string for title of plot}

\item{background}{character string specifying backround color of plot}

\item{CN_low_colour}{character string specifying low value of colour gradient}

\item{CN_high_colour}{character string specifying high value of colour gradient}

\item{x_lab_size}{integer specifying the size of the X label}

\item{facet_lab_size}{integer specifying the size of the faceted labels}

\item{layers}{a valid ggplot layer to over-ride default parameters}

\item{Y_lab_size}{integer specifying the size of the Y label}

\item{x}{object of class data frame containing columns chromosome, start, end, segmean, sample}

\item{genome}{character string specifying UCSC genome from which data is derived}

\item{plot_title}{character string for title of plot}

\item{background}{character string specifying backround color of plot}

\item{CN_low_colour}{character string specifying low value of colour gradient}

\item{CN_high_colour}{character string specifying high value of colour gradient}

\item{x_lab_size}{integer specifying the size of the X label}

\item{Y_lab_size}{integer specifying the size of the Y label}

\item{facet_lab_size}{integer specifying the size of the faceted labels}

\item{x}{object of class data frame containing columns Chromosome, Start, Stop, SegMean, Sample.name}

\item{y}{object of class data frame containing option ......}
}
\value{
ggplot object

ggplot object

character string specifying input passed quality check
}
\description{
given a data frame construct a plot to display CN information for a group of samples

given a data frame construct a plot to display CN information for a group of samples

given a data frame construct a plot to display CN information for a group of samples
}

