% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnSpec_qual.R
\name{cnSpec_qual}
\alias{cnSpec_qual}
\title{Construct CN cohort plot}
\usage{
cnSpec_qual(x, y, genome, CNscale)
}
\arguments{
\item{x}{object of class data frame containing columns chromosome, start,
stop, segmean, sample}

\item{y}{object of class data frame containing user supplied chromosome
locations}

\item{genome}{character string specifying a user supplied genome}

\item{CNscale}{Character string specifying if copy number calls supplied are
relative (i.e.copy neutral == 0) or absolute (i.e. copy neutral ==2). One of 
"relative" or "absolute"}
}
\value{
character string specifying input passed quality check
}
\description{
given a data frame construct a plot to display CN information for a group of
samples
}

