% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnView_subsetChr.R
\name{cnView_subsetChr}
\alias{cnView_subsetChr}
\title{subset based on chr}
\usage{
cnView_subsetChr(x, chr)
}
\arguments{
\item{x}{a data frame with columns Chr, Coord, Tumor, Normal, Diff}

\item{chr}{character string specifying UCSC chromosome to plot one of chr...
or all}
}
\value{
object of class data frame
}
\description{
given a data frame with cytogenetic band locations subset out specific
chromosome
}

