% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compIdent.R
\name{compIdent}
\alias{compIdent}
\title{Construct identity snp comparison plot}
\usage{
compIdent(x, genome)
}
\arguments{
\item{x}{data frame with column names sample_name, bamfile}

\item{genome}{Object of class BSgenome specifying the genome}
}
\value{
graphical object
}
\description{
Given the bam file path, count the number of reads at the 24 SNP locations
}
\details{
TODO
}
\examples{
#TODO
}

