% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compIdent.R
\name{compIdent}
\alias{compIdent}
\title{Construct identity snp comparison plot}
\usage{
compIdent(x, genome, target = NULL, debug = FALSE)
}
\arguments{
\item{x}{data frame with column names sample_name, bamfile}

\item{genome}{Object of class BSgenome specifying the genome}

\item{debug}{Boolean specifying if test datasets should be used for
debugging.}

\item{targetbed}{Object of class data frame containing target locations in
1-base format and containing columns "chr", "start", "end", "var", "name"}
}
\value{
graphical object
}
\description{
Given the bam file path, count the number of reads at the 24 SNP locations
}
\details{
compIdent is a function designed to compa
}
\examples{
# Read in BSgenome object (hg19)
library(BSgenome.Hsapiens.UCSC.hg19)
hg19 <- BSgenome.Hsapiens.UCSC.hg19

# Generate plot
compIdent(genome=hg19, debug=TRUE)
}

