% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compIdent_buildMain.R
\name{compIdent_buildMain}
\alias{compIdent_buildMain}
\title{Compare sample identities}
\usage{
compIdent_buildMain(x)
}
\arguments{
\item{x}{Data frame of vaf for each sample}
}
\value{
ggplot2 object
}
\description{
Produce an identity SNP plot displaying VAFs of 24 SNP locations and coverage
information to compare multiple sample identities
}

