% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compIdent_buildMain.R
\name{compIdent_buildMain}
\alias{compIdent_buildMain}
\title{Compare sample identities}
\usage{
compIdent_buildMain(count_tables, sample_names)
}
\arguments{
\item{sample_names}{list of names associated with each sample in counts list}

\item{counts}{list of readcount tables}
}
\value{
none
}
\description{
Produce an identity SNP plot displaying VAFs of 24 SNP locations and coverage
information to compare multiple sample identities
}

