% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covBars.R
\name{covBars}
\alias{covBars}
\title{Construct coverage cohort plot}
\usage{
covBars(x, col = NULL, plot_title = NULL, background = "grey90",
  x_lab_size = 12, y_lab_size = 12, facet_lab_size = 10, layers = NULL)
}
\arguments{
\item{x}{object of class data frame containing rows for the coverage and
columns the sample names}

\item{col}{vector of colors for the coverage bars}

\item{plot_title}{character string for title of plot}

\item{background}{character string specifying backround color of plot}

\item{x_lab_size}{integer specifying the size of the X label}

\item{y_lab_size}{integer specifying the size of the Y label}

\item{facet_lab_size}{integer specifying the size of the faceted labels}

\item{layers}{Additional layers to be plotted, can be a theme but must be a
ggplot layer}
}
\value{
ggplot object
}
\description{
given a matrix construct a plot to display coverage as percentage bars for a
group of samples
}
\examples{
# Create data
x <- matrix(sample(100000,500), nrow=50, ncol=10, dimnames=list(0:49,paste0("Sample",1:10)))

# Call function
covBars(x)
}

