% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dodge_coord_x.R
\name{dodge_coord_x}
\alias{dodge_coord_x}
\title{dodge coordinates}
\usage{
dodge_coord_x(x, rep.fact = 5000, rep.dist.lmt = 500, attr.fact = 0.1,
  adj.max = 0.1, adj.lmt = 0.5, iter.max = 50000)
}
\arguments{
\item{x}{numeric vector of position coordinates on x axis}
}
\value{
numeric vector of dodged position coordinates on x axis
}
\description{
given amino acid position coordinates on, before grouping and dodge on x axis
}

