% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Clinical-class.R,
%   R/Waterfall-class.R
\docType{methods}
\name{drawPlot}
\alias{drawPlot}
\alias{drawPlot,Clinical-method}
\alias{drawPlot,Clinical}
\alias{drawPlot,Waterfall-method}
\alias{drawPlot,Waterfall}
\title{Method drawPlot}
\usage{
drawPlot(object, ...)

\S4method{drawPlot}{Clinical}(object, ...)

\S4method{drawPlot}{Waterfall}(object, ...)
}
\description{
Method drawPlot
}
