% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Clinical-class.R,
%   R/Lolliplot-class.R, R/MutSpectra-class.R, R/Rainfall-class.R,
%   R/Waterfall-class.R
\docType{methods}
\name{drawPlot}
\alias{drawPlot}
\alias{drawPlot,Clinical-method}
\alias{drawPlot,Lolliplot-method}
\alias{drawPlot,MutSpectra-method}
\alias{drawPlot,Rainfall-method}
\alias{drawPlot,Waterfall-method}
\title{Method drawPlot}
\usage{
drawPlot(object, ...)

\S4method{drawPlot}{Clinical}(object, ...)

\S4method{drawPlot}{Lolliplot}(object, ...)

\S4method{drawPlot}{MutSpectra}(object, ...)

\S4method{drawPlot}{Rainfall}(object, ...)

\S4method{drawPlot}{Waterfall}(object, ...)
}
\arguments{
\item{object}{Object of class Waterfall, MutSpectra, or Clinical}

\item{...}{additional arguments to passed}
}
\description{
Method drawPlot
}
\details{
The drawPlot method is used to draw plots created by GenVisR plot 
constructor functions.
}
