% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genCov.R
\name{genCov}
\alias{genCov}
\title{produce a coverage plot}
\usage{
genCov(x, txdb, gr, genome, reduce = F, gene.colour = NULL,
  gene_name = "Gene", label.bg_fill = "black", label.text_fill = "white",
  label.border = "black", label.size = 10, label.width_ratio = c(1, 10),
  cov.colour = "blue", cov.plot_type = "line", cov.layers = NULL,
  base = c(10, 2, 2), transform = c("Intron", "CDS", "UTR"),
  gene.plot_transcript_name = TRUE, gene.transcript_name_size = 6)
}
\arguments{
\item{x}{named list containing data frames with columns stop and cov}

\item{txdb}{A TxDb object for a genome}

\item{gr}{A Granges object specifying a region of interest}

\item{genome}{Object of class BSgenome specifying the genome}

\item{reduce}{Boolean specifying whether to collapse isoforms in the ROI}

\item{gene.colour}{character string specifying the colour of the gene to be plotted}

\item{gene_name}{character string specifying the name of the gene or ROI}

\item{label.bg_fill}{character string giving the colour to fill the label}

\item{label.text_fill}{character string giving the colour to fill the text}

\item{label.border}{character string specifying the colour to fill the border of the label}

\item{label.size}{integer specifying the size of the text within the label}

\item{label.width_ratio}{vector of length 2 giving the ratio of track labels to plot}

\item{cov.colour}{character string specifying the color of the data in the coverage plot}

\item{cov.plot_type}{character string specifying one of line, area for coverage data display}

\item{cov.layers}{additional ggplot2 layers for the coverage plot}

\item{base}{A vector of log bases to transform the data, corresponding to the elements of transform}

\item{transform}{A vector of strings designating what objects to log transform}

\item{gene.plot_transcript_name}{Boolean specifying whether to plot the transcript name}

\item{gene.transcript_name_size}{Integer specifying the size of the transcript name text}
}
\value{
ggplot object
}
\description{
produce a coverage plot displaying gene and coverage information
}

