% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gene_plot.R
\name{gene_plot}
\alias{gene_plot}
\title{plot gene}
\usage{
gene_plot(txdb, gr, genome, reduce = FALSE, transformIntronic = FALSE,
  output_transInt_table = FALSE, gene_colour = NULL, cores = 1)
}
\arguments{
\item{txdb}{A TxDb object for a genome}

\item{gr}{A Granges object specifying a region of interest}

\item{genome}{Object of class BSgenome specifying the genome}

\item{reduce}{Boolean specifying whether to collapse isoforms in the ROI}

\item{transformIntronic}{Boolean specifying whether to perform a log transform on intronic space}

\item{output_transInt_table}{Boolean specifying whether to output a master gene features table instead of a plot when transformIntronic is TRUE}

\item{cores}{Integer specifying the number of cores to use for processing}
}
\value{
ggplot object
}
\description{
given a Granges object plot genomic features within the Granges object
}

