% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneViz.R
\name{geneViz}
\alias{geneViz}
\title{Gene Plot}
\usage{
geneViz(txdb, gr, genome, reduce = FALSE, gene_colour = NULL, base = c(10,
  2, 2), transform = c("Intron", "CDS", "UTR"), plot_transcript_name = TRUE,
  transcript_name_size = 4, layers = NULL)
}
\arguments{
\item{txdb}{A TxDb object for a genome}

\item{gr}{A Granges object specifying a region of interest}

\item{genome}{Object of class BSgenome specifying the genome}

\item{reduce}{Boolean specifying whether to collapse isoforms in the ROI}

\item{gene_colour}{character string specifying the colour of genomic features plotted}

\item{base}{base A vector of log bases to transform the data, corresponding to the elements of transform}

\item{transform}{A vector of strings designating what objects to log transform}

\item{plot_transcript_name}{Boolean specifying whether to plot the transcript name}

\item{transcript_name_size}{Integer specifying the size of the transcript name text}

\item{layers}{additional ggplot2 layers to plot}
}
\value{
ggplot object
}
\description{
given a Granges object plot genomic features within the Granges object
}
\examples{
# need transcript data for reference
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene

# need a biostrings object for reference
library(BSgenome.Hsapiens.UCSC.hg19)
genome <- BSgenome.Hsapiens.UCSC.hg19

# need Granges object
gr <- GRanges(seqnames=c("chr10"),
ranges=IRanges(start=c(89622195), end=c(89729532)), strand=strand(c("+")))

# Plot the graphic
geneViz(txdb, gr, genome)
}

