% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Clinical-class.R,
%   R/MutSpectra-class.R, R/Rainfall-class.R, R/Waterfall-class.R,
%   R/combinedCnLoh-class.R, R/lohSpec-class.R
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,Clinical-method}
\alias{getData}
\alias{getData,ClinicalData-method}
\alias{getData}
\alias{.getData_MutSpectra}
\alias{getData}
\alias{getData,MutSpectraPrimaryData-method}
\alias{getData}
\alias{getData,MutSpectra-method}
\alias{getData}
\alias{.getData_Rainfall}
\alias{getData}
\alias{getData,RainfallPrimaryData-method}
\alias{getData}
\alias{getData,Rainfall-method}
\alias{getData}
\alias{.getData_waterfall}
\alias{getData}
\alias{getData,WaterfallData-method}
\alias{getData}
\alias{getData,Waterfall-method}
\alias{getData}
\alias{.getData_combinedCnLoh}
\alias{getData}
\alias{getData,cnLohData-method}
\alias{getData}
\alias{getData,cnLoh-method}
\alias{getData}
\alias{.getData_lohSpec}
\alias{getData}
\alias{getData,lohData-method}
\alias{getData}
\alias{getData,lohSpec-method}
\alias{getData}
\title{Method getData}
\usage{
getData(object, ...)

\S4method{getData}{Clinical}(object, ...)

\S4method{getData}{ClinicalData}(object, ...)

.getData_MutSpectra(object, name = NULL, index = NULL, ...)

\S4method{getData}{MutSpectraPrimaryData}(object, name = NULL, index = NULL,
  ...)

\S4method{getData}{MutSpectra}(object, name = NULL, index = NULL, ...)

.getData_Rainfall(object, name = NULL, index = NULL, ...)

\S4method{getData}{RainfallPrimaryData}(object, name = NULL, index = NULL,
  ...)

\S4method{getData}{Rainfall}(object, name = NULL, index = NULL, ...)

.getData_waterfall(object, name = NULL, index = NULL, ...)

\S4method{getData}{WaterfallData}(object, name = NULL, index = NULL, ...)

\S4method{getData}{Waterfall}(object, name = NULL, index = NULL, ...)

.getData_combinedCnLoh(object, name = NULL, index = NULL, ...)

\S4method{getData}{cnLohData}(object, name = NULL, index = NULL, ...)

\S4method{getData}{cnLoh}(object, name = NULL, index = NULL, ...)

.getData_lohSpec(object, name = NULL, index = NULL, ...)

\S4method{getData}{lohData}(object, name = NULL, index = NULL, ...)

\S4method{getData}{lohSpec}(object, name = NULL, index = NULL, ...)
}
\arguments{
\item{object}{Object of class Clinical,}

\item{...}{additional arguments to passed}

\item{name}{String corresponding to the slot for which to extract data from.}

\item{index}{Integer specifying the slot for which to extract data from.}
}
\description{
Method getData

Helper function to getData from classes

Helper function to get data from classes

Helper function to getData from classes

Helper function to get data from classes

Helper function to get data from classes
}
\details{
The getData method is an accessor function used to access data held
in GenVisR objects.
}
