% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Clinical-class.R,
%   R/MutSpectra-class.R, R/Waterfall-class.R
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,Clinical-method}
\alias{getData}
\alias{getData,ClinicalData-method}
\alias{getData}
\alias{getData,MutSpectraPrimaryData-method}
\alias{getData}
\alias{getData,WaterfallData-method}
\alias{getData}
\alias{getData,Waterfall-method}
\alias{getData}
\title{Method getData}
\usage{
getData(object, ...)

\S4method{getData}{Clinical}(object, ...)

\S4method{getData}{ClinicalData}(object, ...)

\S4method{getData}{MutSpectraPrimaryData}(object, ...)

\S4method{getData}{WaterfallData}(object, name = NULL, index = NULL, ...)

\S4method{getData}{Waterfall}(object, name = NULL, index = NULL, ...)
}
\arguments{
\item{object}{Object of class Clinical,}

\item{...}{additional arguments to passed}

\item{name}{String corresponding to the slot for which to extract data from
the object WaterfallData, one of "primaryData", "simpleMutationCounts",
"complexMutationCounts", "geneData", or "mutationHierarchy}

\item{index}{Integer specifying the slot for which to extract data from the
object WaterfallData, must be a value between 1-4.}

\item{name}{String corresponding to the slot for which to extract data from
the object WaterfallData, one of "primaryData", "simpleMutationCounts",
"complexMutationCounts", "geneData", or "mutationHierarchy}

\item{index}{Integer specifying the slot for which to extract data from the
object WaterfallData, must be a value between 1-4.}
}
\description{
Method getData
}
\details{
The getData method is an accessor function used to access data held
in GenVisR objects.
}
