% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Clinical-class.R
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,Clinical-method}
\alias{getData,Clinical}
\title{Method getData}
\usage{
getData(object, ...)

\S4method{getData}{Clinical}(object, ...)
}
\arguments{
\item{object}{Object of class}

\item{...}{additional arguments to passed}
}
\description{
Method getData
}
