% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MutSpectra-class.R,
%   R/Waterfall-class.R
\docType{methods}
\name{getGrob}
\alias{getGrob}
\alias{getGrob,MutSpectraPlots-method}
\alias{getGrob,MutSpectra-method}
\alias{getGrob,WaterfallPlots-method}
\alias{getGrob,Waterfall-method}
\title{Method getGrob}
\usage{
getGrob(object, ...)

\S4method{getGrob}{MutSpectraPlots}(object, index = 1, ...)

\S4method{getGrob}{MutSpectra}(object, index = 1, ...)

\S4method{getGrob}{WaterfallPlots}(object, index = 1, ...)

\S4method{getGrob}{Waterfall}(object, index = 1, ...)
}
\arguments{
\item{object}{Object of clas MutSpectra}

\item{...}{additional arguments to passed}

\item{index}{integer specifying the plot index to extract}
}
\description{
Method getGrob
}
