% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/lohSpec-class.R
\docType{methods}
\name{getLohCalculation}
\alias{getLohCalculation}
\alias{getLohCalculation,data.table-method}
\alias{getLohCalculation}
\title{Method getLohCalculation}
\usage{
getLohCalculation(object, ...)

\S4method{getLohCalculation}{data.table}(object, windowData, normal, verbose,
  ...)
}
\arguments{
\item{object}{Object of class VarScanFormat}

\item{...}{additional arguments to passed}

\item{normal}{integer specifying normal vaf}

\item{object}{of class lohData}

\item{window_data}{of class data.table}
}
\description{
Method getLohCalculation
}
