% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/combinedCnLoh-class.R,
%   R/lohSpec-class.R
\docType{methods}
\name{getLohSegmentation}
\alias{getLohSegmentation}
\alias{getLohSegmentation}
\alias{getLohSegmentation,data.table-method}
\alias{getLohSegmentation}
\title{Method getLohSegmentation}
\usage{
getLohSegmentation(object, ...)

\S4method{getLohSegmentation}{data.table}(object, ...)

\S4method{getLohSegmentation}{data.table}(object, ...)
}
\arguments{
\item{object}{Object of class VarScanFormat}

\item{...}{additional arguments to passed}

\item{object}{of class data.table}

\item{chrData}{of class data.table}

\item{object}{of class data.table}

\item{chrData}{of class data.table}
}
\description{
Method getLohSegmentation
}
