% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/combinedCnLoh-class.R,
%   R/lohSpec-class.R
\docType{methods}
\name{getLohSlidingWindow}
\alias{getLohSlidingWindow}
\alias{getLohSlidingWindow}
\alias{getLohSlidingWindow,data.table-method}
\alias{getLohSlidingWindow}
\title{Method getLohSlidingWindow}
\usage{
getLohSlidingWindow(object, ...)

\S4method{getLohSlidingWindow}{data.table}(object, step, windowSize, ...)

\S4method{getLohSlidingWindow}{data.table}(object, step, windowSize, ...)
}
\arguments{
\item{object}{Object of class VarScanFormat}

\item{...}{additional arguments to passed}

\item{step}{integer specifying the step size between the start position of
each window}

\item{windowSize}{integer specifying the window size for loh calcuations}

\item{object}{of class data.table}

\item{object}{of class lohData}

\item{step}{integer specifying the step size between the start position of
each window}

\item{windowSize}{integer specifying the window size for loh calcuations}
}
\value{
Data.table with window start/stop positions

Data.table with window start/stop positions
}
\description{
Method getLohSlidingWindow
}
