% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MutationAnnotationFormat_Virtual-class.R, R/MutationAnnotationFormat-class.R
\docType{methods}
\name{getMeta}
\alias{getMeta}
\alias{getMeta,MutationAnnotationFormat}
\alias{getMeta,MutationAnnotationFormat-method}
\alias{getMeta,MutationAnnotationFormat_Virtual}
\alias{getMeta,MutationAnnotationFormat_Virtual-method}
\title{Method getMeta}
\usage{
getMeta(object, ...)

\S4method{getMeta}{MutationAnnotationFormat_Virtual}(object, ...)

\S4method{getMeta}{MutationAnnotationFormat}(object, ...)
}
\value{
data.table object holding meta information.
}
\description{
Method getMeta
}

