% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GMS-class.R,
%   R/MutationAnnotationFormat-class.R
\docType{methods}
\name{getMeta}
\alias{getMeta}
\alias{getMeta,GMS-method}
\alias{getMeta,GMS}
\alias{getMeta,MutationAnnotationFormat-method}
\alias{getMeta,MutationAnnotationFormat}
\title{Method getMeta}
\usage{
getMeta(object, ...)

\S4method{getMeta}{GMS}(object, ...)

\S4method{getMeta}{MutationAnnotationFormat}(object, ...)
}
\arguments{
\item{object}{Object of class VEP, GMS, or MutationAnnotationFormat}

\item{...}{additional arguments to passed}
}
\description{
Method getMeta
}
