% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GMS_Virtual-class.R,
%   R/GMS-class.R, R/MutationAnnotationFormat_Virtual-class.R,
%   R/MutationAnnotationFormat-class.R, R/VEP_Virtual-class.R, R/VEP-class.R
\docType{methods}
\name{getMeta}
\alias{getMeta}
\alias{getMeta,GMS_Virtual-method}
\alias{getMeta}
\alias{getMeta,GMS-method}
\alias{getMeta}
\alias{getMeta,MutationAnnotationFormat_Virtual-method}
\alias{getMeta}
\alias{getMeta,MutationAnnotationFormat-method}
\alias{getMeta}
\alias{getMeta,VEP_Virtual-method}
\alias{getMeta}
\alias{getMeta,VEP-method}
\alias{getMeta}
\title{Method getMeta}
\usage{
getMeta(object, ...)

\S4method{getMeta}{GMS_Virtual}(object, ...)

\S4method{getMeta}{GMS}(object, ...)

\S4method{getMeta}{MutationAnnotationFormat_Virtual}(object, ...)

\S4method{getMeta}{MutationAnnotationFormat}(object, ...)

\S4method{getMeta}{VEP_Virtual}(object, ...)

\S4method{getMeta}{VEP}(object, ...)
}
\arguments{
\item{object}{Object of class VEP, GMS, or MutationAnnotationFormat}

\item{...}{additional arguments to passed}
}
\description{
Method getMeta
}
