% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GMS-class.R,
%   R/MutationAnnotationFormat-class.R
\docType{methods}
\name{getMutation}
\alias{getMutation}
\alias{getMutation,GMS-method}
\alias{getMutation,GMS}
\alias{getMutation,MutationAnnotationFormat-method}
\alias{getMutation,MutationAnnotationFormat}
\title{Method getMutation}
\usage{
getMutation(object, ...)

\S4method{getMutation}{GMS}(object, ...)

\S4method{getMutation}{MutationAnnotationFormat}(object, ...)
}
\description{
Method getMutation
}
