% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MutationAnnotationFormat_Virtual-class.R, R/MutationAnnotationFormat-class.R
\docType{methods}
\name{getMutation}
\alias{getMutation}
\alias{getMutation,MutationAnnotationFormat}
\alias{getMutation,MutationAnnotationFormat-method}
\alias{getMutation,MutationAnnotationFormat_Virtual}
\alias{getMutation,MutationAnnotationFormat_Virtual-method}
\title{Method getMutation}
\usage{
getMutation(object, ...)

\S4method{getMutation}{MutationAnnotationFormat_Virtual}(object, ...)

\S4method{getMutation}{MutationAnnotationFormat}(object, ...)
}
\value{
data.table object holding genomic mutations and containing column names
"variant_classification", "variant_type", "reference_allele", "variant_allele_1", "variant_allele_2".
}
\description{
Method getMutation
}

