% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GMS_Virtual-class.R,
%   R/GMS-class.R, R/MutationAnnotationFormat_Virtual-class.R,
%   R/MutationAnnotationFormat-class.R, R/VEP_Virtual-class.R, R/VEP-class.R
\docType{methods}
\name{getMutation}
\alias{getMutation}
\alias{getMutation,GMS_Virtual-method}
\alias{getMutation}
\alias{getMutation,GMS-method}
\alias{getMutation}
\alias{getMutation,MutationAnnotationFormat_Virtual-method}
\alias{getMutation}
\alias{getMutation,MutationAnnotationFormat-method}
\alias{getMutation}
\alias{getMutation,VEP_Virtual-method}
\alias{getMutation}
\alias{getMutation,VEP-method}
\alias{getMutation}
\title{Method getMutation}
\usage{
getMutation(object, ...)

\S4method{getMutation}{GMS_Virtual}(object, ...)

\S4method{getMutation}{GMS}(object, ...)

\S4method{getMutation}{MutationAnnotationFormat_Virtual}(object, ...)

\S4method{getMutation}{MutationAnnotationFormat}(object, ...)

\S4method{getMutation}{VEP_Virtual}(object, ...)

\S4method{getMutation}{VEP}(object, ...)
}
\arguments{
\item{object}{Object of class VEP, GMS, or MutationAnnotationFormat}

\item{...}{additional arguments to passed}
}
\description{
Method getMutation
}
