% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MutationAnnotationFormat_Virtual-class.R, R/MutationAnnotationFormat-class.R
\docType{methods}
\name{getPosition}
\alias{getPosition}
\alias{getPosition,MutationAnnotationFormat}
\alias{getPosition,MutationAnnotationFormat-method}
\alias{getPosition,MutationAnnotationFormat_Virtual}
\alias{getPosition,MutationAnnotationFormat_Virtual-method}
\title{Method getPosition}
\usage{
getPosition(object, ...)

\S4method{getPosition}{MutationAnnotationFormat_Virtual}(object, ...)

\S4method{getPosition}{MutationAnnotationFormat}(object, ...)
}
\description{
Method getPosition
}

