% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GMS-class.R,
%   R/MutationAnnotationFormat-class.R
\docType{methods}
\name{getPosition}
\alias{getPosition}
\alias{getPosition,GMS-method}
\alias{getPosition,GMS}
\alias{getPosition,MutationAnnotationFormat-method}
\alias{getPosition,MutationAnnotationFormat}
\title{Method getPosition}
\usage{
getPosition(object, ...)

\S4method{getPosition}{GMS}(object, ...)

\S4method{getPosition}{MutationAnnotationFormat}(object, ...)
}
\arguments{
\item{object}{Object of class VEP, GMS, or MutationAnnotationFormat}

\item{...}{additional arguments to passed}
}
\description{
Method getPosition
}
