% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GMS_Virtual-class.R,
%   R/GMS-class.R, R/MutationAnnotationFormat_Virtual-class.R,
%   R/MutationAnnotationFormat-class.R, R/VEP_Virtual-class.R, R/VEP-class.R
\name{getPosition}
\alias{getPosition}
\alias{getPosition,GMS_Virtual-method}
\alias{getPosition,GMS-method}
\alias{getPosition,MutationAnnotationFormat_Virtual-method}
\alias{getPosition,MutationAnnotationFormat-method}
\alias{getPosition,VEP_Virtual-method}
\alias{getPosition,VEP-method}
\title{Method getPosition}
\usage{
getPosition(object, ...)

\S4method{getPosition}{GMS_Virtual}(object, ...)

\S4method{getPosition}{GMS}(object, ...)

\S4method{getPosition}{MutationAnnotationFormat_Virtual}(object, ...)

\S4method{getPosition}{MutationAnnotationFormat}(object, ...)

\S4method{getPosition}{VEP_Virtual}(object, ...)

\S4method{getPosition}{VEP}(object, ...)
}
\arguments{
\item{object}{Object of class VEP, GMS, or MutationAnnotationFormat}

\item{...}{additional arguments to passed}
}
\description{
Method getPosition
}
