% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/MutationAnnotationFormat_Virtual-class.R,
%   R/MutationAnnotationFormat-class.R
\docType{methods}
\name{getSample}
\alias{getSample}
\alias{getSample,MutationAnnotationFormat_Virtual-method}
\alias{getSample,MutationAnnotationFormat_Virtual}
\alias{getSample,MutationAnnotationFormat-method}
\alias{getSample,MutationAnnotationFormat}
\title{Method getSample}
\usage{
getSample(object, ...)

\S4method{getSample}{MutationAnnotationFormat_Virtual}(object, ...)

\S4method{getSample}{MutationAnnotationFormat}(object, ...)
}
\description{
Method getSample
}
