% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GMS_Virtual-class.R,
%   R/GMS-class.R, R/MutationAnnotationFormat_Virtual-class.R,
%   R/MutationAnnotationFormat-class.R, R/VEP_Virtual-class.R, R/VEP-class.R,
%   R/VarScanFormat_Virtual-class.R
\docType{methods}
\name{getSample}
\alias{getSample}
\alias{getSample,GMS_Virtual-method}
\alias{getSample,GMS-method}
\alias{getSample,MutationAnnotationFormat_Virtual-method}
\alias{getSample,MutationAnnotationFormat-method}
\alias{getSample,VEP_Virtual-method}
\alias{getSample,VEP-method}
<<<<<<< HEAD
\alias{getSample}
\alias{getSample,VarScanFormat_Virtual-method}
\alias{getSample}
=======
>>>>>>> 4fa590fa705dcce36ce96337d88a31bd01bd6792
\title{Method getSample}
\usage{
getSample(object, ...)

\S4method{getSample}{GMS_Virtual}(object, ...)

\S4method{getSample}{GMS}(object, ...)

\S4method{getSample}{MutationAnnotationFormat_Virtual}(object, ...)

\S4method{getSample}{MutationAnnotationFormat}(object, ...)

\S4method{getSample}{VEP_Virtual}(object, ...)

\S4method{getSample}{VEP}(object, ...)

\S4method{getSample}{VarScanFormat_Virtual}(object, ...)
}
\arguments{
\item{object}{Object of class VEP, GMS, or MutationAnnotationFormat}

\item{...}{additional arguments to passed}
}
\description{
Method getSample
}
