% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenVisR.R
\docType{data}
\name{HCC1395_N}
\alias{HCC1395_N}
\title{Normal BAM}
\format{
a data frame with 59 observations and 6 variables
}
\usage{
data(HCC1395_N)
}
\value{
Object of class list
}
\description{
A data set containing read pileups intersecting 24 identity snp locations
from GenVisR::SNPloci. Pileups are from downsampled bams and originate from
normal tissue corresponding to the HCC1395 breast cancer cell line.
}
\keyword{datasets}
