% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenVisR.R
\docType{data}
\name{HCC1395_T}
\alias{HCC1395_T}
\title{Tumor BAM}
\format{a named list with 24 elements and 6 variables}
\usage{
data(HCC1395_T)
}
\value{
Object of class list
}
\description{
A data set containing aligned reads intersecting 24 identity snp locations
from GenVisR::SNPloci. Aligned reads are downsampled and originate from
tumor tissue corresponding to the HCC1395 breast cancer cell line.
}
\keyword{datasets}

