% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GenVisR.R
\docType{data}
\name{hg19chr}
\alias{hg19chr}
\title{hg19 chromosome boundaries}
\format{a data frame with 24 observations and 3 variables}
\usage{
data(hg19chr)
}
\description{
A data set containg chromosome boundaries corresponding to hg19 in base pairs
}
\keyword{datasets}

