% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ideoView.R
\name{ideoView}
\alias{ideoView}
\title{Chromosome plot}
\usage{
ideoView(x, chromosome = "chr1", chr_txt_angle = 45, chr_txt_size = 5,
  layers = NULL)
}
\arguments{
\item{x}{data frame specifying cytogenetic information for a genome,
should contain columns "chrom", "chromStart", "chromEnd", "name", "gieStain"}

\item{chromosome}{character string specifying UCSC chromosome to plot}

\item{chr_txt_angle}{integer specifying angle of text when plotting band text}

\item{chr_txt_size}{integer specifying size of text when plotting band text}

\item{layers}{additional ggplot2 layers for the ideogram}
}
\value{
ggplot object
}
\description{
plot a chromosome in ggplot using UCSC database
}
\examples{
# Obtain cytogenetic information for the genome of interest from attached
# data set cytoGeno
data <- cytoGeno[cytoGeno$genome == 'hg38',]

# Call ideoView for chromosome 1
ideoView(data, chromosome='chr1', chr_txt_size=4)
}

