% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lohSpec-class.R
\docType{class}
\name{lohData-class}
\alias{lohData-class}
\alias{lohData}
\title{Private Class lohData}
\usage{
lohData(object, chr, samples, y, genome, step, window_size, normal, verbose)
}
\arguments{
\item{object}{Object of class VarScan}
}
\description{
An S4 class for the Data of the loh plot object

Constructor for the lohData class.
}
