% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lohSpec-class.R
\name{lohSpec-class}
\alias{lohSpec-class}
\alias{lohSpec}
\title{Class LOH}
\usage{
lohSpec(x = NULL, path = NULL, fileExt = NULL, y = NULL,
  genome = "hg19", gender = NULL, step = 1e+06, window_size = 2500000,
  normal = 0.5, colourScheme = "inferno", plotLayer = NULL,
  method = "slide", out = "plot")
}
\arguments{
\item{step}{Integer value specifying the step size (i.e. the number of base
pairs to move the window). required when method is set to slide
(see details).}

\item{normal}{Boolean specifiying what value to use for normal VAF when 
calcualting average LOH difference. Defaults to .50\% if FALSE. 
If TRUE, will use average normal VAF in each individual sample as value 
to calculate LOH.}

\item{input}{Object of class VarScan.}

\item{Character}{vector specifying the chromosomes of interest.}

\item{samples}{Character vector specifying samples to plot. If not NULL
all samples in "input" not specified with this parameter are removed.}

\item{BSgenome}{Object of class BSgenome to extract genome wide chromosome 
coordinates}

\item{windowSize}{Integer value specifying the size of the window in base
pairs in which to calculate the mean Loss of Heterozygosity (see details).}
}
\description{
An S4 class for the lohSpec plot object

Constructor for the lohSpec class
}
\section{Slots}{

\describe{
\item{\code{lohFreq_plot}}{gtable object for the lohFreq plot}

\item{\code{lohSpec_plot}}{gtable object for the lohSpec plot}

\item{\code{lohData}}{data.table object soring loh data with column names: sample, 
chromosome, position, t_vaf, n_vaf.}
}}

