% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lohSpec-class.R
\name{lohSpec}
\alias{lohSpec}
\alias{lohSpec}
\alias{LohSpec}
\title{Class lohSpec}
\usage{
LohSpec(input, lohSpec = TRUE, chromosomes = "autosomes", samples = NULL,
  BSgenome = BSgenome, step = 1e+06, windowSize = 2500000,
  normal = FALSE, gradientMidpoint = 0.2, gradientColors = c("#ffffff",
  "#b2b2ff", "#000000"), plotAType = "proportion", plotALohCutoff = 0.2,
  plotAColor = "#98F5FF", plotALayers = NULL, plotBLayers = NULL,
  sectionHeights = c(0.25, 0.75), verbose = FALSE)
}
\arguments{
\item{input}{Object of class VarScan.}

\item{chromosomes}{Character vector specifying chromosomes to plot. If not NULL
all chromosomes in "input" not specified with this parameter are removed.}

\item{samples}{Character vector specifying samples to plot. If not NULL
all samples in "input" not specified with this parameter are removed.}

\item{BSgenome}{Object of class BSgenome to extract genome wide chromosome 
coordinates}

\item{step}{Integer value specifying the step size (i.e. the number of base
pairs to move the window). required when method is set to slide
(see details).}

\item{windowSize}{Integer value specifying the size of the window in base
pairs in which to calculate the mean Loss of Heterozygosity (see details).}

\item{normal}{Boolean specifiying what value to use for normal VAF when 
calcualting average LOH difference. Defaults to .50\% if FALSE. 
If TRUE, will use average normal VAF in each individual sample as value 
to calculate LOH.}
}
\description{
An S4 class for the lohSpec plot object

Constructor for the lohSpec class
}
\section{Slots}{

\describe{
\item{\code{lohFreq_plot}}{gtable object for the lohFreq plot}

\item{\code{lohSpec_plot}}{gtable object for the lohSpec plot}

\item{\code{lohData}}{data.table object soring loh data with column names: sample, 
chromosome, position, t_vaf, n_vaf.}
}}

