% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lohView.R
\name{lohView}
\alias{lohView}
\title{Plot LOH data}
\usage{
lohView(x = NULL, path = NULL, fileExt = NULL, y = NULL,
  genome = "hg19", gender = NULL, step = 1e+05, window_size = 250000,
  normal = 50, gradient_midpoint = 20, gradient_low = "#ffffff",
  gradient_mid = "#b2b2ff", gradient_high = "#000000", theme_layer = NULL,
  method = "tile")
}
\arguments{
\item{x}{object of class data frame with columns 'chromosome', 'position',
'n_vaf', 't_vaf', 'sample'}

\item{path}{character string specifying which directory contains
the sample information stored as datasets with columns "chromosome",
"position", "n_vaf", "t_vaf", and "sample" (required if x is not specified)}

\item{fileExt}{character string specifying the file extensions of files
within the path specified (required if path argument is specified)}

\item{y}{object of class data frame with columns 'chromosome', 'start',
'end' specifying chromosomal boundaries for a genome assembly
(required if genome is not specified)}

\item{genome}{character string specifying the genome assembly from which
input data is based}

\item{gender}{vector of length equal to the number of samples, consisting of
elements from the set {"M", "F"}}

\item{step}{integer with the length of divisions (bp) in chromosomes (only
required when tile=FALSE}

\item{window_size}{integer with the size of the sliding window (bp) to be
applied}

\item{normal}{integer specifying the normal VAF frequency used in LOH
calculations}

\item{gradient_midpoint}{object of class numeric specifying the midpoint
for legend's gradient scale}

\item{gradient_low}{object of class character for hex color code for
gradient's lower values}

\item{gradient_mid}{object of class character for hex color code for
gradient's middle values}

\item{gradient_high}{object of class character for hex color code for
gradient's upper values}

\item{theme_layer}{ggplot theme object specifying parameters for non data
elements}

\item{method}{Character string specifying the approach to be used for
displaying Loss of Heterozygosity, one of "tile" or "slide".}
}
\value{
grid object
}
\description{
Produce a graphic visualizing Loss of Heterozygosity in a cohort
}
\examples{
## Insert an example once the first part of this is understood
}

