% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lohView_slidingWindow.R
\name{lohView_slidingWindow}
\alias{lohView_slidingWindow}
\title{obtain LOH data}
\usage{
lohView_slidingWindow(loh_data, step, window_size, normal)
}
\arguments{
\item{loh_data}{data frame with columns "chromosome", "position", "n_vaf",
"t_vaf", "sample" giving raw vaf calls for germline variants}

\item{step}{integer with the length of divisions (bp) in chromosomes}

\item{window_size}{integer with the size of the sliding window (bp) to be
applied}

\item{normal}{integer specifying the normal VAF frequency used in LOH
calculations}
}
\value{
object of class dataframe containing LOH data
}
\description{
Obtain LOH heatmap on entire chromsomes from samples in a cohort
}

