% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lolliplot-class.R
\name{Lolliplot-class}
\alias{Lolliplot-class}
\alias{Lolliplot}
\title{Class Lolliplot}
\usage{
Lolliplot(input, gene, transcript, species = "hsapiens",
  host = "www.ensembl.org", verbose = FALSE)
}
\arguments{
\item{input}{Object of class MutationAnnotationFormat, GMS, VEP, or a data.table with appropriate columns}

\item{gene}{Character string specifying a gene to plot}

\item{transcript}{Character string specifying the ensembl transcript for which to plot, should be a transcript which corresponds
to the gene parameter.}

\item{species}{Character string specifying a species when using biomaRt queries}

\item{host}{Character string specifying a host to connect to when using biomaRt queries}
}
\description{
An S4 class for the lolliplot object
}
\section{Slots}{

\describe{
\item{\code{PlotA}}{gtable object for the top sub-plot}

\item{\code{PlotB}}{gtable object for the bottom sub-plot}

\item{\code{Grob}}{gtable object storing the arranged plot}

\item{\code{primaryData}}{data.table object storing the primary data}

\item{\code{geneData}}{data.table object storing gene and domain coordinates}
}}

