% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lolliplot.qual.R
\name{lolliplot.qual}
\alias{lolliplot.qual}
\title{Check input to lolliplot}
\usage{
lolliplot.qual(x, y)
}
\arguments{
\item{x}{object of class data frame containing columns transcript_name, gene, and amino_acid_change and rows denoting mutations}

\item{y}{object of class data frame containing columns transcript_name, and amino_acid_change and rows denoting mutations}
}
\value{
character string specifying input passed quality check
}
\description{
Perform Basic quality checks for lolliplot input
}

