% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lolliplot.R
\name{lolliplot}
\alias{lolliplot}
\title{Construct Lolliplot from data}
\usage{
lolliplot(data, cosmic = TRUE, fill_value = "trv_type",
  label_column = NULL, plot_text_angle = 45, plot_text_size = 5,
  point_size = 1, gene_colour = "#999999", spread_degree = 15)
}
\arguments{
\item{data}{object of class data frame containing columns transcript_name, gene, and amino_acid_change for specific transcript}

\item{cosmic}{boolean value specifying if cosmic mutations should be retrieved and plotted}
}
\value{
object of class ggplot2
}
\description{
Construct a Lolliplot from object of class data frame giving observed mutations and transcript
}

