% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lolliplot.R
\name{lolliplot}
\alias{lolliplot}
\title{Construct Lolliplot from data}
\usage{
lolliplot(x, y = NULL, z = NULL, fillCol = NULL, labelCol = NULL,
  plot_text_angle = 45, plot_text_size = 5, point_size = 4,
  gene_colour = "#999999", obsA.rep.fact = 5000, obsA.rep.dist.lmt = 500,
  obsA.attr.fact = 0.1, obsA.adj.max = 0.1, obsA.adj.lmt = 0.5,
  obsA.iter.max = 50000, obsB.rep.fact = 5000, obsB.rep.dist.lmt = 500,
  obsB.attr.fact = 0.1, obsB.adj.max = 0.1, obsB.adj.lmt = 0.5,
  obsB.iter.max = 50000, plot_sidechain = FALSE, species = "hsapiens",
  max_VerticalLolli = NULL, layers = NULL, paletteA = NULL,
  paletteB = NULL)
}
\arguments{
\item{x}{object of class data frame containing column names "transcript_name"
, "gene", and "amino_acid_change" with rows denoting mutations for top track}

\item{y}{object of class data frame containing columns "transcript_name", and
"amino_acid_change" with rows denoting mutations for bottom track (optional)}

\item{z}{Object of class data frame containing columns "description",
"start", "stop" for manual denotation of regions of interest}

\item{fillCol}{character string giving the name of the column to shade
variants on, required}

\item{labelCol}{character string specifying column containing text
information to be plotted, (optional) defaults to NULL}

\item{plot_text_angle}{integer specifying angle of text to be plotted if
labelCol is specified}

\item{plot_text_size}{integer specifying the size of the text plotted if
labelCol is specified}

\item{point_size}{integer specifying the size of points plotted}

\item{gene_colour}{color specifying the background fill of the plotted gene}

\item{obsA.rep.fact}{repulsive factor for plotted mutations observed track}

\item{obsA.rep.dist.lmt}{repulsive distance limit for plotted mutations
observed track}

\item{obsA.attr.fact}{attraction factor for plotted mutations observed track}

\item{obsA.adj.max}{maximum position change for each iteration observed track}

\item{obsA.adj.lmt}{position adjustment limit which simulation stops observed
track}

\item{obsA.iter.max}{maximum iterations beyond which to stop the simulation
observed track}

\item{obsB.rep.fact}{repulsive factor for plotted mutations cosmic track}

\item{obsB.rep.dist.lmt}{repulsive distance limit for plotted mutations
cosmic track}

\item{obsB.attr.fact}{attraction factor for plotted mutations cosmic track}

\item{obsB.adj.max}{maximum position change for each iteration cosmic track}

\item{obsB.adj.lmt}{position adjustment limit which simulation stops cosmic
track}

\item{obsB.iter.max}{maximum iterations beyond which to stop the simulation
cosmic track}

\item{plot_sidechain}{boolean value whether to plot the amino acid sidechains
instead of protein domains}

\item{species}{Character string specifying the species corresponding to the
given ensembl transcript id}

\item{max_VerticalLolli}{Integer specifying the upper limit of lolli points
to allow to be stacked vertically}

\item{layers}{additional ggplot2 layers to plot}

\item{paletteA}{Character vector specifying colours for gene features}

\item{paletteB}{Character vector specifying colours for lolli features}
}
\value{
object of class ggplot2
}
\description{
Construct a Lolliplot from object of class data frame giving observed
mutations and an ensembl transcript id
}
\examples{
# Create input data
data <- brcaMAF[brcaMAF$Hugo_Symbol == 'TP53',c('Hugo_Symbol', 'amino_acid_change_WU')]
data <- as.data.frame(cbind(data, 'ENST00000269305'))
colnames(data) <- c('gene', 'amino_acid_change', 'transcript_name')

# Call lolliplot
lolliplot(data)
}

