% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lolliplot.R
\name{lolliplot}
\alias{lolliplot}
\title{Construct Lolliplot from data}
\usage{
lolliplot(x, y = NULL, fill_value = NULL, label_column = NULL,
  plot_text_angle = 45, plot_text_size = 5, point_size = 3,
  gene_colour = "#999999", obsA.rep.fact = 5000, obsA.rep.dist.lmt = 500,
  obsA.attr.fact = 0.1, obsA.adj.max = 0.1, obsA.adj.lmt = 0.5,
  obsA.iter.max = 50000, obsB.rep.fact = 5000, obsB.rep.dist.lmt = 500,
  obsB.attr.fact = 0.1, obsB.adj.max = 0.1, obsB.adj.lmt = 0.5,
  obsB.iter.max = 50000, plot_sidechain = FALSE, taxId = 9606)
}
\arguments{
\item{x}{object of class data frame containing columns transcript_name, gene, and amino_acid_change and rows denoting mutations}

\item{y}{object of class data frame containing columns transcript_name, and amino_acid_change and rows denoting mutations}

\item{fill_value}{character string giving the name of the column to shade variants on "required"}

\item{label_column}{character string specifying column containing text information to be plotted, defaults to NULL}

\item{plot_text_angle}{integer specifying angle of text to be plotted if label_column is specified}

\item{plot_text_size}{integer specifying the size of the text plotted if label_column is specified}

\item{point_size}{integer specifying the size of points plotted}

\item{gene_colour}{color specifying the background fill of the plotted gene}

\item{obsA.rep.fact}{repulsive factor for plotted mutations observed track}

\item{obsA.rep.dist.lmt}{repulsive distance limit for plotted mutations observed track}

\item{obsA.attr.fact}{attraction factor for plotted mutations observed track}

\item{obsA.adj.max}{maximum position change for each iteration observed track}

\item{obsA.adj.lmt}{position adjustment limit which simulation stops observed track}

\item{obsA.iter.max}{maximum iterations beyond which to stop the simulation observed track}

\item{obsB.rep.fact}{repulsive factor for plotted mutations cosmic track}

\item{obsB.rep.dist.lmt}{repulsive distance limit for plotted mutations cosmic track}

\item{obsB.attr.fact}{attraction factor for plotted mutations cosmic track}

\item{obsB.adj.max}{maximum position change for each iteration cosmic track}

\item{obsB.adj.lmt}{position adjustment limit which simulation stops cosmic track}

\item{obsB.iter.max}{maximum iterations beyond which to stop the simulation cosmic track}

\item{plot_sidechain}{boolean value whether to plot the amino acid sidechains instead of protein domains}

\item{taxId}{integer specifying the uniprot taxonomy id for the species of interest}
}
\value{
object of class ggplot2
}
\description{
Construct a Lolliplot from object of class data frame giving observed mutations and an ensembl transcript id
}

