% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lolliplot.R
\name{lolliplot}
\alias{lolliplot}
\title{Construct Lolliplot from data}
\usage{
lolliplot(data, cosmic = TRUE, fill_value = "trv_type",
  label_column = NULL, plot_text_angle = 45, plot_text_size = 5,
  point_size = 1, gene_colour = "#999999", obs.rep.fact = 5000,
  obs.rep.dist.lmt = 500, obs.attr.fact = 0.1, obs.adj.max = 0.1,
  obs.adj.lmt = 0.5, obs.iter.max = 50000, cos.rep.fact = 5000,
  cos.rep.dist.lmt = 500, cos.attr.fact = 0.1, cos.adj.max = 0.1,
  cos.adj.lmt = 0.5, cos.iter.max = 50000, plot_sidechain = FALSE)
}
\arguments{
\item{data}{object of class data frame containing columns transcript_name, gene, and amino_acid_change and rows denoting mutations}

\item{cosmic}{boolean value specifying if cosmic mutations should be retrieved and plotted}

\item{fill_value}{character string giving the name of the column to shade variants on "required"}

\item{label_column}{character string specifying column containing text information to be plotted, defaults to NULL}

\item{plot_text_angle}{integer specifying angle of text to be plotted if label_column is specified}

\item{plot_text_size}{integer specifying the size of the text plotted if label_column is specified}

\item{point_size}{integer specifying the size of points plotted}

\item{gene_colour}{color specifying the background fill of the plotted gene}

\item{obs.rep.fact}{repulsive factor for plotted mutations observed track}

\item{obs.rep.dist.lmt}{repulsive distance limit for plotted mutations observed track}

\item{obs.attr.fact}{attraction factor for plotted mutations observed track}

\item{obs.adj.max}{maximum position change for each iteration observed track}

\item{obs.adj.lmt}{position adjustment limit which simulation stops observed track}

\item{obs.iter.max}{maximum iterations beyond which to stop the simulation observed track}

\item{cos.rep.fact}{repulsive factor for plotted mutations cosmic track}

\item{cos.rep.dist.lmt}{repulsive distance limit for plotted mutations cosmic track}

\item{cos.attr.fact}{attraction factor for plotted mutations cosmic track}

\item{cos.adj.max}{maximum position change for each iteration cosmic track}

\item{cos.adj.lmt}{position adjustment limit which simulation stops cosmic track}

\item{cos.iter.max}{maximum iterations beyond which to stop the simulation cosmic track}

\item{plot_sidechain}{boolean value whether to plot the amino acid sidechains instead of protein domains}
}
\value{
object of class ggplot2
}
\description{
Construct a Lolliplot from object of class data frame giving observed mutations and an ensembl transcript id
}

