% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lolliplot_dodgeCoordY.R
\name{lolliplot_dodgeCoordY}
\alias{lolliplot_dodgeCoordY}
\title{dodge coordinates}
\usage{
lolliplot_dodgeCoordY(x, track = "top")
}
\arguments{
\item{x}{data frame containing columns coord_x_dodge and group}

\item{track}{character vector, one of "top", "bottom" specifying whether to
dodge in a positive or negative fashion}
}
\value{
numeric vector of dodged position coordinates on y axis
}
\description{
given amino acid position coordinates on, before grouping and dodge on y axis
}

