% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{lolliplot_transcriptID2codingSeq}
\alias{lolliplot_transcriptID2codingSeq}
\title{fetch protein length}
\usage{
lolliplot_transcriptID2codingSeq(transcriptID, species = "hsapiens",
  host = "www.ensembl.org")
}
\arguments{
\item{transcriptID}{character string giving ensembl transcript id}

\item{species}{character string to use when searching for ensemblMart dataset}

\item{host}{Host to connect to.}
}
\value{
length in residues of ensembl transcript id
}
\description{
Retrieve protein length from ensembl database given enseml transcript id
}
