% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenVisR.R
\docType{data}
\name{Luc2CNraw}
\alias{Luc2CNraw}
\title{CN data for Luc2}
\format{a data frame with 10000 observations and 5 variables}
\usage{
data(Luc2CNraw)
}
\description{
Downsampled copy number calls for Luc2 chromosome 14 originating from cnv-hmm "Genomic Landscape of Non-Small Cell Lung Cancer in Smokers and Never-Smokers" PMID:22980976
}
\keyword{datasets}

