% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenVisR.R
\docType{data}
\name{LucCNseg}
\alias{LucCNseg}
\title{Truncated CN segments}
\format{a data frame with 3336 observations and 6 variables}
\usage{
data(LucCNseg)
}
\value{
Object of class data frame
}
\description{
A data set in long format containing Copy Number segments for 4 samples
corresponding to "lung cancer" from Govindan et al. Cell. 2012, PMID:22980976
}
\keyword{datasets}
