% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MAF_to_anno.R
\name{MAF_to_anno}
\alias{MAF_to_anno}
\title{Convert MAF File}
\usage{
MAF_to_anno(data_frame)
}
\arguments{
\item{data_frame}{a data frame in MAF format}
}
\value{
a data frame coerced from MAF to TGI format
}
\description{
Convert columns of a mutation annotation file "MAF" into a format recognizable by internal functions
}

