% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergeRegions.R
\name{mergeRegions}
\alias{mergeRegions}
\title{Create Region Table}
\usage{
mergeRegions(gene_features, gr, base, transform)
}
\arguments{
\item{gr}{Granges object specifying the ROI}

\item{base}{A vector of log bases to transform the data, corresponding to the elements of transform}

\item{transform}{A vector of strings designating what objects to log transform}

\item{txdb}{A TxDb object for a genome}
}
\value{
Master region table data frame
}
\description{
Create a master region table by merging isoforms
}

