% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map_coord_space.R, R/mergeRegions.R
\name{mergeRegions}
\alias{map_coord_space}
\alias{mergeRegions}
\title{Map regions to transformed space}
\usage{
map_coord_space(master, coord, base = exp(1))

mergeRegions(gene_features, gr, base = exp(1))
}
\arguments{
\item{master}{an object of class data frame containing columns start, end, width, type, trans_start, trans_end representing a master genomic region with features from isoforms merged}

\item{coord}{an object of class data frame containing columns start and end to map to transformed space}

\item{gr}{Granges object specifying the ROI}

\item{txdb}{A TxDb object for a genome}
}
\value{
an object of class data frame identical to coord but with extra columns for transformed coord

Master region table data frame
}
\description{
Reference a master genomic region file to map original positions to a transformed space

Create a master region table by merging isoforms
}

