% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutationAnnotationFormat.R
\docType{class}
\name{MutationAnnotationFormat-class}
\alias{MutationAnnotationFormat}
\alias{MutationAnnotationFormat-class}
\title{Class MutationAnnotationFormat}
\usage{
\S4method{initialize}{MutationAnnotationFormat}(.Object, path, version, verbose)

MutationAnnotationFormat(path, version = "auto", verbose = FALSE)
}
\arguments{
\item{path}{String specifying the path to a MAF file.}

\item{version}{String specifying the version of the MAF file, if set to auto
the version will be grabbed from the header in the MAF file.}

\item{verbose}{Boolean specifying if progress should be reported while reading
in the MAF file.}
}
\description{
An S4 class to represent data in mutation annotation format

Initalizer method of MutationAnnotationFormat class

Constructor for MutationAnnotationFormat
}
\section{Slots}{

\describe{
\item{\code{path}}{String specifying the path of the MAF file read in.}

\item{\code{version}}{String specifying the version of the MAF file.}

\item{\code{position}}{data.table object with columns "Chromosome", "Start_Position", "End_Position" and "Strand"}

\item{\code{mutation}}{data.table object with columns "Variant_Classification", "Variant_Type", "Reference_Allele", "Tumor_Seq_Allele1", "Tumor_Seq_Allele2"}

\item{\code{sample}}{data.table object with columns "Tumor_Sample_Barcode"}

\item{\code{meta}}{data.table object containing additional information}
}}

