% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutationCOS.R
\name{mutationCOS}
\alias{mutationCOS}
\title{retrieve cosmic mutation observations}
\usage{
mutationCOS(transcriptID, rep.fact, rep.dist.lmt, attr.fact, adj.max, adj.lmt,
  iter.max)
}
\arguments{
\item{transcriptID}{character string specifying ensembl transcript ID to retrieve mutations for}
}
\value{
object of class data frame giving comsic mutation observations
}
\description{
Create a data frame of cosmic mutation observations
}

