% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mutationObs.R
\name{mutationObs}
\alias{mutationObs}
\title{format mutation observations}
\usage{
mutationObs(data, fill_value, label_column, rep.fact, rep.dist.lmt, attr.fact,
  adj.max, adj.lmt, iter.max)
}
\arguments{
\item{data}{object of class data frame with columns trv_type and amino acid change}
}
\value{
object of class data frame giving mutation observations
}
\description{
Create a data frame of mutation observations
}

