% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutSpec.mutation_recurrence_subset.R
\name{mutSpec.mutation_recurrence_subset}
\alias{mutSpec.mutation_recurrence_subset}
\title{Mutation Recurrence Cutoff}
\usage{
mutSpec.mutation_recurrence_subset(x, recurrence_cutoff)
}
\arguments{
\item{x}{a data frame in long format with columns 'gene', 'trv_type', sample}

\item{recurrence_cutoff}{integer specifying removal of entries not seen
in at least "x" percent of samples}
}
\value{
a subset data frame
}
\description{
Subset a MAF file keeping only samples that meet a mutation recurrence cutoff
}

