% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutSpec.mutation_gene_subset.R
\name{mutSpec.mutation_sample_subset}
\alias{mutSpec.mutation_gene_subset}
\alias{mutSpec.mutation_sample_subset}
\title{mutation sample cutoff gene based}
\usage{
mutSpec.mutation_gene_subset(x, genes)
}
\arguments{
\item{x}{a data frame in long format with columns 'gene', 'trv_type'}

\item{genes}{character vector listing genes to plot}
}
\value{
a subset data frame
}
\description{
Subset a internal mutSpec file keeping only samples within the specified gene
 list
}

