% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutSpectra-class.R
\docType{methods}
\name{MutSpectra-class}
\alias{MutSpectra-class}
\alias{MutSpectra}
\alias{MutSpectra}
\alias{Waterfall}
\title{Class MutSpectra}
\usage{
\S4method{initialize}{MutSpectra}(.Object, input, verbose)

Waterfall(input, labelColumn = NULL, samples = NULL, coverage = NULL,
  mutation = NULL, genes = NULL, mutationHierarchy = NULL,
  recurrence = NULL, geneOrder = NULL, geneMax = NULL,
  sampleOrder = NULL, plotA = c("frequency", "burden", NULL),
  plotATally = c("simple", "complex"), plotALayers = NULL,
  plotB = c("proportion", "frequency", NULL), plotBTally = c("simple",
  "complex"), plotBLayers = NULL, gridOverlay = FALSE, drop = TRUE,
  labelSize = 5, labelAngle = 0, sampleNames = TRUE, clinical = NULL,
  sectionHeights = NULL, sectionWidths = NULL, verbose = FALSE,
  plotCLayers = NULL)
}
\arguments{
\item{input}{Object of class MutationAnnotationFormat, GMS, VEP.}

\item{verbose}{Boolean specifying if status messages should be reported}

\item{plotALayers}{list of ggplot2 layers to be passed to the frequency plot.}

\item{plotBLayers}{list of ggplot2 layers to be passed to the proportion plot.}

\item{clinical}{Object of class Clinical, used for adding a clinical data subplot.}

\item{sectionHeights}{Numeric vector specifying relative heights of each plot section,
should sum to one. Expects a value for each section.}

\item{sectionWidths}{Numeric vector specifying relative heights of each plot section,
should sum to one. Expects a value for each section.}

\item{plotCLayers}{list of ggplot2 layers to be passed to the clinical plot.}
}
\description{
An S4 class for the MutSpectra plot object

Initalizer method for the MutSpectra class

Constructor for the MutSpectra class.
}
\section{Slots}{

\describe{
\item{\code{PlotA}}{gtable object for the mutation frequencies.}

\item{\code{PlotB}}{gtable object for the mutation proportions.}

\item{\code{PlotC}}{gtable object for clinical data sub-plot.}

\item{\code{Grob}}{gtable object for the arranged plot.}

\item{\code{primaryData}}{data.table object storing the primary data, should have
column names sample, mutation, frequency, proportion.}

\item{\code{ClinicalData}}{data.table object storing the data used to plot the
clinical sub-plot.}
}}

