% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenVisR.R
\docType{data}
\name{PIK3CA}
\alias{PIK3CA}
\title{Subset MAF file for PIK3CA gene}
\format{
a data frame with 361 observations and 19 variables
}
\usage{
data(PIK3CA)
}
\value{
Object of class data frame
}
\description{
A data set originating from the open access TCGA data (6c93f518-1956-4435-9806-37185266d248),
the data set is composed of mutations for the PIK3CA gene for breast cancer. this is primarily
intended to test the Lolliplot() function.
}
\keyword{datasets}
