% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_clinical.R
\name{plot_clinical}
\alias{plot_clinical}
\title{plot clinical information}
\usage{
plot_clinical(x)
}
\arguments{
\item{x}{a data frame in "long" format giving additional information to be plotted, requires columns "sample", "variable", and "value"}
}
\value{
a grob object
}
\description{
given a data frame with columns names sample, variable, and value create a ggplot2 object
}

