% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a mutation heatmap}
\usage{
plot_heatmap(data_frame, grid, label_x, gene_label_size, file_type)
}
\arguments{
\item{data_frame}{a data frame in MAF format}

\item{grid}{boolean value whether to overlay a grid on the plot}

\item{label_x}{boolean value whether to label the x axis}

\item{gene_label_size}{numeric value indicating the size of the gene labels on the y-axis}

\item{file_type}{character string specifying the file type, one of 'MAF' or 'TGI'}
}
\value{
a ggplot2 object
}
\description{
Plot a Mutation Landscape with variables sample, gene, mutation
}

