% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenVisR.R
\docType{data}
\name{ptenCov}
\alias{ptenCov}
\title{Coverage for PTEN}
\format{a data frame with 10734 observations and 4 variables}
\usage{
data(ptenCov)
}
\value{
Object of class data frame
}
\description{
A data set containing coverage data for PTEN from Juric et al. Nature 2015,
PMID:25409150
}
\keyword{datasets}

