% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GenVisR.R
\docType{data}
\name{ptenCOV}
\alias{ptenCOV}
\title{Coverage for the pten gene}
\format{a data frame with 107338 observations and 4 variables}
\usage{
data(ptenCOV)
}
\value{
Object of class data frame
}
\description{
A data set containing coverage data for pten obtained using "bedtools multicov" from "Convergent loss of PTEN leads to clinical resistance to a PI3K-alpha inhibitor" PMID:25409150
}
\keyword{datasets}

