% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transcriptID2length.R
\name{transcriptID2length}
\alias{transcriptID2length}
\title{Convert trascript ID to protien length}
\usage{
transcriptID2length(transcriptID, up)
}
\arguments{
\item{transcriptID}{String specifying ensembl transcript ID}

\item{up}{uniprot data object from UniProt.ws}
}
\value{
numeric value specifying protien length
}
\description{
Retrieve the length of an ensembl transcript ID in amino acids
}

