% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TvTi.R
\name{TvTi}
\alias{TvTi}
\title{plot transitions/transversions}
\usage{
TvTi(x, y = NULL, type = "Proportion", label_x_axis = TRUE,
  x_axis_text_angle = 45, palette = c("#1b9e77", "#d95f02", "#7570b3",
  "#e7298a", "#66a61e", "#e6ab02"), file_type = "MAF")
}
\arguments{
\item{x}{Object of class data frame containing columns 'sample', reference', 'variant'}

\item{y}{named vector containing expected transition/transversion proportions with names "A->C or T->G", "A->G or T->C", "A->T or T->A", "G->A or C->T", "G->C or C->G", "G->T or C->A" or a data frame with column names "Prop", "trans_tranv" and levels of trans_tranv matching "A->C or T->G", "A->G or T->C", "A->T or T->A", "G->A or C->T", "G->C or C->G", "G->T or C->A"}

\item{type}{Object of class character specifying whether to plot the Proportion or Frequency, one of "Prop"}

\item{label_x_axis}{boolean specifying wheter to label x axis}

\item{x_axis_text_angle}{Integer specifying the angle to labels on x_axis}

\item{palette}{Character vector of length 6 specifying colors for trans/tranv type}

\item{file_type}{Character string specifying the format the input is in, one of 'MAF', 'MGI'}
}
\value{
Object of class data frame with indels removed
}
\description{
Given a data frame with columns reference, variant, and sample construct a transition/transversion plot
}

