% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TvTi.R
\name{TvTi}
\alias{TvTi}
\title{plot transitions/transversions}
\usage{
TvTi(x, y = NULL, type = "Proportion", label_x_axis = TRUE,
  x_axis_text_angle = 45, palette = c("#2A2650", "#95fE52", "#F9C59B",
  "#0F722C", "#D7C7E9", "#FFB93F"), file_type = "MAF", layers = NULL)
}
\arguments{
\item{x}{Object of class data frame containing columns 'sample', reference', 'variant'}

\item{y}{named vector containing expected transition/transversion proportions with names "A->C or T->G", "A->G or T->C", "A->T or T->A", "G->A or C->T", "G->C or C->G", "G->T or C->A" or a data frame with column names "Prop", "trans_tranv" and levels of trans_tranv matching "A->C or T->G", "A->G or T->C", "A->T or T->A", "G->A or C->T", "G->C or C->G", "G->T or C->A"}

\item{type}{Object of class character specifying whether to plot the Proportion or Frequency, one of "Proportion", "Frequency"}

\item{label_x_axis}{boolean specifying wheter to label x axis}

\item{x_axis_text_angle}{Integer specifying the angle of labels on the x_axis}

\item{palette}{Character vector of length 6 specifying colors for each trans/tranv type}

\item{file_type}{Character string specifying the format the input is in, one of 'MAF', 'MGI'}

\item{layers}{Additional ggplot2 layers to add}
}
\value{
ggplot2 object
}
\description{
Given a data frame with columns reference, variant, and sample construct a transition/transversion plot
}
\examples{
TvTi(brcaMAF, type='Frequency',
palette=c("#77C55D", "#A461B4", "#C1524B", "#93B5BB", "#4F433F", "#BFA753"), x_axis_text_angle=60)
}

