% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TvTi_alignPlot.R
\name{TvTi_alignPlot}
\alias{TvTi_alignPlot}
\title{align TvTi plots on y axis}
\usage{
TvTi_alignPlot(p1 = NULL, p2 = NULL, p3 = NULL)
}
\arguments{
\item{p1}{main plot}

\item{p2}{left expected value subplot}

\item{p3}{bottom clinical subplot}
}
\value{
ggplot object
}
\description{
align transition/transversion plots
}

