% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Waterfall-class.R
\docType{methods}
\name{Waterfall-class}
\alias{Waterfall}
\alias{Waterfall-class}
\title{Class Waterfall}
\usage{
\S4method{initialize}{Waterfall}(.Object, input, labelColumn)

Waterfall(input, labelColumn = NULL)
}
\arguments{
\item{input}{MutationAnnotationFormat class holding genomic information.}
}
\description{
An S4 class for the waterfall plot object

Initalizer method for the Waterfall class

Constructor for the Waterfall class.
}
\section{Slots}{

\describe{
\item{\code{plotA}}{gtable object for the top sub-plot.}

\item{\code{plotB}}{gtable object for the left sub-plot.}

\item{\code{plotC}}{gtable object for the main plot.}

\item{\code{plotD}}{gtable object for the bottom sub-plot.}

\item{\code{Grob}}{gtable object for the arranged plot.}

\item{\code{Data}}{list of data.table objects storing the plotted data.}
}}

