% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Waterfall-class.R
\docType{methods}
\name{Waterfall-class}
\alias{Waterfall-class}
\alias{Waterfall}
\alias{Waterfall}
\title{Class Waterfall}
\usage{
\S4method{initialize}{Waterfall}(.Object, input, labelColumn, samples, coverage,
  noSynonymous, genes, mutationHierarchy, recurrence, geneOrder, geneMax,
  sampleOrder, plotA, plotATally, plotALayers, plotB, plotBTally, plotBLayers,
  gridOverlay, drop, labelSize, labelAngle, sampleNames, clinical,
  sectionHeights, sectionWidths, plotCLayers, verbose)

Waterfall(input, labelColumn = NULL, samples = NULL, coverage = NULL,
  noSynonymous = FALSE, genes = NULL, mutationHierarchy = NULL,
  recurrence = NULL, geneOrder = NULL, geneMax = NULL,
  sampleOrder = NULL, plotA = c("frequency", "burden", NULL),
  plotATally = c("simple", "complex"), plotALayers = NULL,
  plotB = c("proportion", "frequency", NULL), plotBTally = c("simple",
  "complex"), plotBLayers = NULL, gridOverlay = FALSE, drop = TRUE,
  labelSize = 5, labelAngle = 0, sampleNames = TRUE, clinical = NULL,
  sectionHeights = NULL, sectionWidths = NULL, verbose = FALSE,
  plotCLayers = NULL)
}
\arguments{
\item{input}{Object of class MutationAnnotationFormat, or GMS.}

\item{labelColumn}{Character vector specifying a column name from which to
extract labels for cells.}

\item{samples}{Character vector specifying samples to plot. If not NULL
all samples in "input" not specified with this parameter are removed.}

\item{coverage}{Integer specifying the size in base pairs of the genome
covered by sequence data from which mutations could be called. Required for
the mutation burden sub-plot (see details and vignette).}

\item{noSynonymous}{Boolean specifying if silent mutations should be removed
from the plot, this will not affect the mutation burden calculation.}

\item{mutationHierarchy}{Data.table object with rows specifying the order of
mutations from most to least deleterious and column names "mutation" and
"color". Used to change the default colors and/or to give priority to a
mutation for the same gene/sample (see details and vignette).}

\item{recurrence}{Numeric value between 0 and 1 specifying a
mutation recurrence cutoff. Genes which do not have mutations in the
proportion of samples defined are removed.}

\item{geneOrder}{Character vector specifying the order in which to plot
genes.}

\item{geneMax}{Integer specifying the maximum number of genes to be plotted.
Genes kept will be choosen based on the reccurence of mutations in samples.
Unless geneOrder is specified.}

\item{plotA}{String specifying the type of plot for the top sub-plot, one of
"burden", "frequency", or NULL for a mutation burden (requires coverage to be
specified), frequency of mutations, or no plot respectively.}

\item{plotATally}{String specifying one of "simple" or "complex" for a
simplified or complex tally of mutations respectively.}

\item{plotALayers}{list of ggplot2 layers to be passed to the plot.}

\item{plotB}{String specifying the type of plot for the left sub-plot, one of
"proportion", "frequency", or NULL for a plot of gene proportions frequencies
, or no plot respectively.}

\item{plotBTally}{String specifying one of "simple" or "complex" for a
simplified or complex tally of genes respectively.}

\item{plotBLayers}{list of ggplot2 layers to be passed to the plot.}

\item{gridOverlay}{Boolean specifying if a grid should be overlayed on the
waterfall plot.}

\item{drop}{Boolean specifying if unused mutations should be dropped from the
legend.}

\item{labelSize}{Integer specifying the size of label text}

\item{labelAngle}{Numeric value specifying the angle of label text}

\item{sampleNames}{Boolean specifying if samples should be labeled on the plot.}

\item{clinical}{Object of class Clinical, used for adding a clinical data subplot.}

\item{sectionHeights}{Numeric vector specifying relative heights of each plot section,
should sum to one. Expects a value for each section.}

\item{sectionWidths}{Numeric vector specifying relative heights of each plot section,
should sum to one. Expects a value for each section.}

\item{plotCLayers}{list of ggplot2 layers to be passed to the plot.}

\item{verbose}{Boolean specifying if status messages should be reported}

\item{sampOrder}{Character vector specifying the order in which to plot
samples.}
}
\description{
An S4 class for the waterfall plot object

Initalizer method for the Waterfall class

Constructor for the Waterfall class.
}
\section{Slots}{

\describe{
\item{\code{PlotA}}{gtable object for the top sub-plot.}

\item{\code{PlotB}}{gtable object for the left sub-plot.}

\item{\code{PlotC}}{gtable object for the main plot.}

\item{\code{PlotD}}{gtable object for the bottom sub-plot.}

\item{\code{Grob}}{gtable object for the arranged plot.}

\item{\code{primaryData}}{data.table object storing the primary data, should have
column names sample, gene, mutation, label.}

\item{\code{simpleMutationCounts}}{data.table object storing simplified mutation
counts, should have column names sample, mutation, Freq, mutationBurden}

\item{\code{complexMutationCounts}}{data.table object storing mutation counts per
mutation type should have column names sample, mutation, Freq, mutationBurden.}

\item{\code{geneData}}{data.table object storing gene counts, should have column
names gene, mutation, count.}

\item{\code{ClinicalData}}{data.table object stroring the data used to plot the
clinical sub-plot.}

\item{\code{MutationHierarchy}}{data.table object storing the hierarchy of mutation
type in order of most to least important and the mapping of mutation type to
color. Should have column names mutation, color, and label.}
}}

