% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{waterfall}
\alias{waterfall}
\title{Construct a oncoprint}
\usage{
waterfall()
}
\description{
This function has been removed, please use Waterfall() (capital W) Tutorial can be found at: https://currentprotocols.onlinelibrary.wiley.com/doi/full/10.1002/cpz1.252
}
