% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{waterfall_Custom2anno}
\alias{waterfall_Custom2anno}
\title{Convert Custom File}
\usage{
waterfall_Custom2anno(x, label_col)
}
\arguments{
\item{x}{a data frame with columns having values for sample, gene, mutation
type}

\item{label_col}{Character string specifying the column name of a
label column (optional)}
}
\value{
a data frame coerced from custom to annotation format
}
\description{
Convert columns of a Custom annotation file into a format
recognizable by internal functions
}
