% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{waterfall_geneRecurCutoff}
\alias{waterfall_geneRecurCutoff}
\title{Mutation Recurrence Cutoff}
\usage{
waterfall_geneRecurCutoff(x, recurrence_cutoff)
}
\arguments{
\item{x}{data frame in long format with columns 'gene', 'trv_type', 'sample'}

\item{recurrence_cutoff}{integer specifying removal of entries not seen
in at least "x" percent of samples}
}
\value{
a subset data frame
}
\description{
Subset a MAF file keeping only samples that meet a mutation recurrence cutoff
}
