% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_geneSort.R
\name{waterfall_geneSort}
\alias{waterfall_geneSort}
\title{sort MAF File by gene}
\usage{
waterfall_geneSort(x)
}
\arguments{
\item{x}{a data frame in annotation format}
}
\value{
a data frame sorted such that genes with more mutations are ordered
first
}
\description{
order a MAF file ranking genes with more mutations higher
}

