% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{waterfall_sampSort}
\alias{waterfall_sampSort}
\title{sort samples in an internal waterfall file.}
\usage{
waterfall_sampSort(x, sampOrder = NULL)
}
\arguments{
\item{x}{a data frame in long format with column names "sample",
"gene", "trv_type"}

\item{sampOrder}{Character vector specifying the order of samples to plot.}
}
\value{
a vector of samples in a sorted order
}
\description{
perform a hiearchial sort on samples based on the presence of mutations in
an ordered list of genes if a sample order is unspecified.
}
