% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/VEP_Virtual-class.R,
%   R/VEP-class.R
\docType{methods}
\name{writeData}
\alias{writeData}
\alias{writeData,VEP_Virtual-method}
\alias{writeData,VEP_Virtual}
\alias{writeData,VEP-method}
\alias{writeData,VEP}
\title{Method writeData}
\usage{
writeData(object, ...)

\S4method{writeData}{VEP_Virtual}(object, file, sep, ...)

\S4method{writeData}{VEP}(object, file, ...)
}
\arguments{
\item{object}{Object of class VEP}

\item{...}{additional arguments to passed}

\item{file}{Character string specifying a file to send output to.}

\item{sep}{Delimiter used when writing output, defaults to tab.}
}
\description{
Method writeData
}
